% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{AC}
\alias{AC}
\title{absolute change}
\usage{
AC(xTs, base = 0, lag = 1, ...)
}
\arguments{
\item{xTs}{xts object}

\item{base}{choose -1 (or less) to look into the future}

\item{lag}{observations backwards}
}
\description{
absolute change
}
