% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{addCashLogReturns}
\alias{addCashLogReturns}
\title{add cash log returns (CASHlogrets)}
\usage{
addCashLogReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object (only takes the index)}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

NEVER USED - RETIRED - DO NOT USE

}
}
