% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addCrashData}
\alias{addCrashData}
\title{add poor performance data (CRASHACML)}
\usage{
addCrashData(xTs = NULL, Predictee = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{Predictee}{target. xtsAttributes(mktdat)[["rettarget"]](Default)}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

Of the 'target', find the two(2) worst performing months
over the next(future six(6)) months
and then sum those two(2) bad worst performing months together.

calls

Predictee <- xtsAttributes(mktdata)[["rettarget"]]
sumOrdersXts(mktdata[, Predictee], r =  -6:-1, nt = 2)

see ? sumOrdersXts

}
}
