% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addCurrLeadCashAPCReturns}
\alias{addCurrLeadCashAPCReturns}
\title{add current and leading cash APC returns (CASHapcrets)}
\usage{
addCurrLeadCashAPCReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
