% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{addCurrLeadCashLogReturns}
\alias{addCurrLeadCashLogReturns}
\title{add current cash log returns (CASHlogrets)}
\usage{
addCurrLeadCashLogReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
