% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{addCurrLeadSP500LogReturns}
\alias{addCurrLeadSP500LogReturns}
\title{add current SP500 log returns (SP500logrets)}
\usage{
addCurrLeadSP500LogReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
