% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addCurrLeadSymbolAPCReturns}
\alias{addCurrLeadSymbolAPCReturns}
\title{add leading and current Symbol APC returns (SYMBOLlogrets)}
\usage{
addCurrLeadSymbolAPCReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
