% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addCurrLeadSymbolReturns}
\alias{addCurrLeadSymbolReturns}
\title{add leading and current Symbol APC returns (SYMBOLapcrets)}
\usage{
addCurrLeadSymbolReturns(xTs = NULL, Symbol = NULL, src = NULL,
  IsTarget = NULL, IsATarget = NULL, SymplifyGeneratorFUN = NULL,
  ReturnsGeneratorFUN = NULL, ReturnsGeneratorFUNArgs = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{Symbol}{getSymbols Symbol}

\item{src}{getSymbols source}

\item{IsTarget}{is the objective}

\item{IsATarget}{is a member of many possible objectives}

\item{SymplifyGeneratorFUN}{Function that Formats the output}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
\examples{
\dontrun{


addCurrLeadSymbolReturns(mktdata, Symbol = "^GSPC", src = "yahoo"
  , IsATarget = TRUE, SymplifyGeneratorFUN = "toMonthlyData"
  , ReturnsGeneratorFUN = "Returns"
  , ReturnsGeneratorFUNArgs = list(is.na.zero = TRUE,  Fun = list(Fun = "APC", lag = 1))
)

addCurrLeadSymbolReturns(mktdata, Symbol = "GDP",  src = "FRED2"
  , IsTarget = TRUE,  SymplifyGeneratorFUN = "fancifyXts"
  , ReturnsGeneratorFUN = "Returns"
  , ReturnsGeneratorFUNArgs = list(is.na.zero = TRUE,  Fun = list(Fun = "APC", lag = 3))
)

}
}
