% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{addCurrLeadWilshire5000LogReturns}
\alias{addCurrLeadWilshire5000LogReturns}
\title{add current Willshire 5000 Index log returns (WILL5000INDlogrets)}
\usage{
addCurrLeadWilshire5000LogReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
