% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addEomData}
\alias{addEomData}
\title{get the end of month data}
\usage{
addEomData(xTs = NULL, Symbol = NULL, src = NULL,
  SymplifyGeneratorFUN = NULL, CarryForward = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{Symbol}{getSymbols Symbol}

\item{src}{getSymbols source}

\item{SymplifyGeneratorFUN}{Function that Formats the output.}

\item{CarryForward}{default(NULL) after simplification perform carry forward method(if any)
of data to replace future missing values}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
