% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addLeadingWilshire5000LogReturns}
\alias{addLeadingWilshire5000LogReturns}
\title{add Willshire 5000 Index log returns (will5000idxlogrets)}
\usage{
addLeadingWilshire5000LogReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
add Willshire 5000 Index log returns (will5000idxlogrets)
}
