% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{addSP500EyeBallWts}
\alias{addSP500EyeBallWts}
\title{add SP500 log weights returns using eyeball (GSPClogrets)}
\usage{
addSP500EyeBallWts(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
