% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{addSP500LogReturns}
\alias{addSP500LogReturns}
\title{add SP500 log returns (GSPClogrets)}
\usage{
addSP500LogReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
