% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addSP500MachineWts}
\alias{addSP500MachineWts}
\title{add SP500 log weights returns using Machine learning (GSPClogrets)}
\usage{
addSP500MachineWts(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
