% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addSymbolAPCReturns}
\alias{addSymbolAPCReturns}
\title{add Symbol APC returns (SYMBOLlogrets)}
\usage{
addSymbolAPCReturns(xTs = NULL, Symbol = Symbol, src = src)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
