% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addSymbolEyeBallWts}
\alias{addSymbolEyeBallWts}
\title{add Symbol Change weights returns using eyeball (SYMBOLCHANGErets)}
\usage{
addSymbolEyeBallWts(xTs = NULL, Symbol = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{Symbol}{See ? symbolEyeBallWts}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
