% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addSymbolMachineWts}
\alias{addSymbolMachineWts}
\title{add Symbol Change weights returns using Machine learning (SYMBOLCHANGErets)}
\usage{
addSymbolMachineWts(xTs = NULL, Predictee = NULL, Predictors = NULL,
  IndicatorGeneratorFUN = NULL, NumbReplicaCopiesMultiple = NULL, ...)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
