% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addUMCSentimentEomData}
\alias{addUMCSentimentEomData}
\title{add University of Michigan: Consumer Sentiment}
\usage{
addUMCSentimentEomData(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

NOT YET USED ANYWHERE

}
}
