% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addUnRateEomData}
\alias{addUnRateEomData}
\title{add UNRATE (unrate)}
\usage{
addUnRateEomData(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
