% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{addWillShire5000EyeBallWts}
\alias{addWillShire5000EyeBallWts}
\title{add Willshire 5000 Index log weights returns using eyeball (WILL5000INDlogrets)}
\usage{
addWillShire5000EyeBallWts(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
