% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addWillShire5000Wts}
\alias{addWillShire5000Wts}
\title{add Willshire 5000 Index log weights returns (will5000idxlogrets)}
\usage{
addWillShire5000Wts(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
add Willshire 5000 Index log weights returns (will5000idxlogrets)
}
