% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addWilshire5000LogReturns}
\alias{addWilshire5000LogReturns}
\title{add Willshire 5000 Index log returns (WILL5000INDlogrets)}
\usage{
addWilshire5000LogReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
