% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{addWts}
\alias{addWts}
\title{add weights (_wts)}
\usage{
addWts(xTs = NULL, xTs1 = NULL)
}
\arguments{
\item{xTs}{target xts object}

\item{xTs1}{source xts object of new weights}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
