% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{ANN}
\alias{ANN}
\title{annualized an xts object}
\usage{
ANN(x)
}
\arguments{
\item{x}{xts object of 'percent change'}
}
\value{
xts object with values multiplied to get annualization
}
\description{
\preformatted{

Uses the xts index

}
}
\examples{
\dontrun{

ANN(
  xts(1:3, c(zoo::as.Date("1980-01-31"), zoo::as.Date("1980-02-29"), zoo::as.Date("1980-03-31")))
)
           [,1]
1980-01-31   12
1980-02-29   24
1980-03-31   36

}
}
