% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{AOMN}
\alias{AOMN}
\title{average order minimum (AOMN)}
\usage{
AOMN(x, mm = 4, m = 3)
}
\arguments{
\item{x}{xts object}

\item{mm}{numer of last mm observations(including self)}

\item{m}{number of lowest m observations}
}
\description{
\preformatted{
Of the last mm observations(including self), find the lowest m observations,
and then average those m observations
}
}
\examples{
\dontrun{
Of the last three(3) observations(including self), find the lowest two(2) observations,
and then average those two(2) observations.

xts(matrix(1:5,ncol = 1, dimnames = list(list(), list("Data"))), zoo::as.Date(0:4))

Data
1970-01-01    1
1970-01-02    2
1970-01-03    3
1970-01-04    4
1970-01-05    5

AOMN(xts(matrix(1:5,ncol = 1, dimnames = list(list(), list("Data"))), zoo::as.Date(0:4)), mm = 3, m  = 2)

AOMN
1970-01-03  1.5
1970-01-04  2.5
1970-01-05  3.5

}
}
