% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{APC}
\alias{APC}
\title{absolute percent change}
\usage{
APC(xTs, base = 0, lag = 1, ...)
}
\arguments{
\item{xTs}{xts object}

\item{base}{choose -1 (or less) to look into the future}

\item{lag}{observations backwards}
}
\description{
absolute percent change
}
