% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{APCReturns}
\alias{APCReturns}
\title{absolute proportional change results}
\usage{
APCReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object of absolute values}
}
\value{
xts object of absolute proportional change returns
}
\description{
\preformatted{

xTs values less that zero will generate a numeric error

}
}
\examples{
\dontrun{
require(xts)
xTs  <- xts(10:12,zoo::as.Date(0:2))
apcr <- APCReturns(xTs)
apcr
#                         apcrets
# 1970-01-01 0.000000000000000000
# 1970-01-02 0.095310179804324768
# 1970-01-03 0.087011376989629685

}
}
