% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{APCTC}
\alias{APCTC}
\title{absolute percent change}
\usage{
APCTC(x, base = 0, lag = 1, ...)
}
\arguments{
\item{x}{xts object}

\item{base}{choose -1 (or less) to look into the future}

\item{lag}{observations backwards}
}
\description{
\preformatted{
"absolute proportional change" multiplied by 100.00
See ? APC # absolute proportional change
}
}
\examples{
\dontrun{
See ? APC
}
}
