% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{appendAllOtherWts}
\alias{appendAllOtherWts}
\title{add all other wts to make PerformanceAnalytics::Return.portfolio happy}
\usage{
appendAllOtherWts(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
