% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{appendCashAPCWts}
\alias{appendCashAPCWts}
\title{add cash weights of apc returns (CASHapcrets)}
\usage{
appendCashAPCWts(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
