% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{appendCashWts}
\alias{appendCashWts}
\title{add cash weights returns (cashlogrets)}
\usage{
appendCashWts(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
add cash weights returns (cashlogrets)
}
