% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{applyAggregateDT}
\alias{applyAggregateDT}
\title{apply data.table aggregator on subsets of rows}
\usage{
applyAggregateDT(x, Fun, By, ...)
}
\arguments{
\item{x}{data.table to rung the aggrcation}

\item{Fun}{(default none: required) function to pass}

\item{By}{column to aggregate upon}

\item{...}{dots passed to data.table aggregator
and to to Fun.  Also, FunStr may be passed a flag to signal
not to run match.call(Fun). match.call is only
ever allowed to run once on Fun. (R language)
Also, DescTools::DoCall only returns
as.character(substitute(Fun)) == "Fun"}
}
\value{
data.table of aggregates
}
\description{
apply data.table aggregator on subsets of rows
}
\examples{
\dontrun{
MyFun <- function(x, Fun, By, ...) {

  FunStr <- as.character(substitute(Fun))
  Fun    <- match.fun(Fun)
  Dots <- list(...)
  DescTools::DoCall(applyAggregateDT, c(list(), list(x), Fun = Fun, FunStr = FunStr, By = By, Dots))

}
MyFun(as.data.table(mtcars[,c("cyl", "mpg", "disp")]), Fun = mean, By = "cyl", na.rm = TRUE)

}
}
\references{
\cite{period.apply on xts object takes about 50x longer than running it on it's core data and recasting. #278 \url{https://github.com/joshuaulrich/xts/issues/278}}

\cite{[R-Forge #5754] GForce functions and row- + col-wise operations on .SD #523 \url{https://github.com/Rdatatable/data.table/issues/523}}
}
