% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary_predictive.R
\name{as.quantmod.data.frame}
\alias{as.quantmod.data.frame}
\title{data(sample_matrix)
sample_xts <- as.xts(sample_matrix)
quantmodSample <- as.quantmod(as.data.frame(sample_xts), outcomename = "Close", order.by = index(sample_xts))}
\usage{
\method{as.quantmod}{data.frame}(x, outcomename, order.by, na.rm = TRUE,
  ...)
}
\description{
data(sample_matrix)
sample_xts <- as.xts(sample_matrix)
quantmodSample <- as.quantmod(as.data.frame(sample_xts), outcomename = "Close", order.by = index(sample_xts))
}
