% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{as.quantmod.default}
\alias{as.quantmod.default}
\alias{as.quantmod}
\alias{as.quantmod.data.frame}
\title{converts to a quantmod object}
\usage{
\method{as.quantmod}{default}(x, outcomename, order.by, na.rm = TRUE,
  ...)

as.quantmod(x, outcomename, order.by, na.rm = TRUE, ...)

\method{as.quantmod}{data.frame}(x, outcomename, order.by, na.rm = TRUE,
  ...)
}
\description{
converts to a quantmod object

data(sample_matrix)
sample_xts <- as.xts(sample_matrix)
quantmodSample <- as.quantmod(as.data.frame(sample_xts), outcomename = "Close", order.by = index(sample_xts))
}
