% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{AssignEnv}
\alias{AssignEnv}
\title{extract a list of R objects and assign to an environment}
\usage{
AssignEnv(List, nms = NULL, envir = parent.frame())
}
\arguments{
\item{List}{collection of R objects}

\item{nms}{names of List object (default: everything)}

\item{environment}{to assign R objects}
}
\value{
invisible
}
\description{
mostly for 'space saving' and cleanliness
}
\examples{
\dontrun{
List <- list(a=1, b=2)
# a, b
AssignEnv(List, c("a","b"))
}
}
