% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{balanceByUBLData}
\alias{balanceByUBLData}
\title{create/remove more or less observations determined by an UBL function}
\usage{
balanceByUBLData(
  xTs,
  Fmla,
  TrainStart = NULL,
  TrainEnd = NULL,
  UBLFunction = NULL,
  ...
)
}
\arguments{
\item{xTs}{xts object of training data}

\item{Fmla}{formula that is sent to the UBL function}

\item{TrainStart}{absolute training start date}

\item{TrainEnd}{absolute training end date}

\item{UBLFunction}{package UBL *Regress function ( default is ImpSampRegress ),
entered as enclosed in a "string" or bare}

\item{...}{passed to the UBL function
if the UBLFunction is ImpSampRegress, then defaults are the following:
thr.rel = 0.5,  C.perc = list(1.0, 2.5),
rel: values less than zero are important, values greater than zero are not important}
}
\description{
\preformatted{

Produce more or less duplicate/near observations.

}
}
