% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{balanceFocusedData}
\alias{balanceFocusedData}
\title{give more observations to the focused data}
\usage{
balanceFocusedData(
  xTs,
  NumbReplicaCopiesMultiple = NULL,
  SlicesFocusedData,
  SlicesAllData,
  TrainStart,
  TrainEnd
)
}
\arguments{
\item{xTs}{xts object of training data}

\item{NumbReplicaCopiesMultiple}{NULL(default) How much "more" to to
overbalance the "focused data" compareed to its corresponding "all data"
One(1) is the same as NULL(default).
Two(2) or any other fraction, e.g. 2.5 means multiply the internally derived
"number of replica copies" by this value.}

\item{SlicesFocusedData}{list of index slice date vectors}

\item{SlicesAllData}{list of index slice date vectors}

\item{TrainStart}{absolute training start date}

\item{TrainEnd}{absolute training end date}
}
\description{
\preformatted{

Produce more focused data duplicate observations
such that "focused data" is at least the same number of observations
as "all observations less focused observations"

}
}
