% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{buildData}
\alias{buildData}
\title{create one data object from multiple sources}
\usage{
buildData(formula, na.rm = TRUE, return.class = "zoo",
  source.envir = NULL, ...)
}
\arguments{
\item{formula}{as quantmod buildData}

\item{na.rm}{as quantmod buildData}

\item{return.class}{as quantmod buildData}

\item{source.envir}{find xts Symbols in this environment}
}
\value{
"return.class" as quantmod buildData
}
\description{
\preformatted{

Applying transformations via standard R formula mechanism.

UNUSED
See quantmod buildData.
Also, this function can also acquire symbols
that are stored in an environment.

}
}
