% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{buildModel.train}
\alias{buildModel.train}
\title{S3-like method buildModel.train}
\usage{
buildModel.train(quantmod, training.data, ...)
}
\description{
# tuneGrid ( production tester )
tg <- expand.grid(
  nrounds   =  10, # TEN TREES
  eta       =  c(0.1,0.01),
  max_depth =  c(4,6,8,10),
  gamma     =  0,
  colsample_bytree = c(1,0.5),
  min_child_weight = 1,
  subsample        = c(1,0.5)
)
# tuneGrid ( non-production tester )
tg <- expand.grid(
  nrounds   =  50, # TEST 10 trees - DEV 50 trees - OTHER 500 trees
  eta       =  c(0.1,0.01),
  max_depth =  c(4,7,10),
  gamma     =  0,
  colsample_bytree = 1,
  min_child_weight = 1,
  subsample        = 1
)
}
\details{
cl <- parallel::makeCluster(parallel::detectCores(logical = FALSE)) 
doParallel::registerDoParallel(cl)
assign("cl", cl, envir = asNamespace("base")); rm(cl)

# default
# note: caret::trainControl(. . . , allowParallel = TRUE)
UnRateMachinetradeModel()

parallel::stopCluster(cl)
doParallel::stopImplicitCluster()

tc <- caret::trainControl(method = "cv", number = 5)

TODO [ ] : fully WORKED EXAMPLE: combine WITH below
builtModel <- buildModel(specmodel, method="train", training.per=c("1970-12-31","2006-12-31")
  , method_train = "xgbTree", tuneGrid = tg, trControl = tc)
}
