% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{CASHClms}
\alias{CASHClms}
\title{get the CASH columns}
\usage{
CASHClms(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
column names
}
\description{
\preformatted{

MAY NOT BE USED ANYMORE?

}
}
