% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{cashLogReturns}
\alias{cashLogReturns}
\title{cash log returns (cashlogrets)}
\usage{
cashLogReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with the same index as xTs
}
\description{
cash log returns (cashlogrets)
}
