% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{cashReturns}
\alias{cashReturns}
\title{cash APC returns (CASHapcrets)}
\usage{
cashReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object (only takes the index)}
}
\value{
xts object with the same index as xTs
}
\description{
\preformatted{

all SYMBOLS must lead and be UPPERCASE

}
}
