% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{cashWts}
\alias{cashWts}
\title{cash APC weights}
\usage{
cashWts(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

Currently, this number is just only all of the other weight
columns (*_wts) subtracted from one(1)

}
}
