% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{combineLogReturns}
\alias{combineLogReturns}
\title{join two Log Returned xts objects}
\usage{
combineLogReturns(xTs = NULL, xTs1 = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{xTs1}{xts object to merge into xTs}
}
\value{
xts object
}
\description{
join two Log Returned xts objects
}
