% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{combineXts}
\alias{combineXts}
\title{join two xts objects}
\usage{
combineXts(xTs = NULL, xTs1 = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{xTs1}{xts object to merge into xTs}
}
\value{
xts object
}
\description{
\preformatted{

Handles the edge case: if BOTH have no coredata
(then merge.xts produces ONE no-data no-index xts object)
then instead, preserve the indexes.

}
}
\examples{
\dontrun{
> combineXts( initXts(NULL),xts(,zoo::as.Date(0)) )
Data:
numeric(0)

Index:
 Date[1:1], format: "1970-01-01"
> combineXts( xts(,zoo::as.Date(1)), xts(,zoo::as.Date(0)) )
Data:
numeric(0)

Index:
 Date[1:2], format: "1970-01-01" "1970-01-02"
}
}
