% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{dfize}
\alias{dfize}
\alias{xtsize}
\title{safely make a xts into a df and safely make df into an xts}
\usage{
dfize(xtso)

xtsize(dfo)
}
\description{
library(xts)
sample_xts <- as.xts(sample_matrix)
str(head(xtsize(dfize(sample_xts))))
}
\details{
edge case tests

xts1 <- xts(NA_real_, zoo::as.Date(0))
colnames(xts1) <- "col1"
xts1 <- xts1[0,]

exiting from: dfize(xts1)
'data.frame':   0 obs. of  2 variables:
 $ datetime:Class 'Date'  num(0)
 $ col1    : num

> xtsize(dfize(xts1))
     col1
> str(xtsize(dfize(xts1)))
An 'xts' object of zero-width
> index(xtsize(dfize(xts1)))
[1] "Date of length 0"
}
