% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{Coredata}
\alias{Coredata}
\title{garentees return value value is a 'matrix of at least one dimension'or NULL}
\usage{
Coredata(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
matrix
}
\description{
handles the edge case of coredata(<vector of element of size 1 or size 0>)
meant specifically to input empty coredata data into xts(, index)
xts(,index) only accepts a non zero-dimension matrix or a NULL
}
\examples{
\dontrun{
# > require(xts)
# > Coredata(NULL)
# NULL
#
# > Coredata(numeric(0))
# NULL
#
# > Coredata(11:13)
#      [,1]
# [1,]   11
# [2,]   12
# [3,]   13
}
}
