% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{Current}
\alias{Current}
\title{current}
\usage{
Current(xTs = NULL, Shift = NULL, NoColNameChange = NULL)
}
\description{
\preformatted{

Pads beginning date as necessary.

Get the current month near eom date.

}
}
