% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{currentCashLogReturns}
\alias{currentCashLogReturns}
\title{current cash log returns (CASHlogrets)}
\usage{
currentCashLogReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object (only takes the index)}
}
\value{
leading xts object
}
\description{
current cash log returns (CASHlogrets)
}
