% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{currentCashReturns}
\alias{currentCashReturns}
\title{current cash APC returns (CASHapcrets)}
\usage{
currentCashReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object (only takes the index)}
}
\value{
leading xts object
}
\description{
\preformatted{

}
}
