% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{currentRetsClms}
\alias{currentRetsClms}
\title{get the currentrets columns}
\usage{
currentRetsClms(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
column names
}
\description{
\preformatted{

}
}
