% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{currentSymbolReturns}
\alias{currentSymbolReturns}
\title{get the Symbol current APC returns}
\usage{
currentSymbolReturns(Symbol = NULL, src = NULL,
  SymplifyGeneratorFUN = NULL, ReturnsGeneratorFUN = NULL)
}
\value{
xts object
}
\description{
\preformatted{

}
}
