% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{currentWilshire5000LogReturns}
\alias{currentWilshire5000LogReturns}
\title{get the Wilshire 5000 Index current log returns}
\usage{
currentWilshire5000LogReturns()
}
\value{
xts object
}
\description{
\preformatted{

}
}
