% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{customSorting}
\alias{customSorting}
\title{custom sort a vector}
\usage{
customSorting(Vector, InitOrder, CI = FALSE)
}
\arguments{
\item{Vector}{vector to be sorted}

\item{InitOrder}{starting custom sorting ( without the excess )}

\item{CI}{case insensitive}
}
\description{
excess is appended to the end ( CI sort or CS sort )
}
\examples{
\dontrun {
customSorting( c("a","v", "E2", "c","l", "e3" ,"h","o","date"), InitOrder = c("date", "o", "h", "l", "c", "v", "a"), CI = TRUE )
}
}
