% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{daysAfterEvent}
\alias{daysAfterEvent}
\title{days after the event}
\usage{
daysAfterEvent(x, dates)
}
\arguments{
\item{x}{xts object}

\item{dates}{vector of Dates of the events}
}
\value{
xts object with one new column
}
\description{
\preformatted{

}
}
\examples{
\dontrun{

# S&P 500
gspc <- To.Monthly(Cl(getSymbols("^GSPC", auto.assign = FALSE, from = "1900-01-01")), indexAt = "lastof", OHLC = F)
colnames(gspc) <- "GSPC"
# returns
gspc.ret <- TTR::ROC(gspc, type = "discrete")

# top 20 worst monthly returns in the S&P500 history
CrashDates <- zoo::as.Date(rownames(head(as.matrix(gspc.ret)[order(as.matrix(gspc.ret)),1, drop = F], 20)))

daysAfterEvent(gspc.ret, dates = CrashDates)

}
}
