% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{DBDF2CREATETableStmt}
\alias{DBDF2CREATETableStmt}
\title{from column names and datatypes, make a CREATE TABLE statement}
\usage{
DBDF2CREATETableStmt(df, con, Symbol, schname)
}
\arguments{
\item{df}{data.frame (with column names)}

\item{con}{DBI database connection}
}
\description{
from column names and datatypes, make a CREATE TABLE statement
}
