% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{delaySinceLastObs}
\alias{delaySinceLastObs}
\title{get the number of consecutive NA values since the last non-NA observations}
\usage{
delaySinceLastObs(x)
}
\arguments{
\item{x}{vector or single column zoo/xts object}
}
\value{
vector of distances from the last non-NA observation
}
\description{
\preformatted{

At one time imported from the archived package/function rowr rowApply

}
}
\examples{
\dontrun{

library(xts)

delaySinceLastObs(c(101,NA,NA,NA,102,NA,NA))
[1] 0 1 2 3 0 1 2

delaySinceLastObs(xts(matrix(c(101,NA,NA,NA,102,NA,NA), ncol = 1), zoo::as.Date(0:6)))
[1] 0 1 2 3 0 1 2

}
}
