% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{DetectEmptyRows}
\alias{DetectEmptyRows}
\title{detect rows with all NA vaues}
\usage{
DetectEmptyRows(x, Var = NULL)
}
\arguments{
\item{x}{data.frame}

\item{Var}{columns determining complete cases
Can be column names or positions}
}
\value{
df with rows of all-NA removed
}
\description{
MOVE ME
complete.cases does not work on tibbles!
Function is written in DataCombine style
}
