% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{DetectOnlyNonEmptyVarsInRows}
\alias{DetectOnlyNonEmptyVarsInRows}
\title{detect columns(Var) where the values are not-NA but
the rest of the columns are NA everywhere else}
\usage{
DetectOnlyNonEmptyVarsInRows(x, Var = NULL)
}
\arguments{
\item{x}{data.frame}

\item{Var}{columns determining complete cases.
Can be column names or positions}
}
\value{
df with rows of all-NA removed
}
\description{
Function is written in DataCombine style
}
