% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{dfGetPKEYNames}
\alias{dfGetPKEYNames}
\title{determine the column names of a data.frame's key columns}
\usage{
dfGetPKEYNames(df, keys)
}
\arguments{
\item{df}{data.frame sent to function dfToCREATETable}

\item{keys}{if passed NULL or 1 then the first column is returned.
if passed a numeric vector then those column by ordinal position are chosen.
if passed "NONE" then a zero length character vector is returned.
if passed a character vector then those are the primary key column.
names themselves}
}
\value{
character vector of column names that make up the PRIMARY KEY
}
\description{
\preformatted{

This is a friendly function with dfToCREATETable.
This convenience function guesses the column names of the
table's PRIMARY KEY columns.

}
}
