% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{doMachineWts}
\alias{doMachineWts}
\title{determine weights using Machine learning}
\usage{
doMachineWts(xTs = NULL, ModelFormula = NULL, ...)
}
\arguments{
\item{xTs}{xts object}

\item{...}{dots passed to further methods balanceFocusedData(parameters),
balanceByUBLData(dots), weightRankings(parameters)}

\item{timeSliceMatrix}{NULL(default) passed through dots.
Input format is the same as "tis::nberDates()". See. ? tis::nberDates.}

\item{NumbReplicaCopiesMultiple}{in dots passed to NumbReplicaCopies
This means how much more ( e.g. "focused" data ) is replicated
compared to "all" data.}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

This is the workhorse function. This is where the magic/logic happens.
Use any other columns (called indicators) that do not have the weights (_wts)
suffix and do not have the same root name compared to each and every
other *_wts column.

}
}
