% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{doOneChoice}
\alias{doOneChoice}
\title{determine the investment choice from the results of a Machine learning}
\usage{
doOneChoice(Fitted, ModelTarget_wts = NULL, ExtremePct = NULL)
}
\arguments{
\item{Fitted}{previously fitted xts object}

\item{ModelTarget_wts}{NULL(default) character name
of the predictee.  Usually, expected to have its name
end in "_wts".}

\item{ExtremePct}{"25"(default) percent passed to quantile.
This represents the low percent of the time that the investor
is "not doing the (good) One Choice".
I.e. (bad) times. E.g., recession, or crashes.}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

Given a Fitted model (xts object).  Determine
whether or not to buy in my favored asset (may by the stock market)
xor to buy in my less favored asset (may be cash).

}
}
