% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{eomData}
\alias{eomData}
\title{get the end of month data}
\usage{
eomData(
  Symbol = NULL,
  src = NULL,
  SymplifyGeneratorFUN = NULL,
  CarryForward = NULL
)
}
\arguments{
\item{Symbol}{getSymbols Symbol}

\item{src}{getSymbols source}

\item{SymplifyGeneratorFUN}{Function that Formats the output.}

\item{CarryForward}{default(NULL) after simplification perform
carry forward method(if any)
of data to replace future missing values.
Typical values may be the following. "NA.LOCF"; carry forward the last NA.
"NA.Shift.EOLM"; shift the entire series up (to the future) to the
last day of the previous month.}
}
\value{
xts object
}
\description{
\preformatted{

}
}
