% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{eomIndex}
\alias{eomIndex}
\alias{eomIndex.default}
\alias{eomIndex.Date}
\alias{eomIndex.xts}
\title{change the index date to the future}
\usage{
eomIndex(x = NULL)

\method{eomIndex}{default}(x = NULL)

\method{eomIndex}{Date}(x = NULL)

\method{eomIndex}{xts}(x = NULL)
}
\arguments{
\item{x}{date object}
}
\value{
xts object or data object with the index moved up
}
\description{
\preformatted{

}

\preformatted{

}
}
\examples{
\dontrun{
# > require(xts)
# > eomIndex(zoo::as.Date("1970-01-12"))
# [1] "1970-01-31"
}
\dontrun{
# > require(xts)
# > xTs <- xts(, zoo::as.Date("1970-01-12"))
# > eomIndex(xTs)
# Data:
# numeric(0)
#
# Index:
#  Date[1:1], format: "1970-01-31"
}
}
