% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{ErrorHandler}
\alias{ErrorHandler}
\title{error handler function meant for tryCatchLog::tryCatchLog}
\usage{
ErrorHandler(e, useENVI = getOption("useENVI"))
}
\arguments{
\item{e}{input from
tryCatchLog::tryCatchLog(. . . , error = function(e) { ErrorHandler(e) })"}

\item{useENVI}{getOption("useENVI") option.
If NULL or TRUE, will use get("envi") to search for
the calling environment location that is stored in "envi".
Otherwise, will just get the calling environment from parent.frame(8).}
}
\description{
\preformatted{

}
}
