% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{existSymbols.cache}
\alias{existSymbols.cache}
\title{xts object symbols}
\usage{
existSymbols.cache(Symbols = NULL, source.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{source.envir}{location of xts objects}

\item{...}{passed unused}
}
\value{
named character vector of Symbols
}
\description{
xts object symbols
}
