% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{existSymbols}
\alias{existSymbols}
\title{xts object symbols}
\usage{
existSymbols(Symbols = "", source.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{source.envir}{location of xts objects}

\item{...}{passed to trg}

\item{src}{source program to look for symbols}
}
\value{
character vector of Symbols
}
\description{
looks in source.envir xor "src" exclusively
}
