% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{explodeXts}
\alias{explodeXts}
\title{expland out xts}
\usage{
explodeXts(xTs1 = NULL, xTs2 = NULL, Fun = NULL, Whiches = NULL,
  AltName = NULL, Prefix = NULL, FixedSep = NULL, quote = FALSE,
  envir = parent.frame(2), ...)
}
\arguments{
\item{xTs1}{xts object}

\item{xTs2}{xts object}

\item{Fun}{function name in the "bare" or in literal quotes("")}

\item{Whiches}{list of possible varying parameters that are expanded
to all possible combinations by expand.grid}

\item{AltName}{string alternate name for "Fun"}

\item{Prefix}{boolan default is FALSE.  TRUE would place the column meta before the column name}

\item{FixedSep}{string divider of meta items}

\item{quote}{boolean passed to DescTools DoCall}

\item{envir}{calling environment}
}
\value{
new xts object of new derived columns
}
\description{
from an xts function stub, create an ets object of derived columns
}
\details{
Meant to create many package TTR technical trading column results
CURRENTLY NOT USED

Idea was from

Time series cross-validation 5
January 24, 2013
By Deane-Mayer
http://www.r-bloggers.com/time-series-cross-validation-5/
http://moderntoolmaking.blogspot.com/2013/01/time-series-cross-validation-5.html?utm_source=feedburner&utm_medium=feed&utm_campaign=Feed%3A+ModernToolMaking+%28Modern+Tool+Making%29

NOTE: if any xTs2, then xTs1 and xTs2 are paired/matched column position to column position)
}
\examples{
\dontrun{

# require(quantmod)
# ibm <- getSymbols("IBM", from = "1970-01-01", to = "1970-01-13", auto.assign = FALSE)
# explodeXts(ibm[,c("IBM.Open","IBM.Close")], Fun = "TTR::SMA", Whiches = list(n = 2:3))
}
}
