% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{extremesXts}
\alias{extremesXts}
\title{smallest and largest}
\usage{
extremesXts(x, br = 0:1, n = 1, ...)
}
\arguments{
\item{x}{xts object}

\item{br}{backrange: choose -1:0 (or less) to look into the future}

\item{n}{sum of the results of observations found}
}
\description{
Within a range, find the the exteme one-sided observations.
WORK IN PROGRESS
}
