% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ForecastersReleaseDates}
\alias{ForecastersReleaseDates}
\title{Philadelphia Fed Survey of Professinal Forecasters release dates}
\usage{
ForecastersReleaseDates(env = parent.frame(), MaxAge = NULL, ...)
}
\arguments{
\item{env}{where to create objects. (.GlobalEnv)}

\item{MaxAge}{"4 hours"(default) is longest age allowed, such that the retrieving the
the MOST RECENTLY PUBLISHED FORECASTERS RELEASE DATES FILE from  the local fileystem
will be allowed to be done. If the MaxAge is exceeded then,
the most recent release dates file is refreshed anew from the Philadelphia Fed site.
The format uses as.difftime: "# secs", "# mins", "# hours", "# days", "# weeks"}

\item{...}{dots passed to quantmod___try.download.file}

\item{force}{FALSE(default) re-download data from the Philadelphia Fed ( See the examples. )
Generally, using this parameter "force = TRUE"
is NOT necessary: after MaxAge has been exceeded and if a query needs the
MOST RECENTLY PUBLISHED FORECASTERS RELEASE DATES FILE then this file will be downloaded anew.}
}
\value{
data.frame of YearQtr (yearqtr) TrueDeadlineDate (Date) ReleaseDate (Date)
}
\description{
\preformatted{

About the date/time index . . .

This is the Deadline and Release Dates for the
Survey of Professional Forecasters.

True deadline and news release dates for surveys prior to 1990:Q2 are not known.

*The 1990Q2 survey was not taken in real time, because the Philadelphia Fed
had not yet taken over the survey. Forecasters were asked to provide dated
forecasts from May 1990.

**The 1996Q1 survey was delayed because of the federal government shutdown,
which in turn delayed the release of government statistical data.

***The 2013Q4 survey was delayed because of the federal government shutdown,
which in turn delayed the release of government statistical data.

}
}
\examples{
\dontrun{
ForecastersReleaseDates()
ForecastersReleaseDates(force = TRUE)
}
}
\references{
\cite{Survey of Professinal Forecasters Release Dates \url{https://www.philadelphiafed.org/-/media/research-and-data/real-time-center/survey-of-professional-forecasters/spf-release-dates.txt?la=en}}
}
