% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{fredData}
\alias{fredData}
\title{get data from the St. Louis FRED}
\usage{
fredData(Symbol = NULL, New = NULL, NewMaxAge = NULL)
}
\arguments{
\item{Symbol}{FRED symbol as a string}
}
\value{
xts object of results
}
\description{
get data from the St. Louis FRED
}
\examples{
\dontrun{
EXAMPLE
# > head(fredData("GDP"),1)
#                           gdp
# 1947-01-01 243.16399999999999
}
}
