% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{getModelData}
\alias{getModelData}
\title{NEED YET: e.g. initLag: index(X1) - #, Less: X2 - X1, LogDiff: log(X2/X1), FracChg: X2/X1 - 1, ArithDiff: X2/X1
quantmod getModelData with parameter source.envir}
\usage{
getModelData(x, na.rm = TRUE, source.envir = NULL, ...)
}
\description{
\preformatted{

Get some Symbols from an environment (source.envir).
  Will search first in (source.envir).
  If the Symbol is not found in the enviroment (source.envir),
  then get the Symbol from elsewhere
...
Passed to getSymbols . . .
        useful: from, to
  maybe useful: src
  maybe useful: set per Symbol src using setSymbolLookup

specifyModel(getModelData)
if na.rm == TRUE, then does 'na.exclude',
  BUT 'without rules' PUTS back (rbind) the last observation.

}
}
