% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getSymbols.AAII}
\alias{getSymbols.AAII}
\title{American Association of Individual Investors (AAII) weekly sentiment survey}
\usage{
getSymbols.AAII(Symbols, env, return.class = "xts", ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded
Possible Symbols are the following:
"SENTIMENT" (means get all of the columns);
otherwise, get specific columns;
"Bullish", "Neutral", "Bearish",
"Bullish8WMA", "BullBearSpread",
"BullishAvg", "BullishAvgPStd", "BullishAvgNStd",
"SP500WHigh", "SP500WLow", and "SP500WClose"}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{desired class of returned object.
Can be xts, zoo, data.frame, or xts (default)}

\item{...}{additional parameters}

\item{force}{re-download data from AAII ( not implemented yet )}
}
\value{
A call to getSymbols.aaii will load into the specified
environment one object for each \code{Symbol} specified,
with class defined by \code{return.class}.
}
\description{
Data represents what direction members feel the
stock market will be in next 6 months.
}
\details{
The sentiment survey measures the percentage of individual investors who are
bullish, bearish, and neutral on the stock market short term;
individuals are polled from the AAII Web site on a weekly basis.
Only one vote per member is accepted in each weekly voting period.
}
\examples{
\dontrun{

getSymbols("SENTIMENT", src = "AAII")
getSymbols("BullBearSpread", src = "AAII")
}
}
\references{
\cite{AAII Investor Sentiment Survey \url{https://www.aaii.com/sentimentsurvey}}

\cite{Sentiment Survey Past Results \url{https://www.aaii.com/sentimentsurvey/sent_results}}
}
\seealso{
\code{\link{getSymbols}}
\code{\link{setSymbolLookup}}
}
\author{
Jeffrey A. Ryan

Andre Mikulec (adapted original code to work with AAII sentiment data)
}
\keyword{data}
