% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getSymbols.cache}
\alias{getSymbols.cache}
\title{downloads Symbols to specified env from a local R environment}
\usage{
getSymbols.cache(Symbols = NULL, env, return.class = "xts",
  cache.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{class of returned object}

\item{...}{additional parameters}

\item{source.envir}{where to find xts objects ( location of cache )}
}
\value{
A call to getSymbols.csv will load into the
specified environment one object for each Symbol specified,
with class defined by return.class. Presently this may be
ts, zoo, xts, data.frame, or timeSeries.
}
\description{
\preformatted{

}
}
\examples{
\dontrun{

msft <- getSymbols("MSFT", src = "yahoo", auto.assign = FALSE)

source.envir = list2env(list(MSFT = msft))
saveSymbols(trg = "cache", source.envir = source.envir)
msft2 <- getSymbols(Symbols = "MSFT", src = "cache", auto.assign = F)

}
}
