% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getSymbols.FRED2}
\alias{getSymbols.FRED2}
\title{economic data from the St. Louis FRED}
\usage{
getSymbols.FRED2(Symbols, env, from = "1800-01-01", to = Sys.Date(),
  Edit = NULL, return.class = "xts", ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{env}{where to create objects. (.GlobalEnv)}

\item{from}{data end date in format  "YYYY-MM-DD"}

\item{to}{data start date in format "YYYY-MM-DD"}

\item{Edit}{list of names items or named character vectors of Edit options. (See the example).}

\item{return.class}{desired class of returned object.
Can be xts, zoo, data.frame, or xts (default)}

\item{...}{additional parameters}
}
\value{
A call to getSymbols.YaleU will load into the specified
environment one object for each \code{Symbol} specified,
with class defined by \code{return.class}.
}
\description{
Alternative to the quantmod::getSymbols.FRED
}
\details{
The St. Louis FRED's 'data' date and 'published data' data
and 'frequency' of data differs from "yahoo" data
(Yahoo data dates are the same day.).

This an en extension of quantmod::getSymbols.FRED.
Extra attibutes (e.g. 'publish date': Last_Updated )
are added to the xts object (if returned as an xts object
(recommended) and default).

Also, an extra xts column SYMBOL_PUBDLY  'publication dalay'
is returned giving the number of days since publication.
These publication dates are estimated
based on the current 'publication date' and data Frequency.
(Some Thought had been put into 'instead querying
ALFRED', but that path choice utimately was not taken.)

The user may sent 'from' and 'to' dates.

Throught the 'Edit' parameter, the  user may pass
in data transformation formulas.

For this apply either of the following must be true.
(1) The foruma query paramter formula(fml)
has just one veriable "a" and then this Edit transformation
is applied across all Symbols sent by the user. Or,
(2), if the number of Symbols is two(2) or greater then
then the number of symbols must equal to the number
of variables in the query paramter e.g. if the Symbols
sent are "WILL5000IND","BAMLC0A4CBBBEY" in
getSymbols(c("WILL5000IND","BAMLC0A4CBBBEY") then the
number of paraters must be two(2) and these are
letters of the aphabet starting with "a" and must be
lowercase. An example is 'fml="a-b"' (Se The Example.
}
\examples{
\dontrun{

# common usage
if(!exists("DGS3MO")) getSymbols("DGS3MO", src = "FRED2")

# genernal useage
getSymbols(c("DGS3MO", "GDP"), src = "FRED2")

# rare usage
# get the Equity Premium, see the references
getSymbols(c(c("WILL5000IND","BAMLC0A4CBBBEY")), src = "FRED2",
  Edit = c(transformation = c("pc1","pch"), fq = "Daily, Close", fml="a-b")  )

# reading text tables
Text2DF <- function(x) {
    read.csv(textConnection(x),
     sep = ":",
     strip.white = TRUE,
     colClasses = "character") -> x
     x
}


# Edit possible query parameters

Units: transformation
Modify frequency: fq
Aggregation method: fam
Formula: fml
[Final] Units: fgst

# <Time Series>: id

# Settable 'per' each individual Time Series
# Settable when two or more Time Series are chosen

# The first time series is (a).
# The second time series is (b). Etc. . . .
# The final time series the \\*result\\* of all of the individual
# time series and it is  \\*not\\* the last time series ( e.g. not "z")

# Underscores separate each Time Series from each other

Id <- vector(mode = "character")

# examples
# &id=WILL5000IND
# &id=WILL5000IND_BAMLC0A4CBBBEY

Id <- c("WILL5000IND","BAMLC0A4CBBBEY")



# Units: transformation
# Settable 'per' each individual Time Series
# Underscores separate each Time Series from each other

Units_transformation <-
"
  Code: Description
  lin: Index
  chg: Change, Index
  ch1: Change from Year Ago, Index
  pch: Percent Change
  pc1: Percent Change from Year Ago
  pca: Compounded Annual Rate of Change
  cch: Continuously Compounded Rate of Change
  cca: Continuously Compounded Annual Rate of Change
  nbd: Index (Scale value to 100 for chosen date)
"
Units_transformation <- Text2DF(Units_transformation)

# examples
# &transformation=lin
# &transformation=cca_pc1

Units_transformation <- c("cca","pc1")



# Modify frequency: fq
# Settable in the First Time Series
# Settable in the Final Time Series (removed from the First Time Series)
# Note: When "Daily, Close" is chosen,  Aggregation method is not available.

Modify_frequency_fq <-
"
  Code: Description
  Daily, Close: Daily, Close
  Weekly, Ending Friday: Weekly, Ending Friday
  Weekly, Ending Thursday: Weekly, Ending Thursday
  Weekly, Ending Wednesday: Weekly, Ending Wednesday
  Weekly, Ending Tuesday: Weekly, Ending Tuesday
  Weekly, Ending Monday: Weekly, Ending Monday
  Weekly, Ending Sunday: Weekly, Ending Sunday
  Weekly, Ending Saturday: Weekly, Ending Saturday
  Biweekly, Ending Wednesday: Biweekly, Ending Wednesday
  Biweekly, Ending Monday: Biweekly, Ending Monday
  Monthly: Monthly
  Quarterly: Quarterly
  Semiannual: Semiannual
  Annual: Annual
"
Text2DF(Modify_frequency_fq)

# example
# &fq=Annual
# &&fq=Daily\%2C\%20Close

Modify_frequency_fq <- "Annual"



# Aggregation method: fam
# Settable in the First Time Series
# Settable in the Final Time Series (removed from the First Time Series)

Aggregation_method_fam <-
"
  Code: Description
  avg: Average
  sum: Sum
  eop: End of Period
"
Aggregation_method_fam <- Text2DF(Aggregation_method_fam)

# example
# &fam=eop

Aggregation_method_fam <- "eop"



# Formula: fml

# Settable in the First Time Series
# Settable in the Final Time Series (removed from the First Time Series)

Formula_fml <- vector(mode = "character")[1]

# example
# &fml=a-b

Formula <- "a-b"



# [Final] Units: fgst

# Settable in the First Time Series
# Settable in the Final Time Series (removed from the First Time Series)

Final_Units_fgst <-
"
  lin:
  chg: Change
  ch1: Change from Year Ago
  pch: Percent Change
  pc1: Percent Change from Year Ago
  pca: Compounded Annual Rate of Change
  cch: Continuously Compounded Rate of Change
  cca: Continuously Compounded Annual Rate of Change
  log: Natural Log
  nbd: Index (Scale value to 100 for chosen date)
"
Final_Units_fgst <- Text2DF(Final_Units_fgst)

# example
# &fgst=log


}
}
\references{
\cite{The equity premium \url{https://fredblog.stlouisfed.org/2016/07/the-equity-premium/}}

\cite{customize the The equity premium [Edit Graph] \url{https://fred.stlouisfed.org/graph/?g=6LsS}}

\cite{FRED ECONOMIC DATA: FEDERAL RESEARVE BANK OF ST. LOUIS \url{https://fred.stlouisfed.org/}}

\cite{getSymbols('DTWEXB',src='FRED') broken #209 \url{https://github.com/joshuaulrich/quantmod/issues/209}}
}
\seealso{
\code{\link{getSymbols}}
\code{\link{setSymbolLookup}}
}
\author{
Jeffrey A. Ryan

Andre Mikulec (adapted original code to work with features)
}
\keyword{data}
