% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary_predictive.R
\name{getSymbols}
\alias{getSymbols}
\title{quantmod getSymbols with source.envir}
\usage{
getSymbols(Symbols = NULL, env = parent.frame(),
  reload.Symbols = FALSE, verbose = FALSE, warnings = TRUE,
  src = "yahoo", symbol.lookup = TRUE,
  auto.assign = getOption("getSymbols.auto.assign", TRUE),
  source.envir = NULL, ...)
}
\description{
get some Symbols from an environment (source.envir)
  will search first in (source.envir)
  if the Symbol is not found in the enviroment (source.envir),
  then get the Symbol from elsewhere
NOTE: do not do: "source.envir = e, "env = e"
      when auto.assign = TRUE(default), .GetSymbols is placed where "env = e"
}
\details{
e <- new.env()
NOT the same as "env = e"
assign("AAPL", getSymbols(list(AAPL = "yahoo"), auto.assign = F), envir = e)
ls.str(e)
AAPL : An 'xts' object on 2007-01-03/2018-04-24 containing:
  Data: num [1:2847, 1:6] 12.3 12 12.3 12.3 12.3 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:6] "AAPL.Open" "AAPL.High" "AAPL.Low" "AAPL.Close" ...
  Indexed by objects of class: [Date] TZ: UTC
  xts Attributes:
List of 2
 $ src    : chr "yahoo"
 $ updated: POSIXct[1:1], format: "2018-04-25 18:12:26"

getSymbols(list(AAPL = "yahoo", MSFT = "yahoo"), source.envir = e)
AAPLSymbol <- getSymbols(list(AAPL = "yahoo"), auto.assign = FALSE, source.envir = e)
}
