% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getSymbols.Simfa}
\alias{getSymbols.Simfa}
\title{bonds}
\usage{
getSymbols.Simfa(Symbols, env, return.class = "xts", ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded
Possible Symbols are the following:
"SimfaBondMarketData" (means get all of the columns);
otherwise, get specific columns;
TO BE NAMED LATER}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{desired class of returned object.
Can be xts, zoo, data.frame, or xts (default)}

\item{...}{additional parameters}

\item{force}{FALSE(default) re-download data from Simfa ( See the examples. )
The hidden variable ".SimfaBondMarketData_path2file" located in the
environment of parameter env is the last known location of the "xls" file.
Generally, using this parameter "force = TRUE"
is NOT necessary: the hidden variables are persistent through the
entire R session.  Also, send parameter "verbose = TRUE" to see
the *path2file location.}
}
\value{
A call to getSymbols.Simfa will load into the specified
environment one object for each \code{Symbol} specified,
with class defined by \code{return.class}.
}
\description{
US Corporate Bond Issuance
Bonds Outstanding
  (Upper Right Corner)
  DOWNLOAD: JPG|CSV|XLS ( automation: need phantom js ) **HERE**
    XLS: "Issuance" tab ( has data through the previous month)
# XLS bottom ( SEE DEC 31 2018: EXACT SAME AS THE ARTICLE )
         Year	M	 Municipal 	 Treasury 	 Mortgage-Related 	 Corporate Debt 	 Federal Agency Securities 	 Asset-Backed 	 Total
YTD '17
YTD '18
% Change
}
\details{
https://www.sifma.org/wp-content/uploads/2017/08/Corporate-US-Corporate-Issuance-SIFMA.xls
US Corporate Bond Issuance
https://www.sifma.org/resources/research/us-corporate-bond-issuance/
Securities Industry and Financial Markets Association
SIFMA
https://www.sifma.org/

referred by Jeff Cox
AND
Christopher Whalen
twitter handle: rcwhalen

When the Bid Goes to Zero |
https://goo.gl/2tUeJp  |
https://twitter.com/rcwhalen/status/1072619534819971072
AND
When the Bid Goes to Zero
December 11, 2018
By: R. Christopher Whalen
https://www.theinstitutionalriskanalyst.com/single-post/2018/12/11/When-the-Bid-Goes-to-Zero
}
\examples{
\dontrun{

# common usage
if(!exists("NCInvGrd")) getSymbols("NCInvGr", src = "Simfa")

getSymbols(c("NCInvGrd", "NCHighYld"), src = "Simfa")

# force requery of data from the Simfa site
# will collect one new xls (historical obs) file
getSymbols(c("NCInvGrd", "NCHighYld"), src = "Simfa", force = TRUE)

# all columns in one xts object
getSymbols("SimfaBondMarketData", src = "Simfa")

}
}
\references{
\cite{US Corporate Bond Issuance \url{https://www.sifma.org/resources/research/us-corporate-bond-issuance/}}

\cite{R. Christopher Whalen - When the Bid Goes to Zero \url{https://www.theinstitutionalriskanalyst.com/single-post/2018/12/11/When-the-Bid-Goes-to-Zero}}
}
\seealso{
\code{\link{getSymbols}}
\code{\link{setSymbolLookup}}
}
\author{
Jeffrey A. Ryan

Andre Mikulec (adapted original code to work with Simfa Bond market data)
}
\keyword{data}
