% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{getSymbols.UMich}
\alias{getSymbols.UMich}
\title{University of Michigan: Consumer Sentiment}
\usage{
getSymbols.UMich(Symbols, env, return.class = "xts", ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded
Possible Symbols are the following:
"UMichSentiment" and "UMCSENT". This currently the same data.}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{desired class of returned object.
Can be xts, zoo, data.frame, or xts (default)}

\item{...}{additional parameters}

\item{force}{FALSE(default) re-download data from FRED(historical)
and the University of Michigan (current/last_known) ( See the examples. )
The hidden variable ".UMichSentiment_path2file" located in the
environment of parameter env is the last known location of the "html" file.
Generally, using this parameter "force = TRUE"
is NOT necessary: the hidden variables are persistent through the
entire R session.  Also, send parameter "verbose = TRUE" to see
the *path2file location.}
}
\value{
A call to getSymbols.AAII will load into the specified
environment one object for each \code{Symbol} specified,
with class defined by \code{return.class}.
}
\description{
From FRED (historical) and briefing.com (recent) University of
Michigan: Consumer Sentiment voting results data
}
\examples{
\dontrun{

# common usage
if(!exists("UMCSENT")) getSymbols("UMCSENT", src = "UMich")

getSymbols(c("UMCSENT"), src = "UMich")

# force requery of data from the UMich site
# will collect one new FRED (historical obs) data and one new html (latest obs) file
getSymbols(c("UMCSENT"), src = "UMich", force = TRUE)

# all columns in one xts object
getSymbols("UMichSentiment", src = "UMich")

}
}
\references{
Recent data is from
Univ. of Michigan Consumer Sentiment - Prelim:
\cite{Univ. of Michigan Consumer Sentiment - Prelim \url{https://www.briefing.com/investor/calendars/economic/releases/mich.htm}}

Original data is from here:
\cite{surveys of consumers UNIVERSITY OF MICHIGAN \url{https://data.sca.isr.umich.edu/tables.php}}

Historical data is from here.
At the request of the source, the data is delayed by 1 month:
\cite{University of Michigan: Consumer Sentiment \url{https://fred.stlouisfed.org/series/UMCSENT}}

\cite{University of Michigan: Consumer Sentiment raw data \url{https://fred.stlouisfed.org/data/UMCSENT.txt}}
}
\seealso{
\code{\link{getSymbols}}
\code{\link{setSymbolLookup}}
}
\author{
Jeffrey A. Ryan

Andre Mikulec (adapted original code to work with AAII sentiment data)
}
\keyword{data}
