% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getSymbols.USFedPhil}
\alias{getSymbols.USFedPhil}
\title{Survey of Professional Forecasters data from the Philadelphia FED}
\usage{
getSymbols.USFedPhil(Symbols, env, return.class = "xts",
  SurveyDecades = "All", DataPath = "./USFedPhilForecastersData",
  MaxAge = NULL, Fun = mean, na.rm = TRUE, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded
To get all of the data use: "USFedPhilForecastingData".  Possible individual
Symbols are listed in the links on the Forcasters' main page
that reads (see below) ... (1) Historical Data Files for the Survey
of Professional Forecasters (2) Documentation: PDF Provides information on
all variables, transformations, and files in the survey.
E.g. "UNEMP2". Symbols will be returned
as (xts) data with names as excelcolumn+Fun"
E.g. "UNEMP2.mean"}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{desired class of returned object.
Can be xts, zoo, data.frame, or xts (default)}

\item{SurveyDecades}{"All"(default). Vector of options include the following
"1970s", "1980s", "1990s", "2000s", and "2010s".
Multiple options can be chosen and  sent in a character
vector. e.g. c("2000s", "2010s")}

\item{DataPath}{"./USFedPhilForecastersData"(default). New/old location of the
new/old downloaded xls files from the Philadelphia Fed's site. Windows
users can also enter this path using instead two back slashes.
Abolute paths can also be used.}

\item{MaxAge}{"4 hours"(default) is longest age allowed, such that the retrieving the
the MOST RECENTLY PUBLISHED FORECASTERS EXCEL FILE from  the local fileystem
will be allowed to be done. If the MaxAge is exceeded then,
the most recent excel file is refreshed anew from the Philadelphia Fed site.
The format uses as.difftime: "# secs", "# mins", "# hours", "# days", "# weeks"}

\item{Fun}{mean with na.rm = TRUE (default) Aggregate function to
apply to the period data. This becomes the symbol Suffix. E.g. "mean" will
produce the suffix "MEAN"}

\item{na.rm}{TRUE(default) pased to Fun. If Fun does not have na.rm parameter
or Fun does not have dots ... parameter then na.rm is not passed}

\item{...}{additional parameters}

\item{force}{FALSE(default) re-download data from USFedPhil ( See the examples. )
Generally, using this parameter "force = TRUE"
is NOT necessary: after MaxAge has been exceeded and if a query needs the
MOST RECENTLY PUBLISHED FORECASTERS EXCEL FILE then this file will be downloaded anew.}
}
\value{
A call to getSymbols.USFedPhil will load into the specified
environment one object for each \code{Symbol} specified,
with class defined by \code{return.class}.
}
\description{
# WORK IN PROGRESS
# WORK IN PROGRESS
}
\details{
# Popular start page
https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/data-files

# all of the data (1968 to the present) in five(5) excel files
Individual Forecasts for the Survey of Professional Forecasters
https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/historical-data/individual-forecasts

# Some, compare and constrasts may be
'forcasters'(prediction) w.s. FRED (acutally what happened

# unemployment

Civilian Unemployment Rate (UNEMP)
https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/data-files/unemp

Civilian Unemployment Rate (UNRATE)
https://fred.stlouisfed.org/series/UNRATE/

# prices ( e.g. inflation )

Price Index for Gross National Product/Gross Domestic Product (PGDP)
https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/data-files/pgdp

Gross National Product (chain-type price index) (A001RV1Q225SBEA)
https://fred.stlouisfed.org/series/A001RV1Q225SBEA

# gross domestic product

Nominal Gross National Product/Gross Domestic Product (NGDP)
https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/data-files/ngdp

Gross Domestic Product (GDP)
https://fred.stlouisfed.org/series/GDP

# Corporate Profits

Corporate Profits After Tax (CPROF)
https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/data-files/cprof

Corporate Profits After Tax (without IVA and CCAdj) (CP)
https://fred.stlouisfed.org/series/CP

NOTE: (for simplicity) xts time index is the middle data of the calendar quarter

release dates UNIMPLEMENTED
https://www.philadelphiafed.org/-/media/research-and-data/real-time-center/survey-of-professional-forecasters/spf-release-dates.txt?la=en

NOT USED ANYWHERE
}
\examples{
\dontrun{

# common usage
if(!exists("UNEMP2.mean")) getSymbols("UNEMP2", src = "USFedPhil")

getSymbols(c("UNEMP2", "PGDP", "NGDP", "CPROF"), src = "USFedPhil")

# standard deviation
getSymbols("UNEMP2", src = "USFedPhil", Fun = sd)

# just recent data
getSymbols("UNEMP2", src = "USFedPhil", SurveyDecades = "2010s")

# force requery of data from the Philladelphia site
# will collect new xls file(s)
getSymbols("UNEMP2", src = "USFedPhil", force = TRUE)

# all columns in one VERY WIDE xts object
getSymbols("USFedPhilForecastingData", src = "USFedPhil")

}
}
\references{
\cite{Historical Data Files for the Survey of Professional Forecasters \url{https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/data-files}}

\cite{Documentation: PDF Provides information on all variables, transformations, and files in the survey. \url{https://www.philadelphiafed.org/-/media/research-and-data/real-time-center/survey-of-professional-forecasters/spf-documentation.pdf?la=en}}

\cite{Survey of Professional Forecasters \url{https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters}}

\cite{Federal Researve Bank of Philadelphia \url{https://www.philadelphiafed.org/}}
}
\seealso{
\code{\link{getSymbols}}
\code{\link{setSymbolLookup}}
}
\author{
Jeffrey A. Ryan

Andre Mikulec (adapted original code to work with the
US Federal Researve Bank of Philadelphia's Survey of Professional Forecasters)
}
\keyword{data}
