% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getSymbols.YaleU}
\alias{getSymbols.YaleU}
\title{stock market data from Yale University}
\usage{
getSymbols.YaleU(Symbols, env, return.class = "xts", ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded
Possible Symbols are the following:
"YaleUStockMarketData" (means get all of the columns);
otherwise, get specific columns;
"SP500Price", "Dividends", "Earnings",
"CPIAUCNSY" (Consumer Price Index for All Urban Consumers: All Items (of Yale)),}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{desired class of returned object.
Can be xts, zoo, data.frame, or xts (default)}

\item{...}{additional parameters}

\item{force}{FALSE(default) re-download data from YaleU ( See the examples. )
The hidden variable ".YaleUstockMarketData_path2file" located in the
environment of parameter env is the last known location of the "xls" file.
Generally, using this parameter "force = TRUE"
is NOT necessary: the hidden variables are persistent through the
entire R session.  Also, send parameter "verbose = TRUE" to see
the *path2file location.}
}
\value{
A call to getSymbols.YaleU will load into the specified
environment one object for each \code{Symbol} specified,
with class defined by \code{return.class}.
}
\description{
NOT USED ANYWHERE
}
\examples{
\dontrun{

# common usage
if(!exists("CAPE")) getSymbols("CAPE", src = "YaleU")

getSymbols(c("CAPE", "Dividends", "Earnings"), src = "YaleU")

# force requery of data from the Yale University site
# will collect one new xls (historical obs) file
getSymbols(c("CAPE", "Dividends", "Earnings"), src = "YaleU", force = TRUE)

# all columns in one xts object
getSymbols("YaleUstockMarketData", src = "YaleU")

}
}
\references{
\cite{ONLINE DATA ROBERT SHILLER \url{http://www.econ.yale.edu/~shiller/data.htm}}

\cite{Home Page of Robert J. Shiller - Yale University \url{http://www.econ.yale.edu/~shiller/}}
}
\seealso{
\code{\link{getSymbols}}
\code{\link{setSymbolLookup}}
}
\author{
Jeffrey A. Ryan

Andre Mikulec (adapted original code to work with Yale University stock market data)
}
\keyword{data}
