% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{getWhat}
\alias{getWhat}
\title{get OS environment variables or R options}
\usage{
getWhat(x, Storage = NULL)
}
\arguments{
\item{x}{character vector of size one(1)  of what to get}

\item{Storage}{"OS"(default). Otherwise "Options"}
}
\value{
value otherwise NULL
}
\description{
Get truths or values in memory.
Used as an alternative instead of expensive re-quering.
If x is stored ins the OS environment, then x is retrieved as a string.
To be consistent with OS storage, a value stored in options are
also stored a string.  Retrieval exceptions are "TRUE","FALSE".
These are retrieved as TRUE/FALSE.
}
\examples{
\dontrun{

getWhat("TAB_X_COL_Y_EXISTS")
# not 'preferred get location'
# choose "OS" xor "Options"
getWhat("TAB_X_COL_Y_EXISTS", Storage = "Options")

}
}
