% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{\%Like\%}
\alias{\%Like\%}
\title{convenience function calling stringr::str_detect}
\usage{
x \%Like\% pattern
}
\arguments{
\item{x}{string}

\item{pattern}{ICU regular expression pattern}
}
\value{
logical
}
\description{
convenience function calling stringr::str_detect
}
\examples{
\dontrun{
 "1234" \%Like\% "^\\\\d{4}$"
}
}
