% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary_predictive.R
\name{ifelse}
\alias{ifelse}
\alias{ifelse.default}
\alias{ifelse.xts}
\title{comparison of two objects}
\usage{
ifelse(test, yes, no)

\method{ifelse}{default}(test, yes, no)

\method{ifelse}{xts}(test, yes, no)
}
\arguments{
\item{test}{boolean test}

\item{yes}{result of pass}

\item{no}{result of no-pass}
}
\value{
xts object with the same index as xTs1
}
\description{
comparison of two objects
}
