% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{indClms}
\alias{indClms}
\title{get the indicator columns}
\usage{
indClms(xTs = NULL)
}
\arguments{
\item{xTs}{xts objectt}
}
\value{
column names
}
\description{
columns that do not have a corresponding column having
its ending in "_wts" and do not have a do not have the
same root name compared to each and every other *_wts column.
}
\details{
CURRENTLYNOT USED
}
\examples{
\dontrun{
# > require(xts)
# > xTs <- xts(matrix(1:3, ncol = 3, dimnames = list(NULL,c("a","b","b_wts"))),zoo::as.Date(0))[0]
# > indClms(xTs)
# [1] "a"
}
}
