% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{indexSlices}
\alias{indexSlices}
\title{Rob Hyndman style time slices}
\usage{
indexSlices(xTs, SlicesAllData, TrainStart, TrainEnd)
}
\arguments{
\item{xTs}{data}

\item{SlicesAllData}{list of index slice date vectors}

\item{TrainStart}{absolute training start date}

\item{TrainEnd}{absolute training end date}
}
\value{
list of indexSlicesObs and indexSlicesOutObs position indexes
meant to be passed to caret::trainControl index and indexOut
}
\description{
create time slice indexes to be passed to aret::trainControl index and indexOut
My visual observation is that this model DOES \*worse\* than general cross validateion
This is because, per time slice, less observations exist to TRAIN/TEST over
}
