% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{initEnv}
\alias{initEnv}
\title{sets the enviroment
space-saver - meant to be used at the beginning of a function}
\usage{
initEnv(init = NULL)
}
\value{
envi, options, and TZ are set
}
\description{
Variable ops from the calling environment sets R options
Environment variable TZ using the calling environment variable oldtz
sets.
}
\examples{
\dontrun{
# > options(max.print=88888L)
# initEnv()
# > getOption("max.print")
# [1] 99999
}
}
