% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{initMktData}
\alias{initMktData}
\title{internally detect and process inbound variable or string named like "mkdata"}
\usage{
initMktData(xTs = NULL, InBndxTs = NULL)
}
\arguments{
\item{xTs}{object or NULL or bare symbol or string like "mktdata"}

\item{InBndxTs}{This is the xTs name as a symbol or a string}
}
\value{
to the parent's parent env:
\describe{
\item{xTs}{(class xts object)}
\item{InBndxTs(Boolean)}{the string or symbol like "mktdata"}
\item{isInBndxTsMktData(Boolean)}{is the string or symbol like "mktdata":}
}
}
\description{
\preformatted{
Assign to the calling environment the variables:

  xTs:
    the result of work: could be the same as inbound xTs
                        could be massaged by initXts()
  InBndxTs:
    the inbound xTs symbols or name
    user must pre-call gather from "InBndxTs <- as.character(substitute(xTs))"

  isInBndxTsMktData:
    boolean: whether or not the name or symbol "InBndxTs" is like "mktdata"

# NOTE: if one wants to intialize use
data <-initXts(NULL) # default index class Date
# do not use
data <- xts::xts()   # default index class POSIXct,POSIXt
# "YYYY-MM-DD" time ranges input into quantmod::modelData
#  will cause error "unambigous data" error
#  uantmod::modelData by default looks for "POSIXct,POSIXt"

}
}
