% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{initPrintEnv}
\alias{initPrintEnv}
\title{sets the printing enviroment}
\usage{
initPrintEnv()
}
\value{
printing environment is set
}
\description{
space-saver - meant to be used at the beginning of a function
}
\examples{
\dontrun{
# initPrintEnv()
}
}
