% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{initXts}
\alias{initXts}
\title{garantee a passed xts object or a zero length xts object}
\usage{
initXts(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object
}
\description{
garantee a passed xts object or a zero length xts object
}
\examples{
\dontrun{
# > initXts(xTs = NULL)
# Data:
# numeric(0)
#
# Index:
# Date of length 0
#
# > initXts(zoo::as.Date(0)[0])
# Data:
# numeric(0)
#
# Index:
# Date of length 0
}
}
