% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{is.na.xts}
\alias{is.na.xts}
\title{is.na about xts objects}
\usage{
\method{is.na}{xts}(x)
}
\arguments{
\item{x}{xts object}
}
\value{
xts object of T/F values
}
\description{
\preformatted{

NOT CURRRENTLY USED ANYWHERE

}
}
\examples{
\dontrun{

library(xts)

xts(c(NA_real_, 0, NA_real_), zoo::as.Date(0:2))

           [,1]
1970-01-01   NA
1970-01-02    0
1970-01-03   NA

is.na(xts(c(NA_real_, 0, NA_real_), zoo::as.Date(0:2)))

            [,1]
1970-01-01  TRUE
1970-01-02 FALSE
1970-01-03  TRUE

NOTE
is.na(xts::xts(matrix(c(NA_real_, 0, NA_real_,0), ncol = 2), zoo::as.Date(0:1)))
            [,1]  [,2]
1970-01-01  TRUE  TRUE
1970-01-02 FALSE FALSE

}
}
