% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{isWhatFnd}
\alias{isWhatFnd}
\title{search OS environment variables or R options}
\usage{
isWhatFnd(x, y = TRUE, Storage = NULL)
}
\arguments{
\item{x}{character vector of size one(1)  of what to check}

\item{y}{TRUE(default): value to check against (often TRUE)}

\item{Storage}{"OS"(default). Otherwise "Options" or "Both"}
}
\value{
requested value (often TRUE)
}
\description{
\preformatted{

Look for truths or values in memory.
Used as an alternative instead of expensive re-quering.
If x is stored ins the OS environment, then x is stored as a string.
To be consistent with OS storage, a value stored in options are
also stored a string (see setWhat). Test exceptions are "TRUE","FALSE".
These are tested as "TRUE"/"FALSE" or TRUE/FALSE.

}
}
\examples{
\dontrun{

isWhatFnd("TAB_X_COL_Y_EXISTS")
isWhatFnd("TAB_X_COL_Y_EXISTS", TRUE)
# not 'preferred test'
isWhatFnd("TAB_X_COL_Y_EXISTS", "TRUE")
# not 'preferred test location'
isWhatFnd("TAB_X_COL_Y_EXISTS", TRUE, Storage = "Options")
# \\*any\\* one or \\*both\\* of the two
isWhatFnd("TAB_X_COL_Y_EXISTS", TRUE, Storage = "Both")
# \\*any\\* one or \\*both\\* of the two
isWhatFnd("TAB_X_COL_Y_EXISTS", TRUE, Storage = c("OS","Options")

}
}
