% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{LagXts}
\alias{LagXts}
\title{lag an xts object}
\usage{
LagXts(xTs, k = 1, na.pad = TRUE, ...)
}
\arguments{
\item{k}{choose -1 to look into the future}
}
\description{
does not complain when: k > NROW(xTs)
}
