% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{lastUpdatedDate}
\alias{lastUpdatedDate}
\title{number of days since Last Published}
\usage{
lastUpdatedDate(x, AsOfToday = NULL, Last_Updated = NULL,
  Frequency = NULL, DayOfWeek = NULL)
}
\arguments{
\item{x}{xts object returned by getSymbols.FRED2}

\item{AsOfToday}{Sys.Date() (default).
Otherwise overrides; POSIXct, Date or character("YYYY-MM-DD")
Date of refrence. Reality: today or date of run of the "modelling" program.}

\item{Last_Updated}{xts attribute "Last_Updated"(default).
Otherwise overrides; POSIXct, Date or character("YYYY-MM-DD"):}

\item{Frequency}{xts attribute "Frequency"(default).
Otherwise; character overrides.
Values can be "Quarterly". Currently, "Monthly" and "Weekly"
are not yet implmented.
( Future testing is needed to do "Monthly" and "Weekly")}

\item{DayOfWeek}{Read from the xts attributes(default).
Useful when Frequence == "Weekly". Not yet implemented.}
}
\value{
xts object with a new column of show the delay since last update.
}
\description{
The time between the "Last Updated Date" and the "index" date 'published date'
can be important in "model" building.
This amount of time can impact the quality of
the observations.  Therefore, the "model" building program needs
to know about the number of days since the "Last Updated Date."
}
\details{
First:
Of the data of Yahoo finance, the "Last Updated Date" and
the index date 'published date' are the same day. quantmod::getSymbols.yahoo
needs only to return the data and index.
The St. Louis FRED's "Last Updated Date" and index date 'published date'
are different days.

Second: Of the data of Yahoo finance, the Frequency
is "Daily".  The St. Louis FRED's "Frequency" can be"Quarterly"
"Weekly", "Monthly", or "Daily".

From a timeseries returned by getSymbols.FRED2.
The xts attribute "Last_Updated" is used.

Used is To.Monthly to convert the series to to a "Month-like" series.
This is typical in makeing long term predictions.

The xts is returned with an extra column added..

The author put some thought had been put into 'instead querying
ALFRED', but that path choice utimately was not taken.
}
\examples{
\dontrun{

# "Quarterly" time series on the St. Louis FRED
GDP <- getSymbols("GDP", src = "FRED2", auto.assign = FALSE)
lastUpdatedDate(GDP)

}
}
