% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{leadingCashLogReturns}
\alias{leadingCashLogReturns}
\title{leading cash log returns (CASHlogrets)}
\usage{
leadingCashLogReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object (only takes the index)}
}
\value{
leading xts object
}
\description{
\preformatted{

}
}
