% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{leadingCashReturns}
\alias{leadingCashReturns}
\title{leading cash APC returns (CASHapcrets)}
\usage{
leadingCashReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object (only takes the index)}
}
\value{
leading xts object
}
\description{
\preformatted{

}
}
