% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{leadingRetsWtsClms}
\alias{leadingRetsWtsClms}
\title{get the leadingrets_wts columns}
\usage{
leadingRetsWtsClms(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
column names
}
\description{
\preformatted{

}
}
