% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{leadingSymbolAPCReturns}
\alias{leadingSymbolAPCReturns}
\title{get the Symbol leading APC returns}
\usage{
leadingSymbolAPCReturns(Symbol = NULL, src = NULL,
  SymplifyGeneratorFUN = NULL)
}
\value{
xts object
}
\description{
\preformatted{

}
}
