% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{leadingSymbolReturns}
\alias{leadingSymbolReturns}
\title{get the Symbol leading APC returns}
\usage{
leadingSymbolReturns(Symbol = NULL, src = NULL,
  SymplifyGeneratorFUN = NULL, ReturnsGeneratorFUN = NULL,
  ReturnsGeneratorFUNArgs = NULL)
}
\value{
xts object
}
\description{
\preformatted{

}
}
