% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{leadingWilshire5000LogReturns}
\alias{leadingWilshire5000LogReturns}
\title{get the Wilshire 5000 Index leading log returns}
\usage{
leadingWilshire5000LogReturns()
}
\value{
xts object
}
\description{
\preformatted{

}
}
