% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{Less}
\alias{Less}
\title{difference of two xTs objects}
\usage{
Less(xTs1 = NULL, xTs2 = NULL)
}
\arguments{
\item{xTs1}{leading xts object}

\item{xTs2}{trailing xts object}
}
\value{
xts object with the same index as xTs1
}
\description{
difference of two xTs objects
}
