% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{liquifyDF}
\alias{liquifyDF}
\title{flatten a data.frame to (hopfully) ONE row}
\usage{
liquifyDF(x, ConstColsRegex = "^id", FactorColsRegex = "(fct|Fact)$",
  FactorColsNAReplace = NULL, FactorColsFixedSep = ".",
  DetailColsFixedSep = ".", SpaceFixedSep = ".",
  AmperstandFixedSep = "And")
}
\description{
flatten a data.frame to (hopfully) ONE row
}
\examples{
\dontrun{

# data.frame(
#   id = c(1, 1, 1, 1),
#   id_fct = c("1", "1", "1", "1"),
#   col1_fct = c('A1', 'A1', 'A2', 'A2'),
#   col2_fct = c('B1', 'B2', 'B1', 'B2'),
#   aggr1 = c(8, 16, 32, 64),
#   aggr2 = c(10008, 10016, 10032, 10064)
# , stringsAsFactors = FALSE
# )  -> DFS
#
# > DFS
#   id id_fct col1_fct col2_fct aggr1 aggr2
# 1  1      1       A1       B1     8 10008
# 2  1      1       A1       B2    16 10016
# 3  1      1       A2       B1    32 10032
# 4  1      1       A2       B2    64 10064
#
# > liquifyDF(DFS)
#   id id_fct a1__b1____aggr1 a1__b2____aggr1 a2__b1____aggr1 a2__b2____aggr1 a1__b1____aggr2 a1__b2____aggr2 a2__b1____aggr2
# 1  1      1               8              16              32              64           10008           10016           10032
#   a2__b2____aggr2
# 1           10064
#
}
}
