% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{listSymbols.cache}
\alias{listSymbols.cache}
\title{xts object symbols in the cache}
\usage{
listSymbols.cache(source.envir = NULL, ...)
}
\arguments{
\item{source.envir}{location of xts objects}

\item{...}{passed unused}
}
\value{
character vector of Symbols
}
\description{
xts object symbols in the cache
}
