% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{listSymbols}
\alias{listSymbols}
\title{object symbols in a persistent location (dispatcher)}
\usage{
listSymbols(con = NULL, source.envir = NULL, ...)
}
\arguments{
\item{source.envir}{source location of Symbols
xts objects must have the attribute "src"}

\item{...}{passed to src}

\item{src}{source program to look for symbols}
}
\value{
character vector of Symbols
}
\description{
looks in source.envir xor "src" exclusively
}
\examples{
\dontrun{

# Symbols must be case-insenstive unique

getSymbols("IBM", src = "yahoo") # auto.assign = TRUE

# Symbols names found in Symbols and the names of xts objects stored
# in source.envir, must be mutually exclusive ( case in-sensensitive match)
# (because list2env silently drops repeated members)

# save just the Symbol "IBM"
saveSymbols("IBM", trg = "pg")

msft <- getSymbols("MSFT", src = "yahoo", auto.assign = FALSE)
source.envir = list2env(list(MSFT = msft))
listSymbols(source.envir = source.envir)
existSymbols(Symbols = "MSFT", source.envir = source.envir)

saveSymbols(Symbols = "", trg = "cache", source.envir = source.envir)
listSymbols(src = "cache")
existSymbols(Symbols = "MSFT", src = "cache")

# save all of the .getSymbols Symbols and the source.envir Symbols
saveSymbols(trg = "pg", source.envir = source.envir)
listSymbols(src = "pg")
existSymbols(c("MSFT","IBM"), src = "pg")

}
}
