% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{logReturns}
\alias{logReturns}
\title{geometric investment results}
\usage{
logReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object of arithmatic returns}
}
\value{
xts object of geometric returns
}
\description{
\preformatted{

xTs values less that zero will generate a numeric error

}
}
\examples{
\dontrun{
# > require(xts)
# > xTs  <- xts(10:12,zoo::as.Date(0:2))
# > lr <- logReturns(xTs)
# > lr
#                         logrets
# 1970-01-01 0.000000000000000000
# 1970-01-02 0.095310179804324768
# 1970-01-03 0.087011376989629685
#
# > as.vector(coredata(exp(cumsum(lr)) * 10L))
# [1] 10.000000000000000 10.999999999999998 11.999999999999996
}
}
