% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{makeFSTFiles}
\alias{makeFSTFiles}
\title{From source files, create .fst files}
\usage{
makeFSTFiles(
  Dirs = NULL,
  SubDirExpr = NULL,
  FileExpr = NULL,
  ColRmExpr = NULL,
  Force = NULL
)
}
\arguments{
\item{Dirs}{character vector of directories}

\item{SubDirExpr}{NULL(default).  NULL means "all." Character vector of directories}

\item{FileExpr}{NULL(default). NULL means "all." Full file vector of names}

\item{ColRmExpr}{NULL(default) NULL meand "none." Regular expression of columns to remove}

\item{Force}{False(default). If the target .fst file is there, then do not replace it.
Otherwise, replace it.}
}
\value{
NULL silently.  Side-effect of a new .fst file created on disk in the old
directory location.
}
\description{
\preformatted{

Expected main directories (dirs), with exactly one level
of subdirectories. Each subdirectory contains files (with columns).

}
}
\examples{
\dontrun{

# earlest files
makeFSTFiles(Dirs = "W:\\\\AAIISIProDBFs", SubDirExpr = c("12055","12083"), FileExpr = c("si_ci.dbf","si_isq.dbf","si_bsq.dbf","si_cfq.dbf","si_psd.dbf","si_date.dbf"), ColRmExpr = c("^X_NullFlags$","^X.?\\\\d*$"), Force = TRUE)

# all files
makeFSTFiles(Dirs = "W:\\\\AAIISIProDBFs", FileExpr = c("si_ci.dbf","si_isq.dbf","si_bsq.dbf","si_cfq.dbf","si_psd.dbf","si_date.dbf"), ColRmExpr = c("^X_NullFlags$","^X.?\\\\d*$"), Force = TRUE)

}
}
