% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{modelData<-}
\alias{modelData<-}
\title{low-level set the values quantmod object slots}
\usage{
modelData(x, ...) <- value
}
\arguments{
\item{x}{quantmod object}

\item{...}{list of name-value pairs}
}
\description{
# methods/html/slot.html
}
