% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{multitable___mouter}
\alias{multitable___mouter}
\title{multi outer}
\usage{
multitable___mouter(x, ...)
}
\arguments{
\item{x}{list or vectors of values}

\item{...}{dots; one function FUN to applied to the vectors}
}
\value{
array
}
\description{
\preformatted{

Applies reduction of outer amoung multiple vectors.
FUN has a required formals that are similar to "paste". See the example.
This is NOT USED ANYWHERE.

}
}
\examples{
\dontrun{

multitable___mouter(list(LETTERS[1:3], letters[1:2], 1:3), paste0)
multitable___mouter(list(LETTERS[1:3], letters[1:2], 1:3), paste)

# Require FUN function formals to be similar to "paste"
pasteDot <- function(..., sep = ".", collapse = NULL) { paste(..., sep = sep, collapse = collapse) }
multitable___mouter(list(LETTERS[1:3], letters[1:2], 1:3), pasteDot)

# convert the array to a vector
c(multitable___mouter(list(LETTERS[1:3], letters[1:2], 1:3), pasteDot))
}
}
\references{
\cite{R CRAN package multitable \url{https://cran.r-project.org/src/contrib/Archive/multitable}}

\cite{R package multitable \url{https://github.com/stevencarlislewalker/multitable}}
}
\author{
Steven Walker
}
