% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{Names<-}
\alias{Names<-}
\title{if x or value is null or has length zero(0) then
then that item is character(0)}
\usage{
Names(x) <- value
}
\arguments{
\item{x}{names}

\item{value}{result}
}
\description{
if x or value is null or has length zero(0) then
then that item is character(0)
}
