% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{nextMonthfromYesterday.Date}
\alias{nextMonthfromYesterday.Date}
\title{change the date to the future}
\usage{
\method{nextMonthfromYesterday}{Date}(date = NULL)
}
\arguments{
\item{date}{object object}
}
\value{
date object
}
\description{
change the date to the future
}
\examples{
\dontrun{
# > require(xts)
# > nextMonthfromYesterday(zoo::as.Date("1970-01-12"))
# [1] "1970-01-31"
}
}
