% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{nextMonthfromYesterday.default}
\alias{nextMonthfromYesterday.default}
\title{attempt to change the index date to the future}
\usage{
\method{nextMonthfromYesterday}{default}(xTs = NULL)
}
\arguments{
\item{xTs}{default object}
}
\value{
stop
}
\description{
attempt to change the index date to the future
}
