% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{nextMonthfromYesterday}
\alias{nextMonthfromYesterday}
\alias{nextMonthfromYesterday.default}
\alias{nextMonthfromYesterday.Date}
\alias{nextMonthfromYesterday.xts}
\title{change the index date to the future}
\usage{
nextMonthfromYesterday(xTs = NULL)

\method{nextMonthfromYesterday}{default}(xTs = NULL)

\method{nextMonthfromYesterday}{Date}(date = NULL)

\method{nextMonthfromYesterday}{xts}(xTs = NULL)
}
\arguments{
\item{xTs}{xts object or data object}

\item{date}{date object}
}
\value{
xts object or data object with the index moved up
}
\description{
change the index date to the future
}
\examples{
\dontrun{
# > require(xts)
# > nextMonthfromYesterday(zoo::as.Date("1970-01-12"))
# [1] "1970-01-31"
}
\dontrun{
# > require(xts)
# > xTs <- xts(, zoo::as.Date("1970-01-12"))
# > nextMonthfromYesterday(xTs)
# Data:
# numeric(0)
#
# Index:
#  Date[1:1], format: "1970-01-31"
}
}
