% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{nextMonthfromYesterday.xts}
\alias{nextMonthfromYesterday.xts}
\title{change the date to the future}
\usage{
\method{nextMonthfromYesterday}{xts}(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
date object
}
\description{
change the date to the future
}
\examples{
\dontrun{
# > require(xts)
# > xTs <- xts(, zoo::as.Date("1970-01-12"))
# > nextMonthfromYesterday(xTs)
# Data:
# numeric(0)
#
# Index:
#  Date[1:1], format: "1970-01-31"
}
}
