% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{nonImportantClmns}
\alias{nonImportantClmns}
\title{get the non-[current/leading]rets[_wts] columns}
\usage{
nonImportantClmns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
column names
}
\description{
\preformatted{

}
}
