% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{NVAR}
\alias{NVAR}
\alias{NVAR.default}
\alias{NVAR.xts}
\title{number of variables}
\usage{
NVAR(x = NULL)

\method{NVAR}{default}(x = NULL)

\method{NVAR}{xts}(x = NULL)
}
\arguments{
\item{x}{xts object}
}
\value{
integer number of variables
}
\description{
currently NOT USED
}
\examples{
\dontrun{
# > require(xts)
# > NVAR(xts(, zoo::as.Date("1970-01-12")))
# [1] 0
}
}
