% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{oneColumn}
\alias{oneColumn}
\title{returns One column from a database (con)}
\usage{
oneColumn(con, Query, outName, unQuote = NULL)
}
\arguments{
\item{con}{DBI database connection}

\item{Query}{SQL expecting output to be a single value,
no value, or a single value with with comma separted internal parts}

\item{outName}{the new column name}

\item{unQuote}{(default: FALSE) strip beginning(") and end quotes(")}
}
\description{
returns One column from a database (con)
}
\examples{
\dontrun{

pgCurrentUser <- oneColumn(con, "SELECT CURRENT_USER;", "CurrentUser")

pgCurrentSchema <- oneColumn(con, "SELECT current_schema();", "CurrentSchema")

pgCurrentDB <- oneColumn(con, "SELECT current_database();", "CurrentDB")

pgCurrentSearchPath <- oneColumn(con, "SHOW SEARCH_PATH;", "CurrentSearchPath")

pgCurrentTempSchema <- oneColumn(con, "SELECT nspname FROM pg_namespace WHERE oid = pg_my_temp_schema();", "CurrentTempSchema")

}
}
