% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{OneOf}
\alias{OneOf}
\title{choose local minima (or maxima)}
\usage{
OneOf(xTs, base = 0, lag = 0:1, take = "min", Fun = NULL, ...)
}
\arguments{
\item{xTs}{xts object}

\item{base}{choose -1 (or less) to look into the future}

\item{lag}{0:1(default) periods to lag over: this(0) and previous(1)}

\item{take}{"min"(default).  Other option is "max". Finally,
the user can pass in a custom function "Fun."  If Fun is passed
then "take" is ignored.}

\item{Fun}{custom user passed function}

\item{...}{dots passed to Fun}
}
\description{
\preformatted{

Lags an xts column into a 2nd+ columns.
This chooses the lowest(or highest) value rowwise

This is meant for removing noise from a monthly financial time series.
If the goal is a pessimisic (buy "puts") solution
then an optimal choice may be (lag = 0:1, take = "min")

}
}
\examples{
\dontrun{

x <- xts(0:2, zoo::as.Date(0:2))
OneOf(x)
          [,1]
1970-01-01   NA
1970-01-02    0
1970-01-03    1

}
}
