% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{pairWise}
\alias{pairWise}
\title{pairwise interleave of columns}
\usage{
pairWise(x1, x2)
}
\arguments{
\item{x1}{data.frame or xts object}

\item{x2}{data.frame or xts object}
}
\value{
list of length two of two data.frames or xts objects
}
\description{
works better on xts objects ( lapply or plyr::llply structure is held together )
}
\examples{
\dontrun{
#
# list(iris[1:2,1:2], airquality[1:2,1:2])
# [[1]]
#   Sepal.Length Sepal.Width
# 1          5.1         3.5
# 2          4.9         3.0
#
# [[2]]
#   Ozone Solar.R
# 1    41     190
# 2    36     118
#
# > str( pairWise( iris[1:2,1:2], airquality[1:2,1:2] ) )
# List of 2
#  $ :List of 2
#   ..$ Sepal.Length: num [1:2] 5.1 4.9
#   ..$ Ozone       : int [1:2] 41 36
#  $ :List of 2
#   ..$ Sepal.Width: num [1:2] 3.5 3
#   ..$ Solar.R    : int [1:2] 190 118
#
# # 1 by N recycling
# #
#  > str( pairWise( iris[1:2,1:2], airquality[1:2,1, drop = F] ) )
#  List of 2
#   $ :List of 2
#    ..$ Sepal.Length: num [1:2] 5.1 4.9
#    ..$ Ozone       : int [1:2] 41 36
#   $ :List of 2
#    ..$ Sepal.Width: num [1:2] 3.5 3
#    ..$ Ozone      : int [1:2] 41 36
#
# > require(xts)
# > data("sample_matrix", package = "xts")
# > str( pairWise(as.xts(sample_matrix)[,1:2], as.xts(sample_matrix)[,3:4] ) )
# List of 2
#  $ :List of 2
#   ..$ Open:An 'xts' object on 2007-01-02/2007-06-30 containing:
#   Data: num [1:180, 1] 50 50.2 50.4 50.4 50.2 ...
#  - attr(*, "dimnames")=List of 2
#   ..$ : NULL
#   ..$ : chr "Open"
#   Indexed by objects of class: [POSIXct,POSIXt] TZ:
#   xts Attributes:
#  NULL
#   ..$ Low :An 'xts' object on 2007-01-02/2007-06-30 containing:
#   Data: num [1:180, 1] 50 50.2 50.3 50.2 50.1 ...
#  - attr(*, "dimnames")=List of 2
#   ..$ : NULL
#   ..$ : chr "Low"
#   Indexed by objects of class: [POSIXct,POSIXt] TZ:
#   xts Attributes:
#  NULL
#  $ :List of 2
#   ..$ High :An 'xts' object on 2007-01-02/2007-06-30 containing:
#   Data: num [1:180, 1] 50.1 50.4 50.4 50.4 50.2 ...
#  - attr(*, "dimnames")=List of 2
#   ..$ : NULL
#   ..$ : chr "High"
#   Indexed by objects of class: [POSIXct,POSIXt] TZ:
#   xts Attributes:
#  NULL
#   ..$ Close:An 'xts' object on 2007-01-02/2007-06-30 containing:
#   Data: num [1:180, 1] 50.1 50.4 50.3 50.3 50.2 ...
#  - attr(*, "dimnames")=List of 2
#   ..$ : NULL
#   ..$ : chr "Close"
#   Indexed by objects of class: [POSIXct,POSIXt] TZ:
#   xts Attributes:
#  NULL
#
}
}
