% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{pasteDot}
\alias{pasteDot}
\title{paste with the default dot(.) separator}
\usage{
pasteDot(..., sep = ".", collapse = NULL)
}
\arguments{
\item{...}{dots as paste}

\item{sep}{as paste}

\item{collapse}{as paste}
}
\description{
\preformatted{

NOT USED ANYWHERE

}
}
