% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgAddColumnType}
\alias{pgAddColumnType}
\title{add a column to a database table}
\usage{
pgAddColumnType(con, trgt = NULL, schname = NULL, df = NULL)
}
\arguments{
\item{con}{DBI connection PostgreSQL}

\item{trgt}{remote server side string database table name of old data}

\item{schname}{schema name}

\item{df}{local client side data.frame of 'updated data'. Only the column
names are matched to the Server DB table}
}
\description{
matches by colum name. If the column the does not
exist on the Server DB, then the column is added
by ALTER TABLE <trgt> ADD ( column ) called each per new column.
}
\examples{
\dontrun{

 con <- DBI::dbConnect(RPostgreSQL::PostgreSQL(), user = "postgres")
 DBI::dbWriteTable(con, "mtcars", oldData, row.names = FALSE)
 pgAddColumnType(con, trgt = "mtcars", schname = "public", df = data.frame(rnk = 0L))

}
}
