% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgAddTable}
\alias{pgAddTable}
\title{add a table to a database}
\usage{
pgAddTable(con, trgt = NULL, keys = c("rn"), schname, df = NULL)
}
\arguments{
\item{con}{DBI database connection}

\item{keys}{trgt remote server side vector of strings of table
column names that make up a unique id for the row.
keys can not be zero length. keys can not be null.}

\item{schname}{schema name}

\item{df}{data.frame (with column names)}

\item{Symbol}{new table name. E.g Could be a company TICKER or a FRED column.}
}
\description{
add a table to a database
}
