% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgConnect}
\alias{pgConnect}
\title{smartly connect to PostgreSQL database}
\usage{
pgConnect(
  user = NULL,
  password = NULL,
  dbname = NULL,
  schname = NULL,
  host = NULL,
  port = NULL,
  options = NULL,
  forceISOdate = TRUE
)
}
\arguments{
\item{user}{username(default "Symbols") to access database}

\item{password}{password(default "Symbols") to access database}

\item{dbname}{database name (default "Symbols")}

\item{schname}{schema name (default "Symbols")}

\item{host}{database host (default "localhost")}

\item{port}{database port (default 5432)}

\item{options}{pass extra parameters in aa string to the command line}

\item{forceISOdate}{TRUE(default)/FALSE if the communication of date (time stamp) from PostgreSQL
is forced to ISO style at conection}
}
\value{
list of "con" DBI Connection object and "schname" final chosen schema name
}
\description{
\preformatted{

}
}
