% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgDFColClasses}
\alias{pgDFColClasses}
\title{from R data.frame column types, determine Server DB data types}
\usage{
pgDFColClasses(df = NULL, ...)
}
\arguments{
\item{df}{local client side data.frame. If df's class `[` method
drops single dimentions to vectors the call by using
pgDFColClasses(df[, "col1", drop = F])}
}
\value{
named vector of names of column names and
values of Server DB column data types
}
\description{
from R data.frame column types, determine Server DB data types
}
\examples{
\dontrun{

pgDFColClasses(mtcars[, "cyl", drop = F])
            cyl
"NUMERIC(14,3)"

}
}
