% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgInsert}
\alias{pgInsert}
\title{insert into a database table 'new' data}
\usage{
pgInsert(con, trgt = NULL, keys = c("rn"), schname, df = NULL,
  varHint = NULL, valHint = NULL)
}
\arguments{
\item{con}{DBI connection PostgreSQL}

\item{trgt}{remote server side string database table name of old data}

\item{keys}{trgt remote server side vector of strings of table
column names that make up a unique id for the row.
keys can not be zero length. keys can not be null.}

\item{schname}{schema name}

\item{df}{local client side data.frame of limited data
that determines what data is to be returned from the Server DB}

\item{varHint}{optional vector of character column names. Performance optimization
techique to limit the number of rows returned
from the database server to the client(R).
User must specify as paired position values.
e.g. varHint = "dateindex", valHint = "17000"
or e.g. varHint = c("dateindex", "ticker"), valHint = c("17000","'AAPL'").
Position matches one to one with valHint.}

\item{oldData}{previously collected limited server DB data used to
restrict what is updated/inserted.  If not provide by the user then
the oldData will be (re)generated from df.
Here oldData is used to CHECK that the key columns that I
what to insert DO NOT EXISTS on the Server DB
Here, only the keys are checked for 'no duplicates'}
}
\description{
insert into a database table 'new' data
}
\examples{
\dontrun{

# setup
SuBmtcars <- mtcars[c(1,5),1:2]
oldData <- data.table::data.table(SuBmtcars, keep.rownames=TRUE, key="rn")
oldData[1,2] <- NA; oldData[2,3] <- NA

con <- DBI::dbConnect(RPostgreSQL::PostgreSQL(), user = "postgres")
DBI::dbExecute(con, "DROP TABLE IF EXISTS public.mtcars")
DBI::dbWriteTable(con, "mtcars", oldData, row.names = FALSE)

newInsData <- data.table::data.table(mtcars[12, 2, drop = FALSE], keep.rownames=TRUE, key="rn")
pgInsert(con, trgt = "mtcars", keys = "rn", schname = "public",
  df = newInsData, varHint = "cyl", valHint = "8")

}
}
