% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgListSchemaTableColumns}
\alias{pgListSchemaTableColumns}
\title{of a specific PostgreSQL database schema table, show its columns}
\usage{
pgListSchemaTableColumns(con, schname, tblname)
}
\arguments{
\item{con}{PostgreSQL DBI connection}

\item{schema}{name}

\item{table}{name}
}
\value{
vector of characters of column names
The results are ordered.
}
\description{
of a specific PostgreSQL database schema table, show its columns
}
