% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgListSchemaTablePrimaryKeyColumns}
\alias{pgListSchemaTablePrimaryKeyColumns}
\title{of a specific PostgreSQL database schema table, show its primary key columns}
\usage{
pgListSchemaTablePrimaryKeyColumns(con, schname, tblname)
}
\arguments{
\item{con}{PostgreSQL DBI connection}

\item{schema}{name}

\item{table}{name}
}
\value{
vector of characters of primary key column names
The results are ordered.
}
\description{
of a specific PostgreSQL database schema table, show its primary key columns
}
