% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgListSchemaTables}
\alias{pgListSchemaTables}
\title{of a specific PostgreSQL database schema, show its tables}
\usage{
pgListSchemaTables(con, schname)
}
\arguments{
\item{con}{PostgreSQL DBI connection}

\item{schname}{schema name}
}
\value{
vector of characters of table names
The results do not have any order.
}
\description{
of a specific PostgreSQL database schema, show its tables
}
