% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgSchemaTableColumnTypes}
\alias{pgSchemaTableColumnTypes}
\title{of a specific PostgreSQL database schema table, show its columns and data types}
\usage{
pgSchemaTableColumnTypes(con, schname, tblname)
}
\arguments{
\item{con}{PostgreSQL DBI connection}

\item{schema}{name}

\item{table}{name}
}
\value{
data.frame of characters of schema, table, column, and human readable type
ordered by ordinal position
The results are ordered.
}
\description{
\preformatted{

}
}
