% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgSchemaTableLastIndex}
\alias{pgSchemaTableLastIndex}
\title{of a specific PostgreSQL database schema table show it's last date/time index value}
\usage{
pgSchemaTableLastIndex(con, schname = NULL, tblname)
}
\arguments{
\item{con}{PostgreSQL DBI connection}

\item{schname}{schema name}

\item{tblname}{table name}
}
\value{
table last date/time index value
The results do not have any order.
}
\description{
of a specific PostgreSQL database schema table show it's last date/time index value
}
