% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{portfolioLogReturns}
\alias{portfolioLogReturns}
\title{get the porfolio log returns}
\usage{
portfolioLogReturns(xTs = NULL, initVal = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{initVal}{start value of the investor's porfolio}
}
\value{
xts object of geometric returns
}
\description{
\preformatted{

This is calculated by taking the proporation of weights(0-1),
but typically a fraction and then multiplying this proportion
by its corresponding value column.  The sum of weights(_wts)
columns sum to be one(1).

}
}
