% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{portfolioMonthlyReturns}
\alias{portfolioMonthlyReturns}
\title{get the porfolio non-log(arithmatic) returns}
\usage{
portfolioMonthlyReturns(xTs = NULL, initVal = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{initVal}{start value of the investor's porfolio}
}
\value{
xts object of arithmatic returns
}
\description{
\preformatted{

}
}
