% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{predictModel.train}
\alias{predictModel.train}
\title{does caret train}
\usage{
\method{predictModel}{train}(object, data, ...)
}
\description{
\preformatted{

This is from quantmod predictModel.

Example (TODO [ ]) simplify.

}
}
