% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{prepAndDoMachineWtsData}
\alias{prepAndDoMachineWtsData}
\title{prep Machine weights data then call doMachineWts}
\usage{
prepAndDoMachineWtsData(
  xTs = NULL,
  ModelFormula = NULL,
  Predictee = NULL,
  Predictors = NULL,
  IndicatorGeneratorFUN = NULL,
  ModelTarget = NULL,
  ...
)
}
\arguments{
\item{xTs}{xts object}

\item{ModelFormula}{NULL(default) "formula" (IF KNOWN in advance) to be based on
the Predictee string and the columns created by the IndicatorGeneratorFUN function.
If left NULL, then the formula be generated by the string Predictee and the column names returned
by after running the IndicatorGeneratorFUN function.}

\item{Predictee}{string of the name of the Predictee variable}

\item{Predictors}{characters vector of the require columns needed in the
indictor generator function: IndicatorGeneratorFUN}

\item{IndicatorGeneratorFUN}{string of the name of the function
or the function itself that generates more columns (to eventually
be used in the model.)}

\item{ModelTarget}{Exact column name that I am (eventually) trying
to 'financially' optimize. xtsAttributes(xTs)[["rettarget"]](Default)}

\item{...}{dots passed to the indicator generator function IndicatorGeneratorFUN(dots)
and to function doMachineWts(dots)}

\item{ExtremePct}{passed in dots to function doOneChoice}
}
\value{
xts object
}
\description{
prep Machine weights data then call doMachineWts
}
