% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{printCalendar}
\alias{printCalendar}
\title{print calendar of month by month returns}
\usage{
printCalendar(
  xTs = NULL,
  title = NULL,
  PortFolioRetCol = NULL,
  MarginReturns = NULL,
  start = NULL,
  end = NULL,
  DivBy = NULL
)
}
\arguments{
\item{xTs}{xts object}

\item{title}{heading}

\item{PortFolioRetCol}{"portfolio.returns"(default). Column of the 'portfolio returns.'}

\item{MarginReturns}{TRUE(Default) Print PerformanceAnalytics::table.CalendarReturns aggregates
in the margin. If FALSE, is a simpleCalendar instead with no marginal data (because
it may not make sense.  E.g. printing the GDP)}

\item{start}{passed to window.xts}

\item{end}{passed to window.xts}

\item{DivBy}{divide by this value before printing.
Useful to divide Fitted by 100.}
}
\value{
xts object of arithmatic returns
}
\description{
\preformatted{

The vericle left side axis is in "years".
The horizontal axis is in "months".

}

Geometric adding is done across months
An end of year(12 month summary) is in the verticle right side axis
}
