% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{pureBasename}
\alias{pureBasename}
\title{funky rstudio problem (involves "gstat" (and "spacetime"))}
\usage{
pureBasename(path, slash = FALSE)
}
\description{
funky rstudio problem (involves "gstat" (and "spacetime"))
}
