% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantico.R
\docType{package}
\name{quantico}
\alias{quantico}
\alias{quantico-package}
\title{quantico: Quantitive Mathematics and Financial Prediction}
\description{
functions
}
\section{safe data}{


initDate
initXts
initEnv
initPrintEnv
unitEnv

initPorfVal
}

\section{data aquisition}{


fredData
fredEomData
}

\section{target and indicator aquisition}{


Willshire5000*
Cash*
UnRate*
}

\section{quanitative mathematical}{


*LogReturns
}

\section{date manipulation}{


nextMonthfromYesterday
}

\section{data clustering}{


combineXts
}

\section{investment strategy}{


*Wts
}

\section{field identification}{


valueClmns
wtsClmns
indClms
}

\section{total return calculating}{


portfolioLogReturns
portfolioMonthlyReturns
}

\section{printing}{


printTail
printCalendar
}

\section{full strategies and results}{


UnRateEyeBalltradeModel
}

