% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{quantmod___convert.time.series}
\alias{quantmod___convert.time.series}
\title{convert time series}
\usage{
quantmod___convert.time.series(fr, return.class)
}
\arguments{
\item{fr}{as quantmod:::convert.time.series}

\item{return.class}{as quantmod:::convert.time.series}
}
\description{
\preformatted{

See pacakge quantmod quantmod:::convert.time.series.

}
}
