% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{quantmod___try.download.file}
\alias{quantmod___try.download.file}
\title{try to download a file}
\usage{
quantmod___try.download.file(
  url,
  destfile,
  method,
  quiet = FALSE,
  mode = "w",
  cacheOK = TRUE,
  extra = getOption("download.file.extra"),
  ...
)
}
\arguments{
\item{url}{as quantmod:::try.download.file}

\item{destfile}{as quantmod:::try.download.file}

\item{method}{as quantmod:::try.download.file}

\item{quiet}{as quantmod:::try.download.file}

\item{mode}{as quantmod:::try.download.file}

\item{extra}{as quantmod:::try.download.file}

\item{...}{as quantmod:::try.download.file}
}
\description{
\preformatted{

See pacakge quantmod quantmod:::convert.time.series.

}
}
