% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{relativeScale}
\alias{relativeScale}
\title{relative scaling}
\usage{
relativeScale(x, ...)
}
\arguments{
\item{x}{anything with methods for
Ops: +-/* and possibly methods for min/max}

\item{...}{passed to DMwR ReScaling. Do not pass: x, t.mn, and t.mx}
}
\value{
re-scaled from low:mim/max*100' to max:100
}
\description{
relative scaling
}
