% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{relativeScaleImportance}
\alias{relativeScaleImportance}
\title{relative scaling of 2 dimensions}
\usage{
relativeScaleImportance(x, cols = c("importance"), ...)
}
\arguments{
\item{cols}{columns to rescale}

\item{...}{passed to DMwR ReScaling. Do not pass: x, t.mn, and t.mx}

\item{2}{dim object data.frame or two=dim non-matrix, non-array
with methods for
Ops: "+","-","/", "*" and possibly methods for min/max}
}
\value{
re-scaled from low:mim/max*100' to max:100
}
\description{
relative scaling of 2 dimensions
}
