% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{releaseMktData}
\alias{releaseMktData}
\title{Return internally processed outbound variable or string named like "mkdata"}
\usage{
releaseMktData(
  xTs = NULL,
  InBndxTs = NULL,
  isInBndxTsMktData = NULL,
  xTsInvisible = NULL
)
}
\arguments{
\item{xTs}{object or NULL or bare symbol or string like "mktdata"}

\item{InBndxTs}{This is the xTs name as a symbol or a string}

\item{isInBndxTsMktData}{Boolean default(FALSE). This is whether or not
xTs is a symbol or name that is a string like "mktdata"}

\item{xTsInvisible}{FALSE(default) if were to return the xTs then
do so invisibly.  Usefully TRUE in the case of printing.}
}
\value{
xts object or assign a variable name InBndxTs with xts value of the xts object
}
\description{
\preformatted{
Processes initMktData variables xTs, InBndxTs, and isInBndxTsMktData
Returns a "mktdata" variable to the caller's calling environment.
}
}
\examples{
\dontrun{

fooY <- function(xTs = NULL) {

  InBndxTs <- as.character(substitute(xTs))
  initMktData(xTs, InBndxTs)

  # some work
  xTs <- xts(0, zoo::as.Date(0))

  return(releaseMktData(xTs, InBndxTs, isInBndxTsMktData))
}
# fooY("mktdata")
# str(mktdata)
# An ‘xts’ object . . .
#

}
}
