% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{RemoveEmptyRows}
\alias{RemoveEmptyRows}
\title{remove rows with all NA vaues}
\usage{
RemoveEmptyRows(x, Var = NULL)
}
\arguments{
\item{x}{data.frame}

\item{Var}{columns determining complete cases}
}
\value{
df with rows of all-NA removed
}
\description{
MOVE ME
complete.cases does not work on tibbles!
Function is written in DataCombine style
}
