% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{Returns}
\alias{Returns}
\title{absolute proportional change results}
\usage{
Returns(xTs = NULL, ...)
}
\arguments{
\item{xTs}{xts object of absolute values}

\item{...}{dots if a 'returns generator' looks for the list
named "ReturnsGeneratorFUNArgs".}
}
\value{
xts object of absolute proportional change returns
}
\description{
\preformatted{

Must return SYMBOL*xxxrets,
such that the zone of xxx * starts with a non-uppercase character.
"xxx" is currently expected to "not be UPPERCASE characters"?
Later, the xts return result is expected to be passed to "Current" and "Leading"
where "rets" is replaced by "currentrets" and "leadingrets"

xTs values less that zero will generate a numeric error

}
}
\examples{
\dontrun{

require(xts)
xTs  <- xts(10:12,zoo::as.Date(0:2))
apcr <- Returns(xTs, ReturnsGeneratorFUNArgs = list(is.na.zero = TRUE,  Fun = list(Fun = "APC", lag = 1)))
apcr
            V1apc.1rets
1970-01-01 0.000000
1970-01-02 0.100000
1970-01-03 0.090909

}
}
