% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{returnsUnRateEyeBall}
\alias{returnsUnRateEyeBall}
\title{runs the main}
\usage{
returnsUnRateEyeBall()
}
\value{
xts object of monthly return results
}
\description{
space-saver - meant to be used at the beginning of a function
}
\examples{
\dontrun{
EXAMPLE
# EXAMPLE
}
}
