% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{rlang__caller_frame}
\alias{rlang__caller_frame}
\title{Get caller frame}
\usage{
rlang__caller_frame(n = 1)
}
\arguments{
\item{n}{Number of frames to go back.}
}
\description{
\preformatted{
Get caller frame
caller_frame() is deprecated as of rlang 0.3.0.
}
}
