% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{rollEpochRanks}
\alias{rollEpochRanks}
\title{rolling percent ranks grouped by epoch}
\usage{
rollEpochRanks(xTs1, xTs2, window = 10, minimum = window, ranks = 4,
  incBeforeEpochLags = FALSE)
}
\arguments{
\item{xTs1}{xts object}

\item{xTs2}{xts object of single column (or multiple column if I do interaction)
xts object with a data flag of the groupings.
This is most likely calculated from this package's
days/months\*Since/After\*Event functions.}
}
\description{
The idea is group rolling ranks from the end of crash/recession
to the beginning of the next crash/recession (so that the
recession to recession social/economic behavior may be
recaptured.)
}
\details{
[Maybe] this is (weakly) designed to run through "eXplode."

Note: these are not good ways to do "accurate" rolling [percent] ranks.
TTR::runPercentRank is awfully bad (I do not know what the author was thinking).
fromo::running_apx_quantiles is unbalanced.
( so I can not use in findInterval to get ranks)

Note: Consider, sort(method = "radix")/order/tail as a
fast(simple) way to approximate "ranks"
}
\examples{
\dontrun{


# S&P 500
gspc <- To.Monthly(Cl(getSymbols("^GSPC", auto.assign = FALSE, from = "1900-01-01")), indexAt = "lastof", OHLC = F)
colnames(gspc) <- "GSPC"
# returns
gspc.ret <- TTR::ROC(gspc, type = "discrete")

# top 20 worst monthly returns in the S&P500 history
CrashDates <- zoo::as.Date(rownames(head(as.matrix(gspc.ret)[order(as.matrix(gspc.ret)),1, drop = F], 20)))

mse <- monthsSinceEvent(gspc.ret, dates = CrashDates)
mse <- mse[ , c("GSPC","GSPC_TAE_MTHSNC", "GSPC_RUN_MTHSNC")]
# 20 (or near so) epochs
unique(mse$GSPC_RUN_MTHSNC)

# nothing (no crash date) before 1962
tail(mse["/1962"], 10)

# e.g. (will be used on function
# from crash to crash
# "RUN month(MTH) since(SNC) event"
smse <- split(mse[, c("GSPC", "GSPC_TAE_MTHSNC")], coredata(mse[, "GSPC_RUN_MTHSNC"]))

# 20 periods
length(smse)
[1] 20

# sample testing/demo data
## mse <- mse[114:118,]

# note USUALLY "incBeforeEpochLags == FALSE"
# to "fully" separate ONE epoch from the other.
rollEpochRanks(mse[, c("GSPC", "GSPC_TAE_MTHSNC")], xTs2 = mse[, "GSPC_RUN_MTHSNC"], incBeforeEpochLags = TRUE)

}
}
