% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{Rows}
\alias{Rows}
\title{Allows row indexing without knowledge of dimensionality or class.}
\usage{
Rows(data, rownums, ...)
}
\arguments{
\item{data}{any R object}

\item{rownums}{indices of target rows}

\item{...}{dots may include the user-sent parameter "drop = TRUE"}
}
\description{
\preformatted{

This is an improved version of rowr::rows.
This will not drop tables/arrays/data.frames
of one dimension down to dimensionless vectors.

}
}
\examples{
\dontrun{

Rows(state.x77[, 1, drop = FALSE], 13:15)
         Population
Illinois      11197
Indiana        5313
Iowa           2861

Rows(iris[, 1, drop = FALSE], 13:15)
   Sepal.Length
13          4.8
14          4.3
15          5.8

}
}
