% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{RPC}
\alias{RPC}
\title{relative percent change}
\usage{
RPC(xTs, base = 0, lag = 1, ...)
}
\arguments{
\item{xTs}{xts object}

\item{base}{choose -1 (or less) to look into the future}

\item{lag}{observations backwards}
}
\description{
most useful for tracking: velocity, acceleration, and jerk
To get accelleration and jerk use with
diffXts and differences = 1 and 2 respectively.
}
