% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{runRanks}
\alias{runRanks}
\title{rolling Ranks}
\usage{
runRanks(x, window = 10, ranks = 4, exact.multiplier = 0.5)
}
\arguments{
\item{x}{xts object}

\item{window}{lag to determine the ranks}

\item{ranks}{number of ranks. A lower value
means a lower rank number.}
}
\value{
xts object
}
\description{
TTR::runPercentRank gives skewed results
 (but with the results properly ordered)
 This function uses that "proper ordering" and makes
 usable running ranks.
}
\details{
Note: every so-often this does "screw-up"
RUN the  window = 3, ranks = 3 examples

# Last fortran (maybe better)
https://github.com/joshuaulrich/TTR/blob/9b30395f7604c37ea12a865961d81666bc167616/src/percentRank.f
}
\examples{
\dontrun{

runRanks(xts(sample(10,10,T), zoo::as.Date(0:9)), window = 4)

runRanks(xts(sample(10,10,T), zoo::as.Date(0:9)), window = 3, ranks = 3)

}
}
