% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{runRanksDT}
\alias{runRanksDT}
\title{backward looking and forward looking Ranks using data.table}
\usage{
runRanksDT(x, ranks = 4, cumulative = F, exact.multiplier = 0.5)
}
\arguments{
\item{x}{xts object}

\item{ranks}{4(default) number of ranks. A lower value
means a lower rank number.}
}
\value{
xts object
}
\description{
This is the rank over the "entire" data series (the window).
In general this is not useful. NOT USED ANYWHERE.
}
\examples{
\dontrun{

runRanksDT(xts(sample(10,10,T), zoo::as.Date(0:9)))

runRanksDT(xts(sample(10,10,T), zoo::as.Date(0:9)), ranks = 3)

runRanksDT(xts(sample(10,10,T), zoo::as.Date(0:9)), ranks = 2)

}
}
