% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{runRanksTTR}
\alias{runRanksTTR}
\title{rolling Ranks}
\usage{
runRanksTTR(x, window = 10, ranks = 4, cumulative = F,
  exact.multiplier = 0.5)
}
\arguments{
\item{x}{xts object}

\item{window}{10(default) lag to determine the ranks.}

\item{ranks}{4(default) number of ranks. A lower value
means a lower rank number.}

\item{cumulative}{FALSE(default) use from-inception calculation?
Not tested. So beware.}

\item{exact.multiplier}{The weight applied to identical values
in the window. Must be between 0 and 1, inclusive.
See ? TTR::runPercentRank}
}
\value{
xts object
}
\description{
Wrapper around "TTR::runPercentRank."
TTR::runPercentRank gives skewed values
(but with the value are in the correct order).
This function uses that "proper ordering" and makes
usable running ranks.
}
\details{
# Last fortran (maybe better)
https://github.com/joshuaulrich/TTR/blob/9b30395f7604c37ea12a865961d81666bc167616/src/percentRank.f
}
\examples{
\dontrun{

runRanksTTR(xts(sample(10,10,T), zoo::as.Date(0:9)), window = 4)

runRanksTTR(xts(sample(10,10,T), zoo::as.Date(0:9)), window = 3, ranks = 3)

runRanksTTR(xts(sample(10,10,T), zoo::as.Date(0:9)), window = 4, ranks = 2)

runRanksTTR(xts(sample(10,10,T), zoo::as.Date(0:9)), window = 4, cumulative = TRUE)

}
}
