% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{safeClms}
\alias{safeClms}
\title{get the column names}
\usage{
safeClms(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
zero length vector or column names
}
\description{
\preformatted{

This is used internally.

}
}
