% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{saveSymbols.cache}
\alias{saveSymbols.cache}
\title{save xts objects as .xts in source.envir ( cache )}
\usage{
saveSymbols.cache(source.envir = NULL, target.envir = NULL, ...)
}
\arguments{
\item{source.envir}{location of xts objects}

\item{target.envir}{location of where to store xts objects ( location of cache )}

\item{...}{not used}
}
\description{
save xts objects as .xts in source.envir ( cache )
}
\examples{
\dontrun{
msft <- getSymbols("MSFT", src = "yahoo", auto.assign = FALSE)
source.envir = list2env(list(MSFT = msft))
saveSymbols(Symbols = "", trg = "cache", source.envir = source.envir)
}
}
