% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{saveSymbols.pg}
\alias{saveSymbols.pg}
\title{saveSymbols pg}
\usage{
saveSymbols.pg(Symbols = NULL, con = NULL, source.envir = NULL,
  field.names = c("Open", "High", "Low", "Close", "Volume", "Adjusted"),
  db.fields = c("o", "h", "l", "c", "v", "a"), keys = NULL,
  placeNewRecords = NULL, varHint = NULL, valHint = NULL,
  user = NULL, password = NULL, dbname = NULL, schname = NULL,
  host = "localhost", port = 5432, options = NULL,
  forceISOdate = TRUE, ...)
}
\description{
saveSymbols pg
}
