% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{saveSymbols}
\alias{saveSymbols}
\title{saves xts object symbols to a persistent location (dispatcher)}
\usage{
saveSymbols(Symbols = NULL, env = parent.frame(),
  source.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded}

\item{env}{where to create(save) objects. Setting env=NULL is equal to auto.assign=FALSE}

\item{source.envir}{source location of Symbols}

\item{...}{pass through parameters}
}
\description{
saves xts object symbols to a persistent location (dispatcher)
}
\examples{
\dontrun{

saveSymbols("IBM", trg = "pg", schname = "money",
  source.envir = list2env(list(IBM = getSymbols("IBM", auto.assign = F))))

}
}
