% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{setWhat}
\alias{setWhat}
\title{set OS environment variables or R options}
\usage{
setWhat(x, y = TRUE, Storage = NULL)
}
\arguments{
\item{x}{character vector of size one(1)  of what to set}

\item{y}{TRUE(default): value to set (often TRUE).
Can be a specific value that is stored as a character or not.
E.g. TRUE/"TRUE", 1/"1"}

\item{Storage}{"OS"(default). Otherwise "Options" or "Both"}
}
\value{
TRUE(if success) and FALSE(otherwise)
}
\description{
Set truths or values in memory.
Used as an alternative instead of expensive re-quering.
If x is stored in the OS environment, then x is stored as a string.
To be consistent with OS storage, a value stored in options are
also stored a string.
x is tested using isWhatFnd.
}
\examples{
\dontrun{

setWhat("TAB_X_COL_Y_EXISTS")
setWhat("TAB_X_COL_Y_EXISTS", TRUE)
# not 'preferred set'
setWhat("TAB_X_COL_Y_EXISTS", "TRUE")
# not 'preferred set location'
setWhat("TAB_X_COL_Y_EXISTS", "TRUE", Storage = "Options")
setWhat("TAB_X_COL_Y_EXISTS", "TRUE", Storage = "Both")
setWhat("TAB_X_COL_Y_EXISTS", "TRUE", Storage = c("OS","Options"))

}
}
