% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{simpleCalendar}
\alias{simpleCalendar}
\title{print a simple Calendar}
\usage{
simpleCalendar(R, digits = 1, as.perc = TRUE, geometric = TRUE)
}
\arguments{
\item{R}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{digits}{number of digits to round results to for presentation}

\item{as.perc}{TRUE/FALSE if TRUE, multiply simple returns by 100 to get %}

\item{geometric}{utilize geometric chaining (TRUE) or simple/arithmetic chaining (FALSE)
to aggregate returns, default TRUE}
}
\value{
xts object of arithmatic returns
}
\description{
\preformatted{

Based on PerformanceAnalytics table.CalendarReturns.
It does not do aggregate column marginals
}
}
