% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{SortinoRatioLoss}
\alias{SortinoRatioLoss}
\title{for package iml function FeatureImp$new
True Sortino Ratio 'loss function'}
\usage{
SortinoRatioLoss(actual, predicted, ...)
}
\arguments{
\item{...}{lev = NULL, model = NULL : passed to SortinoRatioSummary}

\item{see}{? iml::FeatureImp}

\item{see}{? iml::FeatureImp}
}
\description{
for package iml function FeatureImp$new
True Sortino Ratio 'loss function'
}
