% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{SortinoRatioSummary}
\alias{SortinoRatioSummary}
\title{for package caret: True Sortino Ratio Summary}
\usage{
SortinoRatioSummary(data, lev = NULL, model = NULL)
}
\description{
for package caret: True Sortino Ratio Summary
}
