% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{SP500EomData}
\alias{SP500EomData}
\title{Get the S&P 500 Index eom price from Yahoo}
\usage{
SP500EomData()
}
\value{
xts object of end of month returns
}
\description{
\preformatted{

}
}
