% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{SP500EyeBallWts}
\alias{SP500EyeBallWts}
\title{add SP500 weights using eyeball}
\usage{
SP500EyeBallWts(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

This is the workhorse function. This is where the magic/logic happens.
Use any other columns (called indicators) that do not have the weights (_wts)
suffix and do not have the same root name compared to each and every
other *_wts column.

}
}
