% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary_predictive.R
\name{specifyModel}
\alias{specifyModel}
\title{quantmod specifyModel with parameter source.envir}
\usage{
specifyModel(formula, na.rm = TRUE, source.envir = NULL, ...)
}
\description{
get some Symbols from an environment (source.envir)
  will search first in (source.envir)
  if the Symbol is not found in the enviroment (source.envir),
  then get the Symbol from elsewhere
...
passed to getSymbols
        useful: from, to
  maybe useful: src
  maybe useful: set per Symbol src using setSymbolLookup
}
\details{
specifyModel(getModelData) original code ( not changed by me )
if na.rm == TRUE,  then does 'na.exclude'
  BUT 'without rules' PUTS back (rbind) the last observation

NOTE
TTR acceptable
avoid
specifyModel
Error in runSum(x, n) : Series contains non-leading NAs
zoo::na.trim(tail(x), sides = "right")

example ( better: try a do.call )
Symbols <- unlist(lapply( c("MSFT","AAPL","WMT","COST"), function(x) {
                          l <- list(); l[[x]] <- getSymbols(x, auto.assign = FALSE); l
                          }), recursive = FALSE)
Symbols <- list2env(Symbols)
ls.str(Symbols)

getSymbols( c("AAPL","ORCL"), source.envir = Symbols)
rm("AAPL","ORCL")

quantmod <- specifyModel(Next(ClCl(WMT)) ~ Lag(OpCl(AAPL)) + Lag(LoHi(COST),0:2)
  , source.envir = Symbols
  , from ="2007-01-01"
  , to   ="2011-12-31"
  )
}
