% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{sumOrdersXts}
\alias{sumOrdersXts}
\title{specific orders of n  (smallest/largest) of values within a range}
\usage{
sumOrdersXts(x, base = 0, r = 0:1, n = 1, View = "min",
  na.pad = TRUE, ...)
}
\arguments{
\item{x}{xts object}

\item{base}{passed to LagXts.  See ? LagXts.}

\item{r}{(back) range: choose -2:-1 (or less) to look into the future.
Passed to LagXts k paramter}

\item{n}{the order of the minimum to be found.
n == 1 means smalled value. n == 2 means second smalled
value, etc.}

\item{View}{"min"(default): 'n == 1' means look for the lowest value(s).
Otherwise, "max" n == 1 means look for the highest values(s).
Note, to peform "max" math, the coredata is simply muliplied by
negative one(-1) before the code is executed.
After the code is executed the coredata is re-multiplied by
negative one(-1).}

\item{na.pad}{passed to LagXts}

\item{...}{dots passed to LagXts}

\item{rt}{alternate way to specify rt == 2 means r == 1:2.
rt == -2 means r == -2:-1.  rt will override r.  This can be
useful when passing to eXplodeXts}

\item{nt}{alternate way to specify nt == 2 means n == 1:2.
nt == -2 means n == 2:1.  nt will override n.  This can be
useful when passing to eXplodeXts.}
}
\description{
Within a range, find the smallest observation
of order (n). n == 1 smallest. n == 2 2nd smallest.
}
\details{
within range, within a subrange of smallest
observations defined by nt, find the sum
of those small values.

Expected, to be used as a smoothing
item on past St. Louis FRED data.  Typical use
is to choose the minimal of the previous two(2)
observations before/or/after applying an SMA(x, 3).

Other, expected be used on future data. Find
the lowest valued two(2) months in the next(future)
six(6) months and add(sum) those values together.
}
\examples{
\dontrun{

matrix(1:5,ncol = 1, dimnames = list(list(), list("Data")))

           Data
1970-01-01    1
1970-01-02    2
1970-01-03    3
1970-01-04    4
1970-01-05    5

# futures
LagXts(xts(matrix(1:5,ncol = 1, dimnames = list(list(), list("Data"))), zoo::as.Date(0:4)),
  k =  c(-3:-1))

          lead3 lead2 lead1
1970-01-01     4     3     2
1970-01-02     5     4     3
1970-01-03    NA     5     4
1970-01-04    NA    NA     5
1970-01-05    NA    NA    NA

# of the three(3) (future)leading values I want to find two(2) of
# the three(2) worst months and sum their 'two' values together.

# nt == 2 means n == 1:2 (1st lowest and 2nd lowest)
#
sumOrdersXts(xts(matrix(1:5,ncol = 1, dimnames = list(list(), list("Data"))), zoo::as.Date(0:4)),
  r =  -3:-1, nt = 2)

          Data_RAGG
1970-01-01         5 # lead2 + lead1
1970-01-02         7 # lead2 + lead1


# find two(2) of the three(2) best months and
# sum their 'two' values together.
#
# View == "max" # n == 1 means greatest
#
sumOrdersXts(xts(matrix(1:5,ncol = 1, dimnames = list(list(), list("Data"))), zoo::as.Date(0:4)),
  r =  -3:-1, View = "max", nt = 2)

           Data_RAGG
1970-01-01         7 # lead3 + lead2
1970-01-02         9 # lead3 + lead2

}
}
