% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{symbolAPCReturns}
\alias{symbolAPCReturns}
\title{symbols eom absolute proportional change from the previous month}
\usage{
symbolAPCReturns(Symbol = NULL, src = NULL,
  SymplifyGeneratorFUN = NULL)
}
\description{
symbols eom absolute proportional change from the previous month
}
