% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{symbolData}
\alias{symbolData}
\title{getSymbol data from the target}
\usage{
symbolData(Symbol = NULL, src = NULL, New = NULL, NewMaxAge = NULL,
  ...)
}
\arguments{
\item{Symbol}{single target symbol as a string}

\item{src}{source of the data: Curently, "yahoo"(default).  Other: "FRED"}

\item{New}{TRUE(default) of the Symbol gets the data and stores the data
through the method of getNewSymbols (TRUE).  Otherwise,
get the data and store the data through the method of getSymbols (FALSE).}

\item{NewMaxAge}{default("4 hours"), if New = TRUE, then this
is the timeout period of the stored data. After NewMaxAge time has passed,
the the data will be re-queried from Yahoo corporation.
of the data}

\item{...}{dots passed}
}
\value{
xts object of results
}
\description{
\preformatted{

}
}
\examples{
\dontrun{

# St. Louis FRED

head(symbolData("GDP", src = "FRED"),1)
                          GDP
1947-01-01 243.16399999999999

Get all of the records from the "cache" or "pg".
If the cache and pg is older than 4 hours then
the data will be acquired anew from the source
and loaded into the "pg" and the "cache" and returned
to the user.

default New == TRUE (default)
means use the function get\\*New\\*Symbols
checks the xts "updated" attribute
  checks (1)cache then (2)pg if data is older than NewMaxAge = "4 hours"
  checks "cache": manually check cache by ls(all.names = TRUE)
  checks "pg": manually check database by using: SELECT * FROM "Symbols"."Symbols"
then will get the new data from the "cache" or "pg".

symbolData(Symbol = "UNRATE", src = "FRED")

# Full test . . .
# 1. remove cache data: rm(.UNRATE) (if there)
# 2. drop database table "Symbols"."UNRATE"
# 3. remove corresponding record from "Symbols"."Symbols"
# 4. symbolData(Symbol = "UNRATE", src = "FRED")

# Partial tests

2.1 Manually remove some (few) bottom records from the
database table "Symbols"."UNRATE"
Below just insert only-new (few) records
# placeNewRecords == "AddOnlyNew" (default)

symbolData(Symbol = "UNRATE", src = "FRED", NewMaxAge = "1 secs")

# or

# 2.2
# placeNewRecords =="TruncateTable"

symbolData(Symbol = "UNRATE", src = "FRED", NewMaxAge = "1 secs", placeNewRecords = "TruncateTable")

# Yahoo

symbolData(Symbol = "^GSPC"

symbolData(Symbol = "^GSPC", NewMaxAge = "1 secs")

symbolData(Symbol = "^GSPC", NewMaxAge = "1 secs", placeNewRecords = "TruncateTable")

}
}
