% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{symbolEomData}
\alias{symbolEomData}
\title{get the latest end of month value in the month from the target}
\usage{
symbolEomData(Symbol = NULL, src = NULL)
}
\arguments{
\item{Symbol}{target symbol as a string}

\item{src}{See ? symbolData}
}
\value{
xts object of the last observation with the
month date rounded up to the last day of the month
}
\description{
\preformatted{

}
}
\examples{
\dontrun{

# weekly
ff <- head(symbolEomData("FF", src = "FRED"),1)
# ff
#              FF
# 1954-07-31 0.63

# weekly
gspc <- head(symbolEomData("^GSPC"),1)
# gspc

}
}
