% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{symbolMachineWts}
\alias{symbolMachineWts}
\title{add Symbol Change weights using Machine learning}
\usage{
symbolMachineWts(xTs = NULL, Predictee = NULL, Predictors = NULL,
  IndicatorGeneratorFUN = NULL, NumbReplicaCopiesMultiple = NULL, ...)
}
\arguments{
\item{xTs}{xts object}

\item{Predictee}{string of the name of the Predictee variable}

\item{Predictors}{characters vector of the require columns needed in the
indictor generator function: IndicatorGeneratorFUN}

\item{IndicatorGeneratorFUN}{string of the name of the function
or the function itself that generates more columns (to eventually
be used in the model.)}

\item{NumbReplicaCopiesMultiple}{passed to NumbReplicaCopies
This means how much more ( e.g. "focused" data ) is replicated
compared to "all" data.}

\item{...}{dots passed to the indicator generator function}
}
\value{
xts object
}
\description{
add Symbol Change weights using Machine learning
}
