% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{symbolReturns}
\alias{symbolReturns}
\title{symbols eom absolute proportional change from the previous month}
\usage{
symbolReturns(Symbol = NULL, src = NULL, SymplifyGeneratorFUN = NULL,
  ReturnsGeneratorFUN = NULL)
}
\description{
symbols eom absolute proportional change from the previous month
}
