% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{timeSliceNBER}
\alias{timeSliceNBER}
\title{NBER timeslices}
\usage{
timeSliceNBER(allSlicesStart = NULL, allSlicesEnd = NULL,
  LongTimeSlices = NULL, LongestTimeSlice = NULL)
}
\arguments{
\item{allSlicesStart}{NULL(default), Date of the earlies possbile date.
This Date is the first day(1st) of a month. Note: this filter is applied late.}

\item{allSlicesEnd}{NULL(default), Date of the latest possbile date.
This Date is the last day(last) of a month. Note: this filter is applied late.}

\item{LongTimeSlices}{FALSE(default), if TRUE, include non-recession range that
is before this recession range.}

\item{LongestTimeSlice}{FALSE(default), if TRUE then the start value is the
beginning of "NBER dates" (and limited by allSlicesStart)
LongestTimeSlice = TRUE and LongTimeSlices = TRUE ARE mutually exclusive choices
of each other}
}
\value{
a list of vectors of Dates of recession Ranges
}
\description{
NBER timeslices
}
\examples{
\dontrun{
# str(timeSliceNBER())
# str(timeSliceNBER(allSlicesStart = zoo::as.Date("1969-12-31")))
# str(timeSliceNBER(allSlicesStart = zoo::as.Date("1969-12-31"), LongTimeSlices = TRUE))
# str(timeSliceNBER(allSlicesStart = zoo::as.Date("1969-12-31"), LongestTimeSlice = TRUE))
}
}
