% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{timeSlices}
\alias{timeSlices}
\title{time slices}
\usage{
timeSlices(timeSliceMatrix = NULL, allSlicesStart = NULL,
  allSlicesEnd = NULL, LongTimeSlices = NULL,
  LongestTimeSlice = NULL, OmitSliceFirstDate = NULL)
}
\arguments{
\item{timeSliceMatrix}{NULL(default) required.
Two column matrix of start and end dates.  The left column is named "Start."
The right column is named "End". The cell values are Real values in format YYYYMMDD.
The "Start" column contains beginning of month dates.
The "End" column contains end of month dates.}

\item{allSlicesStart}{NULL(default), Date of the earlies possible date.
This Date is the first day(1st) of a month. Note: this filter is applied LAST.}

\item{allSlicesEnd}{NULL(default), Date of the latest possbile date.
This Date is the last day(last) of a month. Note: this filter is applied late.}

\item{LongTimeSlices}{FALSE(default). If TRUE, include non-date range that
is before this date range.}

\item{LongestTimeSlice}{FALSE(default). If TRUE then the start value is the
beginning of "NBER dates" (and limited by allSlicesStart)
LongestTimeSlice = TRUE and LongTimeSlices = TRUE ARE mutually exclusive choices
of each other}

\item{OmitSliceFirstDate}{FALSE(default), All dates are end of month dates.
A previous timeslice tailing date is the the same date as the next timeslice heading date.
TRUE omits the heading date from each time slice.  This may be necessary to prevent
repetition of data, for example in sending timeslices to machine learning models.}
}
\value{
a list of vectors of Dates of date Ranges
}
\description{
\preformatted{
Given a matrix(Start, End) of cell Real values of YYYYMMDD.
Return a list of date ranges.
}
}
\examples{
\dontrun{
str(timeSlices(timeSliceMatrix = tis::nberDates()))
str(timeSlices(timeSliceMatrix = tis::nberDates(), allSlicesStart = zoo::as.Date("1969-12-31")))
str(timeSlices(timeSliceMatrix = tis::nberDates(), allSlicesStart = zoo::as.Date("1969-12-31")
  , LongTimeSlices = TRUE)
)
str(timeSlices(timeSliceMatrix = tis::nberDates(), allSlicesStart = zoo::as.Date("1969-12-31")
  , LongTimeSlices = TRUE, OmitSliceFirstDate = TRUE)
)
str(timeSlices(timeSliceMatrix = tis::nberDates(), allSlicesStart = zoo::as.Date("1969-12-31")
, LongestTimeSlice = TRUE)
)
# now after seeing all of the date, THEN, choose the filter
str(timeSlices(timeSliceMatrix = tis::nberDates(), allSlicesStart = zoo::as.Date("1969-12-31")
, LongestTimeSlice = TRUE, allSlicesStart = ?, allSlicesEnd = ?)
)
}
}
