% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{To.Monthly}
\alias{To.Monthly}
\title{Convert time series data to an monthly OHLC series}
\usage{
To.Monthly(x, indexAt = "yearmon", drop.time = TRUE, name,
  fillMissing = NULL, ...)
}
\arguments{
\item{x}{? xts::to.monthly and the index(x) class must
have an S3 method for "seq(by = "months")".
Classes "Date" and POSIXt are provided  S3 methods in package "base".}

\item{indexAt}{See ? xts::to.monthly}

\item{drop.time}{See ? xts::to.monthly}

\item{name}{See ? xts::to.monthly}

\item{fillMissing}{TRUE(default). If indexAt is one of "firstof"
or "lastof", then fill in missing month dates, if any.
Otherwise FALSE, just pass throught to xts::to.monthly.}

\item{...}{dots See. ? xts::to.monthly}
}
\value{
See. ? xts::to.monthly
}
\description{
\preformatted{

This is a wrapper around xts to.monthly.
Months between the first date and the last
date will have and entry filled in.
It will get the "firstof" or "lastof"
months between dates as the same in the call.

}
}
\examples{
\dontrun{
x <- xts(c(1,140), zoo::as.Date(c(0,139)))
To.Monthly(x, OHLC = FALSE, indexAt = "firstof")
}
}
