% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{toMonthlyData}
\alias{toMonthlyData}
\title{get the latest value in the month}
\usage{
toMonthlyData(xTs = NULL)
}
\arguments{
\item{xTs}{xts object.
(currently) eomData only works on single column xtx objects.}
}
\value{
xts object of the last observation with the
month date rounded up to the last day of the month
}
\description{
\preformatted{

}
}
\examples{
\dontrun{
# library(xts)
# > xTs <- xts(c(1,NA_real_,2), zoo::as.Date(c(1,11,21)))
# > xTs
#            [,1]
# 1970-01-02    1
# 1970-01-12   NA
# 1970-01-22    2
#
# > eomData(xTs = xTs)
#            [,1]
# 1970-01-31    2
}
}
