% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{trends}
\alias{trends}
\title{trends}
\usage{
trends(xTs, col = NULL, UseAOMX = FALSE)
}
\arguments{
\item{col}{column of the mktdata object to perform the trends}

\item{x}{mktdata object}
}
\description{
trends
}
\examples{
\dontrun{

# called from a 'main'
trends(mktdata, "UNRATE", UseAOMX = TRUE)

}
}
