% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{trendsWithAOMIndicators}
\alias{trendsWithAOMIndicators}
\alias{trendsWithAOMXIndicators}
\alias{trendsWithAOMNIndicators}
\title{collection of curves of the trends Indicator using average order}
\usage{
trendsWithAOMIndicators(x = NULL, ...)

trendsWithAOMXIndicators(x = NULL, ...)

trendsWithAOMNIndicators(x = NULL, ...)
}
\arguments{
\item{x}{xts object}
}
\description{
\preformatted{

}
}
