% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{trimLeadingNAGaps}
\alias{trimLeadingNAGaps}
\title{trim leading NAs and leading gaps of NAs}
\usage{
trimLeadingNAGaps(xTs)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object
}
\description{
Some FRED early data was delivered less
frequently than later FRED data.
Remove that earlier wider-gapped data.
}
\details{
Also, many functions in R CRAN package TTR
will return an error if leading NA values are found
}
