% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{trueSortinoRatio}
\alias{trueSortinoRatio}
\title{(true) sortino ratio}
\usage{
trueSortinoRatio(x, MinRows, rf = 0, na.rm = FALSE)
}
\description{
Sortino Ratio: Are you calculating it wrong?
https://www.rcmalternatives.com/2013/09/sortino-ratio-are-you-calculating-it-wrong/
}
\details{
I CAN GET RID OF THIS FUNCTION

# THIS WORKS CORRECTLY "AS IS"
Add feature zeroMAR to SortinoRatio ( actually DownsideDeviation ) #106
https://github.com/braverock/PerformanceAnalytics/issues/106
}
\examples{
\dontrun{
# trueSortinoRatio(x, n, rf = 0.0, na.rm = FALSE)
}
}
