% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{UMCSentimentData}
\alias{UMCSentimentData}
\title{University of Michigan: Consumer Sentiment}
\usage{
UMCSentimentData()
}
\value{
xts object
}
\description{
From FRED (historical) and briefing.com (recent)
}
\details{
Recent data is from
Univ. of Michigan Consumer Sentiment - Prelim:
\cite{Univ. of Michigan Consumer Sentiment - Prelim \url{https://www.briefing.com/investor/calendars/economic/releases/mich.htm}}

Original data is from here:
\cite{surveys of consumers UNIVERSITY OF MICHIGAN \url{https://data.sca.isr.umich.edu/tables.php}}

Historical data is from here.
At the request of the source, the data is delayed by 1 month:
\cite{University of Michigan: Consumer Sentiment \url{https://fred.stlouisfed.org/series/UMCSENT}}

\cite{University of Michigan: Consumer Sentiment raw data \url{https://fred.stlouisfed.org/data/UMCSENT.txt}}
}
