% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{UMCSentimentEomData}
\alias{UMCSentimentEomData}
\title{end of month University of Michigan: Consumer Sentiment
from FRED (historical) and briefing.com (recent)}
\usage{
UMCSentimentEomData()
}
\value{
xts object
}
\description{
end of month University of Michigan: Consumer Sentiment
from FRED (historical) and briefing.com (recent)
}
