% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{uninitEnv}
\alias{uninitEnv}
\title{unsets the enviroment
space-saver - meant to be used at the beginning of a function}
\usage{
uninitEnv()
}
\value{
options, and TZ are un-set
}
\description{
Variable ops from the calling environment resets R options
Environment variable TZ using the calling environment variable oldtz
resets.
}
\examples{
\dontrun{
# > uninitEnv()
# getOption("digits")
# [1] 5
}
}
