% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{UnRateEyeBalltradeModel}
\alias{UnRateEyeBalltradeModel}
\title{Predicts the FRED WILL5000IND eom returns using UNRATE and the eyeball}
\usage{
UnRateEyeBalltradeModel()
}
\value{
xts object of monthly return results
}
\description{
Predicts the FRED WILL5000IND eom returns using UNRATE and the eyeball
}
\examples{
\dontrun{
EXAMPLE
# EXAMPLE
}
}
