% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{UnRateEyeBalltradeModelGSPC}
\alias{UnRateEyeBalltradeModelGSPC}
\title{Predicts the Yahoo SP500 eom returns using UNRATE and the eyeball}
\usage{
UnRateEyeBalltradeModelGSPC()
}
\value{
xts object of monthly return results
\describe{
  \item{One}{First item}
  \item{Two}{Second item}
}
}
\description{
Predicts the Yahoo SP500 eom returns using UNRATE and the eyeball
}
\examples{
\dontrun{
EXAMPLE
# EXAMPLE
}
}
