% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{UnRateMachinetradeModelGSPC}
\alias{UnRateMachinetradeModelGSPC}
\title{Predicts the SP500 eom returns using UNRATE and Machine learning}
\usage{
UnRateMachinetradeModelGSPC()
}
\value{
xts object of monthly return results
}
\description{
Predicts the SP500 eom returns using UNRATE and Machine learning
}
\examples{
\dontrun{
EXAMPLE
# EXAMPLE
}
}
