% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{UnRateMachinetradeModelWILL5000IND}
\alias{UnRateMachinetradeModelWILL5000IND}
\title{Predicts the FRED WILL5000IND eom returns using UNRATE and Machine learning}
\usage{
UnRateMachinetradeModelWILL5000IND()
}
\description{
Predicts the FRED WILL5000IND eom returns using UNRATE and Machine learning
}
