% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{unsetWhat}
\alias{unsetWhat}
\title{unset OS environment variables or R options}
\usage{
unsetWhat(x, Storage = NULL)
}
\arguments{
\item{x}{character vector of size one(1)  of what to unset}

\item{Storage}{"OS"(default). Otherwise "Options" or "Both"}
}
\value{
NULL invisibly
}
\description{
Unset truths or values in memory.
}
\examples{
\dontrun{

unsetWhat("TAB_X_COL_Y_EXISTS")
# not 'preferred unset location'
unsetWhat("TAB_X_COL_Y_EXISTS", Storage = "Options")
unsetWhat("TAB_X_COL_Y_EXISTS", Storage = "Both")
unsetWhat("TAB_X_COL_Y_EXISTS", Storage = c("OS", "Options"))

}
}
