% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{updatedSymbols.cache}
\alias{updatedSymbols.cache}
\title{"updated" property of xts object symbols in the cache}
\usage{
updatedSymbols.cache(Symbols = NULL, cache.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{source.envir}{location of xts objects}
}
\value{
a named list of "updated" properties
}
\description{
\preformatted{

}
}
