% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{updatedSymbols.pg}
\alias{updatedSymbols.pg}
\title{updatedSymbols pg}
\usage{
updatedSymbols.pg(
  Symbols = NULL,
  con = NULL,
  user = NULL,
  password = NULL,
  dbname = NULL,
  schname = NULL,
  host = "localhost",
  port = 5432,
  options = NULL,
  forceISOdate = TRUE,
  ...
)
}
\description{
updatedSymbols pg
}
