% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{updatedSymbols.PostgreSQL}
\alias{updatedSymbols.PostgreSQL}
\title{'updated' property of symbols in the PostgreSQL (pg) database}
\usage{
updatedSymbols.PostgreSQL(Symbols = "", con = NULL, user = NULL,
  password = NULL, dbname = NULL, schname = NULL,
  host = "localhost", port = 5432, options = NULL,
  forceISOdate = TRUE, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{con}{PostgreSQL DBI connection}

\item{user}{username (default "Symbols") to access database}

\item{password}{password (default "Symbols") to access database}

\item{dbname}{database name (default "Symbols")}

\item{schname}{schema name (default "Symbols")}

\item{host}{database host (default "localhost")}

\item{port}{database port (default 5432)}

\item{options}{pass extra parameters in a string to the command line}

\item{forceISOdate}{TRUE(default)/FALSE if the communication of date (time stamp) from PostgreSQL
is forced to ISO style at conection}

\item{...}{passed unused}
}
\value{
named list of the object 'updated' property
}
\description{
'updated' property of symbols in the PostgreSQL (pg) database
}
