% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{updatedSymbols}
\alias{updatedSymbols}
\title{'updated' date-ish/time-ish property
of an [xts] object symbols in a persistent location (dispatcher)}
\usage{
updatedSymbols(Symbols = NULL, source.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{source.envir}{source location of Symbols}

\item{...}{passed to src}

\item{src}{source program to look for symbols}
}
\value{
a named list of 'updated' properties
}
\description{
looks in source.envir
}
\examples{
\dontrun{

# Symbols must be case-insenstive unique

getSymbols("IBM", src = "yahoo") # auto.assign = TRUE

# Symbols names found in Symbols and the names of xts objects stored
# in source.envir, must be mutually exclusive ( case in-sensensitive match)
# (because list2env uses a case-insenstive comparison and silently drops repeated members)

# save just the Symbol "IBM"
saveSymbols("IBM", trg = "pg")

msft <- getSymbols("MSFT", src = "yahoo", auto.assign = FALSE)
source.envir = list2env(list(MSFT = msft))
listSymbols(source.envir = source.envir)
updatedSymbols(Symbols  = "MSFT", source.envir = source.envir)
existSymbols(Symbols = "MSFT", source.envir = source.envir)

saveSymbols(trg = "cache", source.envir = source.envir)
listSymbols(src = "cache")
updatedSymbols(Symbols  = "MSFT",src = "cache")
existSymbols(Symbols = "MSFT", src = "cache")

# save all of the .getSymbols Symbols and the source.envir Symbols
saveSymbols(trg = "pg", source.envir = source.envir)
listSymbols(src = "pg")
updatedSymbols(Symbols  = "MSFT",src = "pg")
existSymbols(c("MSFT","IBM"), src = "pg")

}
}
