% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{valueClms}
\alias{valueClms}
\title{get the values columns names}
\usage{
valueClms(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
column names
}
\description{
Values column names have an associated column with the
same root.  However the associated column always ends with
the suffix "_wts"
}
\examples{
\dontrun{
# > require(xts)
# > xTs <- xts(matrix(1:3, ncol = 3, dimnames = list(NULL,c("a","b","b_wts"))),zoo::as.Date(0))[0]
# > valueClms(xTs)
# [1] "b"
}
}
