% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{valuePredictorClms}
\alias{valuePredictorClms}
\title{get the un-massaged original 'predictor' columns}
\usage{
valuePredictorClms(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
column names
}
\description{
\preformatted{

}
}
