% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{weightRankings}
\alias{weightRankings}
\title{weights}
\usage{
weightRankings(xTs, target, TrainStart, TrainEnd, weightings = NULL,
  probs = NULL, CaseAdj = NULL)
}
\arguments{
\item{xTs}{data}

\item{target}{the column being predicted}

\item{TrainStart}{absolute training start date}

\item{TrainEnd}{absolute training end date}

\item{weightings}{passed to Hmisc::wtd.quantile weights}

\item{probs}{passed to Hmisc::wtd.quantile probs
Default is seven(7) intervals of probs = c(0.00, 0.01, 0.10, 0.25, 0.75, 0.90, 0.99, 1.00)}

\item{CaseAdj}{new (re)values for values of the intervals 7:1 (from probs)
Must be passed to dplyr::case_when as a quoted list of formulas.
Passed as formula elements:  BareWeightRankings == <oldvalue> ~ <newvalue>
See the default (in the code: tradeModel::AdjustedWeightRankings )}
}
\description{
xgboost weights ( using objective(y hieght) value to determine 'how much I care'(weights))
The weights are then
}
\details{
simply multiplied by the classification error at each iteration of the learning process.

Gradient boosting machines, a tutoria
Front Neurorobot. 2013; 7: 21.
Published online 2013 Dec 4. doi:  10.3389/fnbot.2013.00021
PMCID: PMC3885826l
http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3885826/
}
