% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{willShire5000Wts}
\alias{willShire5000Wts}
\title{add Willshire 5000 Index weights}
\usage{
willShire5000Wts(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
This is the workhorse function. This is where the magic/logic happens.
Use any other columns (called indicators) that do not have the weights (_wts)
suffix and do not have the same root name compared to each and every
other *_wts column.
}
