% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{wilshire5000indEomData}
\alias{wilshire5000indEomData}
\title{Get the Wilshire 5000 Index eom price from FRED}
\usage{
wilshire5000indEomData()
}
\value{
xts object of end of month returns
}
\description{
\preformatted{

}
}
