% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{wilshire5000LogReturns}
\alias{wilshire5000LogReturns}
\title{get the Wilshare 5000 Index log returns}
\usage{
wilshire5000LogReturns()
}
\value{
xts object
}
\description{
get the Wilshare 5000 Index log returns
}
