% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{within.xts}
\alias{within.xts}
\title{inside and exts environment}
\usage{
\method{within}{xts}(data, expr, ...)
}
\description{
unstable structure to debug inside: holds for 2 seconds then bounces out
within.xts IS TOO VOLITILE: CAN NOT browser/rstudio debug inside: SOMETHING IS NOT RIGHT
}
