% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{wtsClms}
\alias{wtsClms}
\title{get the weights(_wts) columns}
\usage{
wtsClms(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
column names
}
\description{
Weights column names always ends in "_wts"
CURRENTLYNOT USED ANYWHERE
}
\examples{
\dontrun{
# > require(xts)
# > xTs <- xts(matrix(1:3, ncol = 3, dimnames = list(NULL,c("a","b","b_wts"))),zoo::as.Date(0))[0]
# > wtsClms(xTs)
# [1] "b_wts
}
}
