% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{wtsCurrentRetsClms}
\alias{wtsCurrentRetsClms}
\title{get the currentrets_wts columns}
\usage{
wtsCurrentRetsClms(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
column names
}
\description{
\preformatted{

Weights column names always end in "currentrets_wts"

}
}
\examples{
\dontrun{
# xTs <- xts(matrix(1:3, ncol = 3, dimnames = list(NULL,c("a","b","b_currentrets_wts"))),zoo::as.Date(0))[0]
# wtsCurrentRetsClms(xTs)
# [1] "b_currentrets_wts"
}
}
