% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{wtsLeadingRetsClms}
\alias{wtsLeadingRetsClms}
\title{get the leadingrets_wts columns}
\usage{
wtsLeadingRetsClms(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
column names
}
\description{
\preformatted{

Weights column names always end in "leadingrets_wts".

}
}
\examples{
\dontrun{
# xTs <- xts(matrix(1:3, ncol = 3, dimnames = list(NULL,c("a","b","b_leadingrets_wts"))),zoo::as.Date(0))[0]
# wtsLeadingRetsClms(xTs)
# [1] "b_leadingrets_wts"
}
}
