% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{xTs2DBDF}
\alias{xTs2DBDF}
\title{convert an OHLC xTs into aN OHLC data.frame MEANT to be loaded into a database}
\usage{
xTs2DBDF(xTs, con)
}
\arguments{
\item{xTs}{OHLC[V][A] object}

\item{con}{DBI database connection}
}
\description{
convert an OHLC xTs into aN OHLC data.frame MEANT to be loaded into a database
}
