% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{xTsCols2SymbolsEnv}
\alias{xTsCols2SymbolsEnv}
\title{column split an xts object into
an environment of column-xts objects}
\usage{
xTsCols2SymbolsEnv(xTs)
}
\arguments{
\item{xTs}{xts object}
}
\description{
uses S3 function: xts:::as.data.frame.xts
}
