% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{yahooData}
\alias{yahooData}
\title{get data from Yahoo corporation}
\usage{
yahooData(Symbol = NULL, New = NULL, NewMaxAge = NULL, ...)
}
\arguments{
\item{Symbol}{FRED symbol as a string}

\item{New}{TRUE(default) of the Symbol gets the data and stores the data
through the method of getNewSymbols (TRUE).  Otherwise,
get the data and store the data through the method of getNewSymbols (FALSE).}

\item{NewMaxAge}{default("4 hours"), if New = TRUE, then this
is the timeout period of the stored data. After NewMaxAge time has passed,
the the data will be re-queried from Yahoo corporation.
of the data}

\item{...}{dots passed}
}
\value{
xts object of results
}
\description{
\preformatted{

 WORKS just like fredData

}
}
\examples{
\dontrun{

# WORKS just like fredData

yahooData(Symbol = "^GSPC"

yahooData(Symbol = "^GSPC", NewMaxAge = "1 secs")

yahooData(Symbol = "^GSPC", NewMaxAge = "1 secs", placeNewRecords = "TruncateTable")

}
}
