% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{yahooEomData}
\alias{yahooEomData}
\title{get the latest value in the month from FRED}
\usage{
yahooEomData(Symbol = NULL)
}
\arguments{
\item{Symbol}{FRED symbol as a string}
}
\value{
xts object of the last observation with the
month date rounded up to the last day of the month
}
\description{
\preformatted{

# Units: Index, Not Seasonally Adjusted
# Frequency: Daily, Close

}
}
\examples{
\dontrun{
EXAMPLE
# # weekly
# > gspc <- head(yahooEomData("^GSPC"),1)
# > gspc
#
}
}
