% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{AC}
\alias{AC}
\title{absolute change}
\usage{
AC(x, base = 0, lag = 1, ...)
}
\arguments{
\item{x}{xts object}

\item{base}{choose -1 (or less) to look into the future}

\item{lag}{observations backwards}

\item{...}{dots passed to LagXts}
}
\description{
\preformatted{

}
}
\examples{
\dontrun{

matrix(1:4,ncol =2)
     [,1] [,2]
[1,]    1    3
[2,]    2    4

AC(xts(matrix(1:4,ncol =2), zoo::as.Date(0:1)), lag = 1:2)
           V1ac.1 V2ac.1 V1ac.2 V2ac.2
1970-01-01     NA     NA     NA     NA
1970-01-02      1      1     NA     NA

}
}
