% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addCashLogReturns}
\alias{addCashLogReturns}
\title{add cash log returns (cashlogrets)}
\usage{
addCashLogReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with merged data into xTs
}
\description{
add cash log returns (cashlogrets)
}
