% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{addCurrLeadSymbolAPCReturns}
\alias{addCurrLeadSymbolAPCReturns}
\title{add leading and current Symbol APC returns (SYMBOLapcrets)}
\usage{
addCurrLeadSymbolAPCReturns(xTs = NULL, Symbol = NULL, src = NULL,
  IsTarget = NULL, IsATarget = NULL, SymplifyGeneratorFUN = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{Symbol}{getSymbols Symbol}

\item{src}{getSymbols source}

\item{IsTarget}{is the objective}

\item{IsATarget}{is a member of many possible objectives}

\item{SymplifyGeneratorFUN}{Function that Formats the output}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

}
}
