% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{APC}
\alias{APC}
\title{absolute percent change}
\usage{
APC(x, base = 0, lag = 1, ...)
}
\arguments{
\item{x}{xts object}

\item{base}{choose -1 (or less) to look into the future}

\item{lag}{observations backwards}
}
\description{
\preformatted{

This is most useful for calculating velocity
and acceleration, (and jerk).
To get accelleration and jerk use this with
diffXts with differences 2 (and 3) respectively.

Ninety-nine (99%) percent of the people in the world SHOULD HAVE BEEN using this one.

}
}
\examples{
\dontrun{

xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2))
           [,1] [,2]
1970-01-01    1    8
1970-01-02   -2   16
1970-01-03   -4   32

APC(xts(matrix(c(1,-2,-4,8,16,32), ncol = 2), zoo::as.Date(0:2)))
           V1apc.1 V2apc.1
1970-01-01      NA      NA
1970-01-02      -3       1
1970-01-03      -1       1

}
}
