% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{as.quantmod}
\alias{as.quantmod}
\alias{as.quantmod.default}
\alias{as.quantmod.data.frame}
\title{converts to a quantmod object}
\usage{
as.quantmod(x, outcomename, order.by, na.rm = TRUE, ...)

\method{as.quantmod}{default}(x, outcomename, order.by, na.rm = TRUE, ...)

\method{as.quantmod}{data.frame}(x, outcomename, order.by, na.rm = TRUE, ...)
}
\description{
\preformatted{

quantmod class conversion

}
}
\details{
\preformatted{

CURRENLY NOT USED

If I have to do preprocessing ( e.g. a treatment )
  therefore the situation may be cheaper to make a quantmod object
    from a data.frame.

if na.rm == TRUE, then does 'na.exclude'
  BUT 'without rules' PUTS back (rbind) the last observation

data(sample_matrix)
sample_xts <- as.xts(sample_matrix)
quantmodSample <- as.quantmod(as.data.frame(sample_xts), outcomename = "Close", order.by = index(sample_xts))

}

\preformatted{
From a data.frame, covert to a quantmod object directly.
}
}
