% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{balanceFocusedData}
\alias{balanceFocusedData}
\title{give more observations to the focused data}
\usage{
balanceFocusedData(xTs, SlicesFocusedData, SlicesAllData, TrainStart,
  TrainEnd)
}
\arguments{
\item{xTs}{xts object of training data}

\item{SlicesFocusedData}{list of index slice date vectors}

\item{SlicesAllData}{list of index slice date vectors}

\item{TrainStart}{absolute training start date}

\item{TrainEnd}{absolute training end date}
}
\description{
produce more focused data duplicate observations
such that focused data as at least the same number of observations
as "all observations - focused observations"
}
