% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary_predictive.R
\name{buildData}
\alias{buildData}
\title{Create one data object from multiple sources, applying transformations via standard R formula mechanism.}
\usage{
buildData(formula, na.rm = TRUE, return.class = "zoo",
  source.envir = NULL, ...)
}
\arguments{
\item{formula}{as quantmod buildData}

\item{na.rm}{as quantmod buildData}

\item{source.envir}{find xts Symbols in this environment}
}
\value{
"return.class" as quantmod buildData
}
\description{
UNUSED
see quantmod buildData
and this function can also acquire symbols that are stored in an environment
}
