% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{cashLogReturns}
\alias{cashLogReturns}
\title{cash log returns (CASHlogrets)}
\usage{
cashLogReturns(xTs = NULL)
}
\arguments{
\item{xTs}{xts object (only takes the index)}
}
\value{
xts object with the same index as xTs
}
\description{
\preformatted{

}
}
