% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{currentSymbolAPCReturns}
\alias{currentSymbolAPCReturns}
\title{get the Symbol current APC returns}
\usage{
currentSymbolAPCReturns(Symbol = Symbol, src = src)
}
\value{
xts object
}
\description{
\preformatted{

}
}
