% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{customSorting}
\alias{customSorting}
\title{custom sort a vector}
\usage{
customSorting(Vector, InitOrder, CI = FALSE)
}
\arguments{
\item{Vector}{vector to be sorted}

\item{InitOrder}{starting custom sorting ( without the excess )}

\item{CI}{case insensitive}
}
\description{
excess is appended to the end ( CI sort or CS sort )
}
\details{
# Custom Sorting in R
# 2014
# https://stackoverflow.com/questions/23995285/custom-sorting-in-r
# AND
# Case insensitive sort of vector of string in R
# 2015
# https://stackoverflow.com/questions/29890303/case-insensitive-sort-of-vector-of-string-in-r
}
\examples{
\dontrun{
customSorting( c("a","v", "E2", "c","l", "e3" ,"h","o","date"), InitOrder = c("date", "o", "h", "l", "c", "v", "a"), CI = TRUE )
[1] "date" "o"    "h"    "l"    "c"    "v"    "a"    "E2"   "e3"
}
}
