% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{dfToCREATETable}
\alias{dfToCREATETable}
\title{from column names and datatypes, make a CREATE TABLE statement}
\usage{
dfToCREATETable(df, con, Symbol, schname)
}
\arguments{
\item{df}{data.frame (with column names)}

\item{con}{DBI database connection}
}
\description{
also register its meta-data
}
