% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{diffXts}
\alias{diffXts}
\title{lag and/or difference and/or use a function(Fun) upon and xts object}
\usage{
diffXts(x, lag = 1, differences = 1, arithmetic = TRUE,
  log = FALSE, na.pad = TRUE, Fun = diff, ...)
}
\arguments{
\item{lag}{as diff.xts}

\item{differences}{as diff.xts}

\item{arithmetic}{as diff.xts}

\item{log}{as diff.xts}

\item{na.pad}{as diff.xts}

\item{Fun}{differencing function.  Meant to change the xTs in some way. (Default diff[.xts])}

\item{...}{dots passed}
}
\description{
lag and/or difference and/or use a function(Fun) upon and xts object
}
