% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{existSymbols.cache}
\alias{existSymbols.cache}
\title{xts object symbols}
\usage{
existSymbols.cache(Symbols = NULL, cache.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{...}{passed unused}

\item{source.envir}{location of xts objects}
}
\value{
R named list of Symbols and results (TRUE/FALSE)
}
\description{
\preformatted{

}
}
