% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{existSymbols}
\alias{existSymbols}
\title{xts object symbols}
\usage{
existSymbols(Symbols = NULL, source.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{source.envir}{location of xts objects}

\item{...}{passed to src}

\item{src}{source program to look for symbols}
}
\value{
R named list of Symbols and results (TRUE/FALSE)
}
\description{
looks in source.envir xor "src" exclusively
}
