% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{fancifyXts}
\alias{fancifyXts}
\title{number of days since Last Published}
\usage{
fancifyXts(x, AsOfToday = NULL, calcMonthlies = NULL,
  Last_Updated = NULL, CalendarAdj = NULL,
  FixLastLastTrueUpDated = NULL, NaCol = NULL, NaLOCF = NULL,
  Frequency = NULL, mkAsOfToday2LastObs = NULL)
}
\arguments{
\item{x}{xts object returned by getSymbols.FRED2 or
an xts object with an attribute of "Last_Updated" and "Frequency".
Or neither ( These values can be entered Manually).}

\item{AsOfToday}{Sys.Date() (default).
Otherwise overrides; POSIXct, Date or character("YYYY-MM-DD")
Date of refrence. Reality: today or date of run of the "modelling" program.
Note, choosing any earlier date does remove generated "Last Updated Dates"
However, choosing a later date does not add "Lst Updated Dates"}

\item{calcMonthlies}{TRUE(default) convert x (xts) object to an
end of month (eom) monthly (using To.Monthly) before
determining  the 'true last updated' dates.
Otherwise, FALSE; means 'x is already an eom monthly so do not apply To.Monthly.}

\item{Last_Updated}{xts attribute "Last_Updated"(default).
Otherwise overrides; POSIXct, Date or character("YYYY-MM-DD")}

\item{CalendarAdj}{"UnitedStates/GovernmentBond"(default).
To prevent 'last adjusted dates' from landing on an
unreasonable weekend, adjust the day forward to the next
US Government working day (e.g St. Louis FRED data).
Otherwise, "any other RQuantLib calender", e.g. "UnitedStates/NYSE".
The choice depends on the past schedule of the data. "NONE' meaning
no adjustment. See "?? RQuantLib::Calendars".}

\item{FixLastLastTrueUpDated}{TRUE(default). Of the current
month (determined by AsOfToday), the program calculates the
last day of the month then determines the its "Last Updated Date"
This 'date' may need custom adjustment.  For the hopefully one
Monthies date greater than today, adjust the last updated date,
to be calcualted using AsOfToday instead of the the last day of
the month. This is meant to corrrect the accuracy of the last
observation. This is meant to keep the usefulness of the
quantmod workflow of having the last observation be available to
be used to make a prediction.  Otherwise FALSE; useful
when no no new data is expected to be added between
now(AsOfDate) and month's end.}

\item{NaCol}{FALSE(Default). In for example, "Quarterly" data,
To.Monthy, will create gaps (observation rows with no data).
Create a new column and fill in the gaps with twos(2) and
the non-gaps with ones(1).}

\item{NaLOCF}{TRUE(Default). In for examle, "Quarterly" data,
To.Monthy, will create gaps (observation rows with no data).
Fill in the empty observations with the last(past.known data).}

\item{Frequency}{xts attribute "Frequency"(default) of Last_Updated.
Otherwise; character overrides.
Values can be "Quarterly". Currently, "Monthly", "Weekly", and "Daily".
Other Frequencies are not yet implmented.}

\item{mkAsOfToday2LastObs}{FALSE(default) Sets the hopefully the last
observation's xts index value to be today's date(AsOfToday)
instead of the end of month date. Generally,
mkAsOfToday2LastObs == TRUE is more trouble that it is worth.}
}
\value{
xts object with a new column of show the delay since last update.
}
\description{
The time between the "Last Updated Date" and the "index" date 'published date'
can be important in "model" building.
This amount of time can impact the quality of
the observations.  Therefore, the "model" building program needs
to know about the number of days since the "Last Updated Date."
}
\details{
First:
Of the data of Yahoo finance, the "Last Updated Date" and
the index date 'published date' are the same day. quantmod::getSymbols.yahoo
needs only to return the data and index.
The St. Louis FRED's "Last Updated Date" and index date 'published date'
are different days.

Second: Of the data of Yahoo finance, the Frequency
is "Daily".  The St. Louis FRED's "Frequency" can be"Quarterly",
"Monthly", "Weekly", or "Daily"

From a timeseries returned by getSymbols.FRED2.
The xts attribute "Last_Updated" is used.

Used is To.Monthly to convert the series to to a "Month-like" series.
This is typical in makeing long term predictions.

The xts is returned with an extra column added..

The author put some thought had been put into 'instead querying
ALFRED', but that path choice utimately was not taken.
}
\examples{
\dontrun{

# "Quarterly" time series on the St. Louis FRED
GDP <- getSymbols("GDP", src = "FRED2", auto.assign = FALSE)
GDP_DLY <- fancifyXts(GDP)

# "Monthly" time series on the St. Louis FRED
PAYEMS <- getSymbols("PAYEMS", src = "FRED2", auto.assign = FALSE)
PAYEMS_DLY <- fancifyXts(PAYEMS)

# "Weekly" time series on the St. Louis FRED
TREAST <- getSymbols("TREAST", src = "FRED2", auto.assign = FALSE)
TREAST_DLY <- fancifyXts(TREAST)

# "Daily" time series on the St. Louis FRED
DGS3MO <- getSymbols("DGS3MO", src = "FRED2", auto.assign = FALSE)
DGS3MO_DLY <- fancifyXts(DGS3MO)

}
}
