% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{getModelData}
\alias{getModelData}
\title{quantmod getModelData with parameter source.envir}
\usage{
getModelData(x, na.rm = TRUE, source.envir = NULL, ...)
}
\description{
get some Symbols from an environment (source.envir)
  will search first in (source.envir)
  if the Symbol is not found in the enviroment (source.envir),
  then get the Symbol from elsewhere
...
passed to getSymbols
        useful: from, to
  maybe useful: src
  maybe useful: set per Symbol src using setSymbolLookup
}
\details{
specifyModel(getModelData)
if na.rm == TRUE, then does 'na.exclude',
  BUT 'without rules' PUTS back (rbind) the last observation
}
