% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getSymbols.AAII}
\alias{getSymbols.AAII}
\title{American Association of Individual Investors (AAII) weekly sentiment survey}
\usage{
getSymbols.AAII(Symbols, env, return.class = "xts", ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded
Possible Symbols are the following:
"AAIIsentiment" (means get all of the columns);
otherwise, get specific columns;
"Bullish", "Neutral", "Bearish",
"Bullish8WMA", "BullBearSpread",
"BullishAvg", "BullishAvgPStd", "BullishAvgNStd",
"SP500WHigh", "SP500WLow", and "SP500WClose"}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{desired class of returned object.
Can be xts, zoo, data.frame, or xts (default)}

\item{...}{additional parameters}

\item{force}{FALSE(default) re-download data from AAII ( See the examples. )
The hidden variables ".AAIIsentiment_path2file" and
".AAIISentimentSurveyPastResults_path2file" located in the
environment of parameter env are the last know location of the "xls" file
and the "html" file.  Generally, using this parameter "force = TRUE"
is NOT necessary: the hidden variables are persistent through the
entire R session.  Also, send parameter "verbose = TRUE" to see
the *path2file locations.}
}
\value{
A call to getSymbols.AAII will load into the specified
environment one object for each \code{Symbol} specified,
with class defined by \code{return.class}.
}
\description{
Data represents what direction members feel the
stock market will be in next 6 months.
}
\details{
The sentiment survey measures the percentage of individual investors who are
bullish, bearish, and neutral on the stock market short term;
individuals are polled from the AAII Web site on a weekly basis.
Only one vote per member is accepted in each weekly voting period.

The latest published date/data is every Thursday.
The delivery time is during Friday (UNVERIFIED)
}
\examples{
\dontrun{

# common usage
if(!exists("BullBearSpread")) getSymbols("BullBearSpread", src = "AAII")

# During Fridays, if the user has a long R session that starts in the morning and
# continues through the evening, the user may want to get a new 'updated'
# Excel 'xls' file with the new results of this weeks vote.
# The Symbol name does not matter.  All symbols are from the same 'xls' file
# The dates are as of Thursday each week
getSymbols(c("Bullish", "Neutral", "Bearish"), src = "AAII")

# force requery of data from the AAII site
# will collect one new xls (historical obs) file and one new html (latest obs) file
getSymbols(c("Bullish", "Neutral", "Bearish"), src = "AAII", force = TRUE)

# all columns in one xts object
getSymbols("AAIIsentiment", src = "AAII")

}
}
\references{
\cite{AAII Investor Sentiment Survey \url{https://www.aaii.com/sentimentsurvey}}

\cite{Sentiment Survey Past Results \url{https://www.aaii.com/sentimentsurvey/sent_results}}
}
\seealso{
\code{\link{getSymbols}}
\code{\link{setSymbolLookup}}
}
\author{
Jeffrey A. Ryan

Andre Mikulec (adapted original code to work with AAII sentiment data)
}
\keyword{data}
