% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getSymbols.cache}
\alias{getSymbols.cache}
\title{Downloads Symbols to specified env from a local R environment}
\usage{
getSymbols.cache(Symbols = NULL, source.envir = NULL, env,
  return.class = "xts", ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded}

\item{source.envir}{where to find xts objects ( location of cache )}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{class of returned object}

\item{...}{additional parameters}
}
\value{
A call to getSymbols.csv will load into the specified environment one object for each Symbol specified, with class defined by return.class. Presently this may be ts, zoo, xts, data.frame, or timeSeries
}
\description{
Downloads Symbols to specified env from a local R environment
}
\examples{
\dontrun{
msft <- getSymbols("MSFT", src = "yahoo", auto.assign = FALSE)
source.envir = list2env(list(MSFT = msft))
saveSymbols(trg = "cache", source.envir = source.envir)
res <- getSymbols(Symbols = "MSFT", src = "cache", auto.assign = F)
}
}
