% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getSymbols.cache}
\alias{getSymbols.cache}
\title{Downloads Symbols to specified env from a local R environment}
\usage{
getSymbols.cache(Symbols, env, return.class = "xts",
  source.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{class of returned object}

\item{source.envir}{where to find xts objects ( location of cache )}

\item{...}{additional parameters}
}
\value{
A call to getSymbols.csv will load into the specified environment one object for each Symbol specified, with class defined by return.class. Presently this may be ts, zoo, xts, data.frame, or timeSeries
}
\description{
Downloads Symbols to specified env from a local R environment
}
