% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{getSymbols}
\alias{getSymbols}
\title{quantmod getSymbols with source.envir}
\usage{
getSymbols(Symbols = NULL, env = parent.frame(),
  reload.Symbols = FALSE, verbose = FALSE, warnings = TRUE,
  src = "yahoo", symbol.lookup = TRUE,
  auto.assign = getOption("getSymbols.auto.assign", TRUE),
  source.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded}

\item{env}{where to create objects. Setting env=NULL is equal to auto.assign=FALSE}

\item{reload.Symbols}{boolean to reload current symbols in specified environment. (FALSE)}

\item{verbose}{boolean to turn on status of retrieval. (FALSE)}

\item{warnings}{boolean to turn on warnings. (TRUE)}

\item{src}{character string specifying sourcing method. (yahoo)}

\item{symbol.lookup}{retrieve symbol's sourcing method from external lookup (TRUE)}

\item{auto.assign}{should results be loaded to env If FALSE, return results instead.
As of 0.4-0, this is the same as setting env=NULL. Defaults to TRUE}

\item{source.envir}{override: source environment to aquire the Symbols}

\item{...}{pass other parameters

get some Symbols from an environment (source.envir)
  will search first in (source.envir)
  if the Symbol is not found in the enviroment (source.envir),
  then get the Symbol from elsewhere
NOTE: do not do: "source.envir = e, "env = e"
      when auto.assign = TRUE(default), .GetSymbols is placed where "env = e"

e <- new.env()
NOT the same as "env = e"
assign("AAPL", getSymbols(list(AAPL = "yahoo"), auto.assign = F), envir = e)
ls.str(e)
AAPL : An 'xts' object on 2007-01-03/2018-04-24 containing:
  Data: num [1:2847, 1:6] 12.3 12 12.3 12.3 12.3 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:6] "AAPL.Open" "AAPL.High" "AAPL.Low" "AAPL.Close" ...
  Indexed by objects of class: [Date] TZ: UTC
  xts Attributes:
List of 2
 $ src    : chr "yahoo"
 $ updated: POSIXct[1:1], format: "2018-04-25 18:12:26"

getSymbols(list(AAPL = "yahoo", MSFT = "yahoo"), source.envir = e)
AAPLSymbol <- getSymbols(list(AAPL = "yahoo"), auto.assign = FALSE, source.envir = e)}
}
\description{
quantmod getSymbols with source.envir
}
