% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getSymbols}
\alias{getSymbols}
\title{quantmod getSymbols with source.envir}
\usage{
getSymbols(Symbols = NULL, env = parent.frame(),
  reload.Symbols = FALSE, verbose = FALSE, warnings = TRUE,
  src = "yahoo", symbol.lookup = TRUE,
  auto.assign = getOption("getSymbols.auto.assign", TRUE),
  source.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded}

\item{env}{where to create objects. Setting env=NULL is equal to auto.assign=FALSE}

\item{reload.Symbols}{boolean to reload current symbols in specified environment. (FALSE)}

\item{verbose}{boolean to turn on status of retrieval. (FALSE)}

\item{warnings}{boolean to turn on warnings. (TRUE)}

\item{src}{character string specifying sourcing method. (yahoo)}

\item{symbol.lookup}{retrieve symbol's sourcing method from external lookup (TRUE)}

\item{auto.assign}{should results be loaded to env If FALSE, return results instead.
As of 0.4-0, this is the same as setting env=NULL. Defaults to TRUE}

\item{source.envir}{override: source environment to aquire the Symbols}

\item{...}{pass other parameters}
}
\description{
Get some Symbols from an environment (source.envir)
It will search first in (source.envir)
If the Symbol is not found in the enviroment (source.envir),
then get the Symbol from elsewhere.
}
\details{
NOTE: do not do: "source.envir = e, "env = e"
when auto.assign = TRUE(default), .GetSymbols is placed where "env = e"
}
\examples{
\dontrun{

e <- new.env(parent = emptyenv())

getSymbols(list(AAPL = "yahoo"), env = e) # bcomes INTERNALLY auto.assign = T
ls.str(envir = e)

# INSTEAD OF "yahoo", get IT from the source.envir
AAPLSymbol <- getSymbols(list(AAPL = "yahoo"), auto.assign = FALSE, source.envir = e)
str(AAPLSymbol)

}
}
