% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getSymbols.Simfa}
\alias{getSymbols.Simfa}
\title{Simfa bond and equity information}
\usage{
getSymbols.Simfa(Symbols, env, return.class = "xts", ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded
Possible Symbols are the following:
"SimfaUSCorporate" (means get all of the columns);
otherwise, get specific columns;
BondsAveMaturity"}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{desired class of returned object.
Can be xts, zoo, data.frame, or xts (default)}

\item{...}{additional parameters}

\item{force}{FALSE(default) re-download data from Simfa ( See the examples. )
The hidden variable ".SimfaBondMarketData_path2file" located in the
environment of parameter env is the last known location of the "xls" file.
Generally, using this parameter "force = TRUE"
is NOT necessary: the hidden variables are persistent through the
entire R session.  Also, send parameter "verbose = TRUE" to see
the *path2file location.}
}
\value{
A call to getSymbols.Simfa will load into the specified
environment one object for each \code{Symbol} specified,
with class defined by \code{return.class}.
}
\description{
# 1 (HAVE)
most valuable information
tab "Issuance" columns "Corporate Debt", "Total"
US Bond Market Issuance and Outstanding
https://www.sifma.org/wp-content/uploads/2017/06/cm-us-bond-market-sifma.xls
https://www.sifma.org/resources/research/us-bond-market-issuance-and-outstanding/
https://www.sifma.org/resources/research/bond-chart/
upper right hand corner, click on "download xls"
click on "explore"
SIFMA
https://www.sifma.org/
}
\details{
same
Click on \[Explore\]
US Bond Market Issuance and Outstanding
  (Upper Right Corner)
  DOWNLOAD: JPG|CSV|XLS ( automation: need phantom js ) \*\*HERE\*\*
  https://www.sifma.org/wp-content/uploads/2017/06/cm-us-bond-market-sifma.xls
    XLS: "Issuance" tab ( has data through the previous month)
# XLS bottom ( SEE DEC 31 2018: EXACT SAME AS THE ARTICLE )
         Year	M	 Municipal 	 Treasury 	 Mortgage-Related 	 Corporate Debt 	 Federal Agency Securities 	 Asset-Backed 	 Total
YTD '17
YTD '18
% Change

# 2 (HAVE)
most valuable information
tab "Average Maturity" column "Years" (Symbol is "BondsAveMaturity")
https://www.sifma.org/wp-content/uploads/2017/08/Corporate-US-Corporate-Issuance-SIFMA.xls
US Corporate Bond Issuance
https://www.sifma.org/resources/research/us-corporate-bond-issuance/
Securities Industry and Financial Markets Association
SIFMA
https://www.sifma.org/
#
Note only the monthly dates for the current year and the previous
year are available.  The remaining data is year-end-data.
The dates intermediate dates are filled in using
xts::to.monthly() and seq()
The monthly dates are interpolated using zoo::na.approx
that is stats::approx. (In R Code that is "approx1").

# 3
most valuable information
tab "US Bond Market" column "Corporate Debt"
US Bond Market Trading Volume
https://www.sifma.org/wp-content/uploads/2017/06/cm-us-bond-market-trading-volume-sifma.xls
https://www.sifma.org/resources/research/us-bond-market-trading-volume/

# 4
US Equity Issuance and Trading Volumes
most valuable information
tab "Underwriting" column "TOTAL EQUITY"
https://www.sifma.org/wp-content/uploads/2017/06/cm-us-equity-sifma1.xls
https://www.sifma.org/resources/research/us-equity-stats/

main pages ( html list of all excel spreadsheets )
https://www.sifma.org/resources/archive/research/statistics/
# a.k.a.
https://www.sifma.org/resources/archive/research/statistics/?aq=&hPP=10&idx=prod_wp_searchable_posts&ap=0&fR%5Btaxonomies.research_type%5D%5B0%5D=Statistics&is_v=1
next two pages
https://www.sifma.org/resources/archive/research/statistics/?aq=&hPP=10&idx=prod_wp_searchable_posts&ap=1&fR%5Btaxonomies.research_type%5D%5B0%5D=Statistics&is_v=1
https://www.sifma.org/resources/archive/research/statistics/?aq=&hPP=10&idx=prod_wp_searchable_posts&ap=2&fR%5Btaxonomies.research_type%5D%5B0%5D=Statistics&is_v=1

referred by Jeff Cox
AND
Christopher Whalen
twitter handle: rcwhalen

When the Bid Goes to Zero |
https://goo.gl/2tUeJp  |
https://twitter.com/rcwhalen/status/1072619534819971072
AND
When the Bid Goes to Zero
December 11, 2018
By: R. Christopher Whalen
https://www.theinstitutionalriskanalyst.com/single-post/2018/12/11/When-the-Bid-Goes-to-Zero
}
\examples{
\dontrun{

# common usage
if(!exists("BondsAveMaturity")) getSymbols("BondsAveMaturity", src = "Simfa")

getSymbols(c("CorporateDebt","BondsAveMaturity", "CorporateDebtTradeVol"), src = "Simfa")

# force requery of data from the Simfa site
# will collect one new xls (historical obs) file
getSymbols(c("CorporateDebt","BondsAveMaturity", "CorporateDebtTradeVol"), src = "Simfa", force = TRUE)

# all columns in one xts object
getSymbols("SimfaUSCorporate", src = "Simfa")

}
}
\references{
\cite{US Corporate Bond Issuance \url{https://www.sifma.org/resources/research/us-corporate-bond-issuance/}}

\cite{R. Christopher Whalen - When the Bid Goes to Zero \url{https://www.theinstitutionalriskanalyst.com/single-post/2018/12/11/When-the-Bid-Goes-to-Zero}}
}
\seealso{
\code{\link{getSymbols}}
\code{\link{setSymbolLookup}}
}
\author{
Jeffrey A. Ryan

Andre Mikulec (adapted original code to work with Simfa Bond market data)
}
\keyword{data}
