% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getSymbols.USFedPhil}
\alias{getSymbols.USFedPhil}
\title{Survey of Professional Forecasters data from the Philadelphia FED}
\usage{
getSymbols.USFedPhil(Symbols, env, return.class = "xts",
  SurveyDecades = "All", DataPath = "./USFedPhilForecastersData",
  MaxAge = NULL, UseFST = TRUE, Fun = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded
To get all of the data use: "USFedPhilForecastingData".  Possible individual
Symbols are listed in the links on the Forcasters' main page
that reads (see below) ... (1) Historical Data Files for the Survey
of Professional Forecasters (2) Documentation: PDF Provides information on
all variables, transformations, and files in the survey.
E.g. "UNEMP2". Symbols will be returned
as (xts) data with names as excelcolumn+Fun"
E.g. "UNEMP2.mean"}

\item{env}{where to create objects. (.GlobalEnv)}

\item{return.class}{desired class of returned object.
Can be xts, zoo, data.frame, or xts (default)}

\item{SurveyDecades}{"All"(default). Vector of options include the following
"1970s", "1980s", "1990s", "2000s", and "2010s".
Multiple options can be chosen and  sent in a character
vector. e.g. c("2000s", "2010s")}

\item{DataPath}{"./USFedPhilForecastersData"(default). New/old location of the
new/old downloaded xls files from the Philadelphia Fed's site. Windows
users can also enter this path using instead two back slashes.
Abolute paths can also be used.}

\item{MaxAge}{"4 hours"(default) is longest age allowed, such that the retrieving the
the MOST RECENTLY PUBLISHED FORECASTERS EXCEL FILE from  the local fileystem
will be allowed to be done. If the MaxAge is exceeded then,
the most recent excel file is refreshed anew from the Philadelphia Fed site.
The format uses as.difftime: "# secs", "# mins", "# hours", "# days", "# weeks"}

\item{UseFST}{TRUE(default) use the "Excel + FST" files (from R package fst).
Initial creation of .fst files uses about 15 seconds of time per file (5 file).
Afterwards, queries are lightning fast.
Otherwiase FALSE, use "Excel Only"  Queries are slower (mostly because of the heavier disk I/O)}

\item{Fun}{mean(default) with na.rm = TRUE(default) Aggregate function to
apply to the period data. This becomes the symbol Suffix. E.g. "mean" will
produce the suffix "mean".}

\item{...}{additional parameters pased to Fun. Also dots are passed to
quantmod___try.download.file}

\item{force}{FALSE(default) re-download data from USFedPhil ( See the examples. )
Generally, using this parameter "force = TRUE"
is NOT necessary: after MaxAge has been exceeded and if a query needs the
MOST RECENTLY PUBLISHED FORECASTERS EXCEL FILE then this file will be downloaded anew.}
}
\value{
A call to getSymbols.USFedPhil will load into the specified
environment one object for each \code{Symbol} specified,
with class defined by \code{return.class}.
}
\description{
About the data . . .
}
\details{
File Structure: Column Header Nomenclature and Forecast Horizons

“1” to “6” (quarterly forecasts) or
“A” and “B” (annual-average forecasts)

The number “1” represents the “forecast” for the quarter prior to the quarter in
which the survey is conducted. The forecasters know the values of the variables for this
quarter at the time they submit their projections. . . . The forecasters are permitted
to forecast a revision

The number “2” represents the forecast for the current quarter, defined as
the quarter in which the survey is conducted. The numbers “3” through “6”
represent the forecasts for the four quarters after the current quarter.
The letters “A” and “B” represent annual average forecasts
for the current year (the year in which the survey is conducted) and
the following year.

About the date/time index . . .

Deadline and Release Dates for the Survey of Professional Forecasters

release dates
https://www.philadelphiafed.org/-/media/research-and-data/real-time-center/survey-of-professional-forecasters/spf-release-dates.txt?la=en

True deadline and news release dates for surveys prior to 1990:Q2 are not known.

NOTE: about "not known" (see below), for simplicity
xts time index of these "not known" periods is the middle date of the calendar quarter

*The 1990Q2 survey was not taken in real time, because the Philadelphia Fed
had not yet taken over the survey. Forecasters were asked to provide dated
forecasts from May 1990.

**The 1996Q1 survey was delayed because of the federal government shutdown,
which in turn delayed the release of government statistical data.

***The 2013Q4 survey was delayed because of the federal government shutdown,
which in turn delayed the release of government statistical data.

# Popular start page
https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/data-files

# all of the data (1968 to the present) in five(5) excel files
Individual Forecasts for the Survey of Professional Forecasters
https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/historical-data/individual-forecasts

# Some, compare and constrasts may be
'forcasters'(prediction) w.s. FRED (acutally what happened)

# unemployment

Civilian Unemployment Rate (UNEMP)
https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/data-files/unemp

Civilian Unemployment Rate (UNRATE)
https://fred.stlouisfed.org/series/UNRATE/

# prices ( e.g. inflation )

Price Index for Gross National Product/Gross Domestic Product (PGDP)
https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/data-files/pgdp

Gross National Product (chain-type price index) (A001RV1Q225SBEA)
https://fred.stlouisfed.org/series/A001RV1Q225SBEA

# gross domestic product

Nominal Gross National Product/Gross Domestic Product (NGDP)
https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/data-files/ngdp

Gross Domestic Product (GDP)
https://fred.stlouisfed.org/series/GDP

# Corporate Profits

Corporate Profits After Tax (CPROF)
https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/data-files/cprof

Corporate Profits After Tax (without IVA and CCAdj) (CP)
https://fred.stlouisfed.org/series/CP
#'
}
\examples{
\dontrun{

# common usage
if(!exists("UNEMP2_mean")) getSymbols("UNEMP2", src = "USFedPhil")

getSymbols(c("UNEMP2", "PGDP2", "NGDP2","CPROF2"), src = "USFedPhil")
getSymbols(c("RR1_TBILL_PCE_3","RR3_TBILL_PCE_3"), src = "USFedPhil", Fun = mean, na.rm = TRUE)

# standard deviation
getSymbols("UNEMP2", src = "USFedPhil", Fun = sd, na.rm = TRUE)

# just recent data
getSymbols("UNEMP2", src = "USFedPhil", SurveyDecades = "2010s")

# force requery of data from the Philladelphia site
# will collect new xls file(s)
getSymbols("UNEMP2", src = "USFedPhil", force = TRUE)

# all columns in one VERY WIDE xts object
getSymbols("USFedPhilForecastingData", src = "USFedPhil")

}
}
\references{
\cite{Survey of Professinal Forecasters Release Dates \url{https://www.philadelphiafed.org/-/media/research-and-data/real-time-center/survey-of-professional-forecasters/spf-release-dates.txt?la=en}}

\cite{Historical Data Files for the Survey of Professional Forecasters \url{https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters/data-files}}

\cite{Documentation: PDF Provides information on all variables, transformations, and files in the survey. \url{https://www.philadelphiafed.org/-/media/research-and-data/real-time-center/survey-of-professional-forecasters/spf-documentation.pdf?la=en}}

\cite{Survey of Professional Forecasters \url{https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters}}

\cite{Federal Researve Bank of Philadelphia \url{https://www.philadelphiafed.org/}}
}
\seealso{
\code{\link{getSymbols}}
\code{\link{setSymbolLookup}}
}
\author{
Jeffrey A. Ryan

Andre Mikulec (adapted original code to work with the
US Federal Researve Bank of Philadelphia's Survey of Professional Forecasters)
}
\keyword{data}
