% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{indexSlices}
\alias{indexSlices}
\title{Rob Hyndman style time slices}
\usage{
indexSlices(xTs, SlicesAllData, TrainStart, TrainEnd)
}
\arguments{
\item{xTs}{data}

\item{SlicesAllData}{list of index slice date vectors}

\item{TrainStart}{absolute training start date}

\item{TrainEnd}{absolute training end date}
}
\value{
list of indexSlicesObs and indexSlicesOutObs position indexes
meant to be passed to caret::trainControl index and indexOut
}
\description{
\preformatted{

Create time slice indexes to be passed to caret trainControl index
and indexOut. My visual observation is that this model
DOES \*worse\* than general cross validation.
This is because, per time slice, less observations
exist to TRAIN/TEST over.

}
}
\examples{
\dontrun{
# determine slices of index and indexOut
indexSlices(
    # data.window = c(TrainingBegin, TrainingEnd)
    # Update currently specified or built model with most recent data
    # remove the last record(NO) (na.rm = FALSE)
    # "2007-01-31" (actual "2001-12-31")
    xTs = modelData(specifiedUnrateModel, exclude.training = TRUE) # not defined yet
  , SlicesAllData = NBERAllData
  , TrainStart = TrainingBegin
  , TrainEnd   = TrainingEnd
) -> Sliced
# determine slices of index and indexOut to caret trainControl
# pass through
indexSlicesObs    <- Sliced[["indexSlicesObs"]]
indexSlicesOutObs <- Sliced[["indexSlicesOutObs"]]
}
}
