% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{initDate}
\alias{initDate}
\title{garantee a date is a date or and empty date}
\usage{
initDate(date = NULL)
}
\arguments{
\item{date}{date}
}
\value{
date
}
\description{
garantee a date is a date or and empty date
}
\examples{
\dontrun{
# > initDate(date = NULL)
# Date of length 0

# > initDate(c())
# Date of length 0
}
}
