% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{initEnv}
\alias{initEnv}
\title{sets the enviroment}
\usage{
initEnv(init = NULL, envir = parent.frame())
}
\arguments{
\item{init}{list with names entries of alernate/other options}

\item{envir}{where to return the options}
}
\value{
environment is set
}
\description{
space-saver - meant to be used at the beginning of a function
}
\examples{
\dontrun{
# > options(max.print=88888L)
# initEnv()
# > getOption("max.print")
# [1] 99999
}
}
