% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{initPorfVal}
\alias{initPorfVal}
\title{initialize}
\usage{
initPorfVal(initVal = NULL)
}
\arguments{
\item{initVal}{portfolio starting value($$). If it is null
then the value is set to by 100000}
}
\value{
set value
}
\description{
\preformatted{

}
}
\examples{
\dontrun{
# > initPorfVal(initVal = NULL)
# [1] 100000
}
}
