% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary_predictive.R
\name{quantmod___is.method.available}
\alias{quantmod___is.method.available}
\alias{is.method.available}
\title{check the .libPaths() to see a package has a training function}
\usage{
quantmod___is.method.available(method, package)

is.method.available(method, package)
}
\description{
extracted because quantmod does not export this
}
