% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{leadingWilshire5000LogReturns}
\alias{leadingWilshire5000LogReturns}
\title{get the Wilshare 5000 Index leadinglog returns}
\usage{
leadingWilshire5000LogReturns()
}
\value{
xts object
}
\description{
\preformatted{

}
}
