% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{liquifyDF}
\alias{liquifyDF}
\title{flatten a data.frame to (hopfully) ONE row}
\usage{
liquifyDF(x, ConstColsRegex = "^dateindex", FactorColsRegex = "Fact$",
  FactorColsNAReplace = NULL, FactorColsFixedSep = ".",
  DetailColsFixedSep = ".", SpaceFixedSep = ".",
  AmperstandFixedSep = "And")
}
\description{
NOTE: only WORKS with ONE unique primary key value
If I want to work with many primary key valuess,
then one would need to split the data upon those values
then next apply this function individually to each chunk
Last, 'smart' merge all of the data.frames together (while
adding new columns contributed by each data.frame).
}
\examples{
\dontrun{
#
# data.frame(
#   dateindexid = c(17000, 17000, 17000, 17000),
#   dateindexFact = c("2016-07-18", "2016-07-18", "2016-07-18", "2016-07-18"),
#   ActionFact = c("Morn", "Morn", "Night", "Night"),
#   ActionAtFact = c("Sell", "Buy", "Sell", "Buy"),
#   INSTR1 = c(8, 16, 32, 64),
#   INSTR2 = c(10008, 10016, 10032, 10064)
# , stringsAsFactors = FALSE
# )  -> DFS
#
# > DFS
#   dateindexid dateindexFact ActionFact ActionAtFact INSTR1 INSTR2
# 1       17000    2016-07-18       Morn         Sell      8  10008
# 2       17000    2016-07-18       Morn          Buy     16  10016
# 3       17000    2016-07-18      Night         Sell     32  10032
# 4       17000    2016-07-18      Night          Buy     64  10064
#
#
# > liquifyDF(DFS)
#   dateindexid dateindexFact Morn.Sell.INSTR1 Morn.Buy.INSTR1 Night.Sell.INSTR1 Night.Buy.INSTR1
# 1       17000    2016-07-18                8              16                32               64
#   Morn.Sell.INSTR2 Morn.Buy.INSTR2 Night.Sell.INSTR2 Night.Buy.INSTR2
# 1            10008           10016             10032            10064
#
#
}
}
