% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{listSymbols.cache}
\alias{listSymbols.cache}
\title{xts object symbols in the cache}
\usage{
listSymbols.cache(cache.envir = NULL, ...)
}
\arguments{
\item{...}{passed unused}

\item{source.envir}{location of xts objects}
}
\value{
character vector of Symbols
}
\description{
\preformatted{

}
}
