% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{nextMonthfromYesterday}
\alias{nextMonthfromYesterday}
\title{change the index date to the future}
\usage{
nextMonthfromYesterday(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
xts object with the index moved up
}
\description{
change the index date to the future
}
