% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgCurrentUser}
\alias{pgCurrentUser}
\alias{pgCurrentSchema}
\alias{pgCurrentDB}
\alias{pgCurrentSearchPath}
\alias{pgSetCurrentSearchPath}
\alias{pgCurrentTempSchema}
\title{get PostgreSQL current user name}
\usage{
pgCurrentUser(con)

pgCurrentSchema(con)

pgCurrentDB(con)

pgCurrentSearchPath(con)

pgSetCurrentSearchPath(con, path)

pgCurrentTempSchema(con)
}
\description{
get PostgreSQL current user name

get PostgreSQL current user schema

get PostgreSQL current user database name

get PostgreSQL current user search path

set PostgreSQL current user search path

get PostgreSQL current user temp schema
}
