% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgCurrentHostAddress}
\alias{pgCurrentHostAddress}
\alias{pgCurrentPort}
\alias{pgCurrentUser}
\alias{pgCurrentSchema}
\alias{pgCurrentDB}
\alias{pgCurrentSearchPath}
\alias{pgSetCurrentSearchPath}
\alias{pgCurrentTempSchema}
\title{get PostgreSQL host address}
\usage{
pgCurrentHostAddress(con)

pgCurrentPort(con)

pgCurrentUser(con)

pgCurrentSchema(con)

pgCurrentDB(con)

pgCurrentSearchPath(con)

pgSetCurrentSearchPath(con, path)

pgCurrentTempSchema(con)
}
\description{
\preformatted{

Find the host name and port using PSQL commands.
https://stackoverflow.com/questions/5598517/find-the-
host-name-and-port-using-psql-commands

}

\preformatted{

Find the host name and port using PSQL commands.
https://stackoverflow.com/questions/5598517/find-the-
host-name-and-port-using-psql-commands

}

\preformatted{

}

\preformatted{

}

\preformatted{

}

\preformatted{

}

\preformatted{

}

\preformatted{

}
}
