% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{pgInsert_STUB}
\alias{pgInsert_STUB}
\title{inserat into a database table 'new' data}
\usage{
pgInsert_STUB(...)
}
\description{
TODO [ ] reorder df columns to match DB Server table LOOK AT . . .
  add that code in here ( see caroline dbWriteTable2 ),
  or
  my implementation, my 'sort using factor code' in a tradeModel function )
}
