% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{predictModel}
\alias{predictModel}
\alias{predictModel.default}
\title{determine the future}
\usage{
predictModel(object, data, ...)

\method{predictModel}{default}(object, data, ...)
}
\description{
\preformatted{

}
}
