% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{printCalendar}
\alias{printCalendar}
\title{print calendar of month by mont returns}
\usage{
printCalendar(xTs = NULL, title = NULL, PortFolioRetCol = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{title}{heading}

\item{PortFolioRetCol}{"portfolio.returns"(default). Column of the 'portfolio returns.'}
}
\value{
xts object of arithmatic returns
}
\description{
\preformatted{

The vericle left side axis is in "years".
The horizontal axis is in "months".

}

Geometric adding is done across months
An end of year(12 month summary) is in the verticle right side axis
}
