% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{printTail}
\alias{printTail}
\title{show the last six(6) records}
\usage{
printTail(xTs = NULL, title = NULL, n = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{title}{heading}

\item{n}{number of rows to print}
}
\value{
invisible xts object
}
\description{
show the last six(6) records
}
