% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{rollApply2}
\alias{rollApply2}
\title{rollapply anything}
\usage{
rollApply2(x, fun, window = len(x), minimum = 1, align = "left", ...)
}
\arguments{
\item{fun}{the function to evaluate}

\item{window}{window width defining the size of the subset available
to the fun at any given point}

\item{minimum}{minimum width of the window.
Will not return results if the window is truncated below this value
at the end of the data set}

\item{align}{whether to align the window right or left}

\item{...}{additional arguments to pass to fun}

\item{data}{any R object}
}
\description{
\preformatted{

This is based on archived package rowr archived function rollApply (rollapply anything).
Uses Rows so single dimension objects are not dropped to vectors.

}
}
\examples{
\dontrun{
all examples of archived package/function rowr::rollApply work

rollApply2(state.x77, function(x) { browser() }, window = 3, min = 3, align = "right")
Called from:
Browse[1]> print(x)
        Population Income Illiteracy Life Exp Murder HS Grad Frost   Area
Alabama       3615   3624        2.1    69.05   15.1    41.3    20  50708
Alaska         365   6315        1.5    69.31   11.3    66.7   152 566432
Arizona       2212   4530        1.8    70.55    7.8    58.1    15 113417

States <- rollApply2(state.x77, function(x) { tail(x,1) }, window = 3, min = 3, align = "right")

}
}
