% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{rollApply2}
\alias{rollApply2}
\title{rollapply anything}
\usage{
rollApply2(x, fun, window = len(x), minimum = 1, align = "left",
  .parallel = FALSE, ...)
}
\arguments{
\item{x}{xts object}
}
\description{
based on rowr rollApply (rollapply anything)
}
\examples{
\dontrun{
all examples of rowr::rollApply work

rollApply2(state.x77, function(x) { browser() }, window = 3, min = 3, align = "right")
Called from:
Browse[1]> print(x)
        Population Income Illiteracy Life Exp Murder HS Grad Frost   Area
Alabama       3615   3624        2.1    69.05   15.1    41.3    20  50708
Alaska         365   6315        1.5    69.31   11.3    66.7   152 566432
Arizona       2212   4530        1.8    70.55    7.8    58.1    15 113417

# seems to be working ( notice: 8 second startup )
rollApply2(state.x77, function(x, a, b) { list(list(x=x), a=a, b=b)}, window = 3, min = 3, align = "right",
  a = 1, b = 2, .parallel = T)

}
}
