% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{rollEpochRanks}
\alias{rollEpochRanks}
\title{rolling percent ranks grouped by epoch}
\usage{
rollEpochRanks(
  xTs1,
  xTs2,
  window = 10,
  ranks = 4,
  incBeforeEpochLags = FALSE,
  originalData = FALSE,
  derivedDataDetailed = FALSE
)
}
\arguments{
\item{xTs1}{xts object}

\item{xTs2}{xts object of single column
(or multiple columns if doing interaction).
This is an xts object with a data flag of the groupings.
This is most likely calculated from this package's
days/months\*Since/After\*Event functions.}

\item{window}{size of zone of rank calculation}

\item{ranks}{number of ranks}

\item{incBeforeEpochLags}{FALSE(default), if TRUE include early data the size of "window - 1".}

\item{originalData}{FALSE(default), if TRUE include oringinal data in the output.}

\item{derivedDataDetailed}{FALSE(default) if TRUE, include the rank as part of the derived data column name.}
}
\value{
xts object
}
\description{
\preformatted{

The idea is group rolling ranks from the end of crash/recession
to the beginning of the next crash/recession (so that the
recession to recession social/economic behavior may be
recaptured.)

[Maybe] this is (weakly) designed to run through "eXplode."

}
}
\examples{
\dontrun{

# S&P 500
gspc <- To.Monthly(Cl(getSymbols("^GSPC", auto.assign = FALSE, from = "1900-01-01")), indexAt = "lastof", OHLC = F)
colnames(gspc) <- "GSPC"
# returns
gspc.ret <- PerformanceAnalytics::Return.calculate(gspc)

# top 20 worst monthly returns in the S&P500 history
CrashDates <- zoo::as.Date(rownames(head(as.matrix(gspc.ret)[order(as.matrix(gspc.ret)),1, drop = F], 20)))

mse <- monthsSinceEvent(gspc.ret, dates = CrashDates)
mse <- mse[ , c("GSPC","GSPC_TAE_MTHSNC", "GSPC_RUN_MTHSNC")]
# 20 (or near so) epochs
unique(mse$GSPC_RUN_MTHSNC)

# nothing (no crash date) before 1962
tail(mse["/1962"], 10)

# e.g. (will be used on function
# from crash to crash
# "RUN month(MTH) since(SNC) event"
smse <- split(mse[, c("GSPC", "GSPC_TAE_MTHSNC")], coredata(mse[, "GSPC_RUN_MTHSNC"]))

# 20 periods
length(smse)
[1] 20

# sample testing/demo data
## mse <- mse[114:118,]

# note USUALLY "incBeforeEpochLags == FALSE"
# to "fully" separate ONE epoch from the other.
# to get/give full separation before/after recession crashes
# incBeforeEpochLags = FALSE
rollEpochRanks(mse[, c("GSPC", "GSPC_TAE_MTHSNC")], xTs2 = mse[, "GSPC_RUN_MTHSNC"], incBeforeEpochLags = TRUE, originalData = TRUE, derivedDataDetailed = TRUE)

#            GSPC GSPC_TAE_MTHSNC GSPC.rollEpochRanks4 GSPC_TAE_MTHSNC.rollEpochRanks4
# 1950-01-31   NA              NA                   NA                              NA

}
}
