% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{runRanksDT}
\alias{runRanksDT}
\title{backward looking and forward looking Ranks using data.table}
\usage{
runRanksDT(
  x,
  ranks = 4,
  ties.method = "average",
  originalData = FALSE,
  derivedDataDetailed = FALSE
)
}
\arguments{
\item{x}{xts object}

\item{ranks}{4(default) number of ranks. A lower value
means a lower rank number.}

\item{ties.method}{"average"(default). Passed to data.table::frank parameter ties.method.}

\item{derivedDataDetailed}{FALSE(default) if TRUE, include the original data in the output.}
}
\value{
xts object
}
\description{
\preformatted{

This is the rank over the "entire" data series (the window).
In general this is not useful. This is NOT USED ANYWHERE.
This uses package data.table

}
}
\examples{
\dontrun{

runRanksDT(xts(sample(4,4,T), zoo::as.Date(0:3)), originalData = TRUE, derivedDataDetailed = TRUE)

runRanksDT(xts(matrix(c(2,2,4,4),ncol=1), zoo::as.Date(0:3)), ties.method = "average", originalData = TRUE, derivedDataDetailed = TRUE)
runRanksDT(xts(matrix(c(2,2,4,4),ncol=1), zoo::as.Date(0:3)), ties.method = "first", originalData = TRUE, derivedDataDetailed = TRUE)
runRanksDT(xts(matrix(c(2,2,4,4),ncol=1), zoo::as.Date(0:3)), ties.method = "max", originalData = TRUE, derivedDataDetailed = TRUE)
runRanksDT(xts(matrix(c(2,2,4,4),ncol=1), zoo::as.Date(0:3)), ties.method = "min", originalData = TRUE, derivedDataDetailed = TRUE)
runRanksDT(xts(matrix(c(2,2,4,4),ncol=1), zoo::as.Date(0:3)), ties.method = "dense", originalData = TRUE, derivedDataDetailed = TRUE)

runRanksDT(xts(sample(10,10,T), zoo::as.Date(0:9)), originalData = TRUE, derivedDataDetailed = TRUE)

runRanksDT(xts(sample(10,10,T), zoo::as.Date(0:9)), ranks = 3, originalData = TRUE, derivedDataDetailed = TRUE)

runRanksDT(xts(sample(10,10,T), zoo::as.Date(0:9)), ranks = 2, originalData = TRUE, derivedDataDetailed = TRUE)

}
}
