% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{saveSymbols.cache}
\alias{saveSymbols.cache}
\title{save xts objects as .xts in source.envir ( cache )}
\usage{
saveSymbols.cache(Symbols = NULL, source.envir = NULL,
  cache.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{source.envir}{location of xts objects}

\item{...}{not used}

\item{target.envir}{location of where to store xts objects ( location of cache )}
}
\description{
save xts objects as .xts in source.envir ( cache )
}
\examples{
\dontrun{

msft <- getSymbols("MSFT", src = "yahoo", auto.assign = FALSE)
source.envir = list2env(list(MSFT = msft))
saveSymbols(trg = "cache", source.envir = source.envir)
saveSymbols(trg = "cache", source.envir = source.envir, cache.envir = .GlobalEnv)

}
}
