% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{saveSymbols.PostgreSQL}
\alias{saveSymbols.PostgreSQL}
\title{saves xts object symbols to a persistent location (database: PostgreSQL)}
\usage{
saveSymbols.PostgreSQL(Symbols = NULL, con = NULL,
  source.envir = NULL, field.names = c("Open", "High", "Low", "Close",
  "Volume", "Adjusted"), db.fields = c("o", "h", "l", "c", "v", "a"),
  user = NULL, password = NULL, dbname = NULL, schname = NULL,
  host = "localhost", port = 5432, options = NULL,
  forceISOdate = TRUE, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{field.names}{names existing in starting columns}

\item{db.fields}{character vector indicating
names of fields to insert}

\item{user}{username(default "Symbols") to access database}

\item{password}{password(default "Symbols") to access database}

\item{dbname}{database name (default "Symbols")}

\item{schname}{schema name (default "Symbols")}

\item{host}{database host (default "localhost")}

\item{port}{database port (default 5432)}

\item{options}{pass extra parameters in aa string to the command line}

\item{forceISOdate}{TRUE(default)/FALSE if the communication of date (time stamp) from PostgreSQL
is forced to ISO style at conection}

\item{...}{pass through parameters}

\item{sourc.envir}{location of xts objects}
}
\description{
# Once only, the end-user must pre-create the [database and [ user and]] schema
}
\details{
CREATE ROLE "Symbols" LOGIN
  NOSUPERUSER INHERIT NOCREATEDB NOCREATEROLE NOREPLICATION;

ALTER USER "Symbols" PASSWORD 'Symbols';

CREATE DATABASE "Symbols"
  WITH OWNER = "Symbols"
       ENCODING = 'UTF8'
       TABLESPACE = pg_default
       LC_COLLATE = 'C'
       LC_CTYPE = 'C'
       CONNECTION LIMIT = -1;

CREATE SCHEMA "Symbols"
  AUTHORIZATION "Symbols";

# # create automatically by function: dfToCREATETable
# CREATE TABLE "Symbols"
# (
#   "Symbols" text NOT NULL,
#   updated timestamp with time zone,
#   "index_R_class" text,
#   src text,
#   CONSTRAINT "Symbols_pkey" PRIMARY KEY ("Symbols")
# )
# WITH (
#   OIDS=FALSE
# );
# ALTER TABLE "Symbols"
#   OWNER TO "Symbols";
}
\examples{
\dontrun{

msft <- getSymbols("MSFT", src = "yahoo", auto.assign = FALSE)
source.envir = list2env(list(MSFT = msft))

# save all of the .getSymbols Symbols and the source.envir Symbols
saveSymbols(trg = "pg", source.envir = source.envir)

unrate <- getSymbols("UNRATE", src = "FRED", auto.assign =  F)
saveSymbols(trg = "pg", source.envir = list2env(list(UNRATE = unrate)))

unrate.db <- getSymbols("UNRATE", src = "pg", auto.assign =  F)
unrate.db <- getSymbols(Symbols = "UNRATE", src = "pg", auto.assign = F)

}
}
