% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{saveSymbols}
\alias{saveSymbols}
\title{saves xts object symbols to a persistent location (dispatcher)}
\usage{
saveSymbols(Symbols = NULL, env = parent.frame(),
  source.envir = NULL, Gathering = "source.envir", ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{env}{location of xts objects placed that had been aquired
with getSymbols("XXX", src = 'yahoo')
Gathering = "DotGetSymbols" and "file.path" look 'here' for Symbols}

\item{source.envir}{source location of Symbols}

\item{Gathering}{places to look for symbols to be collected
Can be "DotgetSymbols" and/or "source.envir"
Symbol names must be case-insenstive unique.}

\item{...}{passed to trg}

\item{file.path}{if provided will save to disk}

\item{trg}{if provided will save to a target "cache"(Symbol name prepended with a ".")
or "pg" (PostgreSQL)
See ? saveSymbols.cache and ? saveSymbols.pg}

\item{target.envir}{if provided will save to to memory
Note: .getSymbols is * NOT UPDATED.*  A user is reponsible for 'self-tracking.'}
}
\description{
First, it will look for Symbols in the .getSymbols file and env and gather them
Next,  it will look for Symbols in source.envir and gather them
}
\details{
If provided file.path, then object will be stored on disk(same as "RData")
If provided trg == "RData", "cache", or "PostgreSQL (or just "pg")
then the objectw be ALSO saved in this OTHER location

Needs either/both "trg" or "file.path"
}
