% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{saveSymbols}
\alias{saveSymbols}
\title{saves xts object symbols to a persistent location (dispatcher)}
\usage{
saveSymbols(Symbols = NULL, env = parent.frame(),
  source.envir = NULL, ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded}

\item{env}{location of xts objects downloaded with getSymbols("XXX", src = 'yahoo')}

\item{source.envir}{source location of Symbols
xts objects must have the attribute "src"}

\item{...}{pass through parameters to saveSymbols.fn}
}
\description{
saves xts object symbols to a persistent location (dispatcher)
}
\examples{
\dontrun{

# Symbols must be case-insenstive unique

getSymbols("IBM", src = "yahoo") # auto.assign = TRUE

# Symbols names found in Symbols and the names of xts objects stored
# in source.envir, must be mutually exclusive ( case in-sensensitive match)
# (because list2env silently drops repeated members)

# save just the Symbol "IBM"
saveSymbols("IBM", trg = "pg")

msft <- getSymbols("MSFT", src = "yahoo", auto.assign = FALSE)
source.envir = list2env(list(MSFT = msft))
# save only the source.envir Symbols
saveSymbols(Symbols = "", trg = "pg", source.envir = source.envir)

# save all of the .getSymbols Symbols and the source.envir Symbols
saveSymbols(trg = "pg", source.envir = source.envir)

# save all to a file
# saveSymbols(source.envir = source.envir, file.path = "C:\\\\Users\\\\Public")

}
}
