% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{saveSymbols.RData}
\alias{saveSymbols.RData}
\title{saves xts object symbols to a persistent location (disk: Symbol.RData file)}
\usage{
saveSymbols.RData(Symbols = NULL, source.envir = NULL,
  file.path = stop("must specify 'file.path'"), ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{source.envir}{source location of Symbols}

\item{file.path}{character string of file (Symbol.Rdata) location}

\item{...}{pass through not used}
}
\description{
see quantmod SaveSymbols
}
