% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{saveSymbols.RData}
\alias{saveSymbols.RData}
\title{saves xts object symbols to a persistent location (disk: Symbol.RData file)}
\usage{
saveSymbols.RData(file.path = stop("must specify 'file.path'"),
  source.envir = NULL, ...)
}
\arguments{
\item{file.path}{character string of file (Symbol.Rdata) location}

\item{source.envir}{source location of Symbols}

\item{...}{pass through parameters}
}
\description{
see quantmod SaveSymbols
}
