% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{saveSymbols.RData}
\alias{saveSymbols.RData}
\title{saves xts object symbols to a persistent location (disk: Symbol.RData file)}
\usage{
saveSymbols.RData(Symbols = NULL,
  file.path = stop("must specify 'file.path'"), source.envir = NULL,
  ...)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol}

\item{file.path}{character string of file (Symbol.Rdata) location}

\item{source.envir}{source location of Symbols}

\item{...}{pass through not used}
}
\description{
\preformatted{

See quantmod SaveSymbols.

}
}
\examples{
\dontrun{

msft <- getSymbols("MSFT", src = "yahoo", auto.assign = FALSE)
source.envir = list2env(list(MSFT = msft))
# save to a file
saveSymbols(source.envir = source.envir, file.path = "C:\\\\Users\\\\Public")

}
}
