% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{SortinoRatioLoss}
\alias{SortinoRatioLoss}
\title{True Sortino Ratio 'loss function'}
\usage{
SortinoRatioLoss(actual, predicted, ...)
}
\arguments{
\item{...}{lev = NULL, model = NULL : passed to SortinoRatioSummary}

\item{see}{? iml::FeatureImp}
}
\description{
\preformatted{

for iml FeatureImp$new

}
}
