% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{SortinoRatioSummary}
\alias{SortinoRatioSummary}
\title{for package caret function trainControl parameter summaryFunction
True Sortino Ratio 'summary function'}
\usage{
SortinoRatioSummary(data, lev = NULL, model = NULL)
}
\description{
for package caret function trainControl parameter summaryFunction
True Sortino Ratio 'summary function'
}
