% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{symbolEyeBallWts}
\alias{symbolEyeBallWts}
\title{Symbol weights using eyeball}
\usage{
symbolEyeBallWts(xTs = NULL, Symbol = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{Symbol}{name prefix given to the generated output columns}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

This is the workhorse function. This is where the magic/logic happens.
Use any other columns (called indicators) that do not have the weights (_wts)
suffix and do not have the same root name compared to each and every
other *_wts column.

}
}
