% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeModel.R
\docType{package}
\name{tradeModel}
\alias{tradeModel}
\alias{tradeModel-package}
\title{tradeModel: Quantitive Mathematics and Financial Prediction}
\description{
functions
}
\section{Safe Data}{


initDate
initXts
initEnv
}

\section{Data Aquisition}{


fredData
fredEomData
}

\section{Target and Indicator Aquisition}{


Willshire5000*
Cash*
UnRate*
}

\section{Quanitative Mathematical}{


*LogReturns
}

\section{Date Manipulation}{


nextMonthfromYesterday
}

\section{Data Clustering}{


combineXts
}

\section{Investment Strategy}{


*Wts
}

\section{Field Identification}{


valueClmns
wtsClmns
indClms
}

\section{Total Return Calculating}{


portfolioLogReturns
portfolioMonthlyReturns
}

\section{Showing and Printing}{


printTail
printCalendar
}

\section{Full Strategies and Results}{


UnRateEyeBalltradeModel
UnRateMachinetradeModel
}

