% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{trends}
\alias{trends}
\title{collection of curves of the trends}
\usage{
trends(xTs, col = NULL, UseAOMX = FALSE)
}
\arguments{
\item{col}{column of the mktdata object to perform the trends}

\item{x}{mktdata object}
}
\description{
collection of curves of the trends
}
\examples{
\dontrun{

# called
trends(xTs, "UNRATE", UseAOMX = TRUE)

}
}
