% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{trendsWithAOMIndicators}
\alias{trendsWithAOMIndicators}
\title{collection of curves of the trends Indicator using average order max (AOMX)}
\usage{
trendsWithAOMIndicators(x = NULL, ...)
}
\description{
\preformatted{

}
}
