% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{UMCSentimentEomData}
\alias{UMCSentimentEomData}
\title{end of month University of Michigan: Consumer Sentiment data}
\usage{
UMCSentimentEomData()
}
\value{
xts object
}
\description{
\preformatted{

}
}
