% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{uninitEnv}
\alias{uninitEnv}
\title{unsets the enviroment}
\usage{
uninitEnv()
}
\value{
environment is un-set
}
\description{
space-saver - meant to be used at the beginning of a function
}
\examples{
\dontrun{
# > uninitEnv()
# getOption("digits")
# [1] 5
}
}
