% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{UnRateEyeBalltradeModel}
\alias{UnRateEyeBalltradeModel}
\title{Predicts the Symbol eom returns using UNRATE and the eyeball}
\usage{
UnRateEyeBalltradeModel(Symbol = NULL, src = NULL, Change = NULL)
}
\value{
xts object of monthly return results
}
\description{
Predicts the Symbol eom returns using UNRATE and the eyeball
}
\examples{
\dontrun{
# Predicts the FRED WILL5000IND eom returns using UNRATE and the eyeball
# UnRateEyeBalltradeModel(Symbol = "WILL5000IND", src = "FRED", Change = "apc")
}
}
