% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{UnRateEyeBalltradeModelGSPC}
\alias{UnRateEyeBalltradeModelGSPC}
\title{Predicts the Yahoo SP500 eom returns using UNRATE and the eyeball}
\usage{
UnRateEyeBalltradeModelGSPC()
}
\description{
Predicts the Yahoo SP500 eom returns using UNRATE and the eyeball
}
