% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{UnRateEyeBalltradeModelWILL5000IND}
\alias{UnRateEyeBalltradeModelWILL5000IND}
\title{Predicts the FRED WILL5000IND eom returns using UNRATE and the eyeball}
\usage{
UnRateEyeBalltradeModelWILL5000IND()
}
\description{
Predicts the FRED WILL5000IND eom returns using UNRATE and the eyeball
}
\examples{
\dontrun{
}
}
