% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{UnRateMachinetradeModel}
\alias{UnRateMachinetradeModel}
\title{Predicts the FRED WILL5000IND eom returns using UNRATE and Machine learning}
\usage{
UnRateMachinetradeModel()
}
\value{
xts object of monthly return results
}
\description{
Predicts the FRED WILL5000IND eom returns using UNRATE and Machine learning
}
\examples{
\dontrun{
EXAMPLE
# EXAMPLE
}
}
