% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{UnRateMachinetradeModel}
\alias{UnRateMachinetradeModel}
\title{Predicts Symbol eom returns using UNRATE and Machine learning}
\usage{
UnRateMachinetradeModel(Symbol = NULL, src = NULL)
}
\description{
Predicts Symbol eom returns using UNRATE and Machine learning
}
\examples{
\dontrun{

# Predicts the FRED WILL5000IND / yahoo S&P500 eom returns using UNRATE and the machine

UnRateMachinetradeModel(Symbol = "WILL5000IND", src = "FRED")
UnRateMachinetradeModel(Symbol = "^GSPC"      , src = "yahoo")
}
}
