% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{UnRateMachinetradeModel}
\alias{UnRateMachinetradeModel}
\title{Predicts Symbol eom returns using UNRATE and Machine learning}
\usage{
UnRateMachinetradeModel(Symbol = NULL, src = NULL, Predictors = NULL,
  IndicatorGeneratorFUN = NULL)
}
\description{
Predicts Symbol eom returns using UNRATE and Machine learning
}
\examples{
\dontrun{

# Predicts the FRED WILL5000IND / yahoo S&P500 eom returns using UNRATE (and/or 'others') and the machine

# done: !!! misses entire 2008 recession !!!
UnRateMachinetradeModel(Symbol = "WILL5000IND", src = "FRED",  Predictors = "UNRATE", IndicatorGeneratorFUN = "unrateEyeballIndicators")
UnRateMachinetradeModel(Symbol = "^GSPC"      , src = "yahoo", Predictors = "UNRATE", IndicatorGeneratorFUN = "unrateEyeballIndicators")


# alt derived data columns from 'trends'
UnRateMachinetradeModel(Symbol = "WILL5000IND", src = "FRED", Predictors = "UNRATE",  IndicatorGeneratorFUN = "trendsWithAOMXIndicators")
#
# technically works: but (Symbol = "^GSPC") gives UPSIDE DOWN predictions
UnRateMachinetradeModel(Symbol = "^GSPC"      , src = "yahoo", Predictors = "UNRATE", IndicatorGeneratorFUN = "trendsWithAOMXIndicators")


# good captures entire 2008 recession 'and' early 2019(current time data: just before edge)
UnRateMachinetradeModel(Symbol = "WILL5000IND", src = "FRED",  Predictors = c("UNRATE","UMCSENT"), IndicatorGeneratorFUN = c("trendsWithAOMXIndicators", "trendsWithAOMNIndicators"))
#
# good captures entire 2008 recession 'and' early 2019(current time data: just before edge)
UnRateMachinetradeModel(Symbol = "^GSPC"      , src = "yahoo", Predictors = c("UNRATE","UMCSENT"), IndicatorGeneratorFUN = c("trendsWithAOMXIndicators", "trendsWithAOMNIndicators"))


# good: captures entire 2008 recession 'and' early 2019(current time data: just before edge)
UnRateMachinetradeModel(Symbol = "WILL5000IND", src = "FRED",  Predictors = c("UNRATE","UMCSENT","Earnings"), IndicatorGeneratorFUN = c("trendsWithAOMXIndicators", "trendsWithAOMNIndicators", "trendsWithAOMNIndicators"))
#
# good: captures entire 2008 recession 'and' early 2019(current time data: just before edge)
UnRateMachinetradeModel(Symbol = "^GSPC"      , src = "yahoo", Predictors = c("UNRATE","UMCSENT","Earnings"), IndicatorGeneratorFUN = c("trendsWithAOMXIndicators", "trendsWithAOMNIndicators", "trendsWithAOMNIndicators"))


# done: extra SP500PriceEarningsRat !!! misses entire 2008 recession !!!
# SP500PriceEarningsRat (not as much impact as I originally thought or 'wrong' impact)
#
UnRateMachinetradeModel(Symbol = "WILL5000IND", src = "FRED",  Predictors = c("UNRATE","UMCSENT","Earnings", "SP500PriceEarningsRat"), IndicatorGeneratorFUN = c("trendsWithAOMXIndicators", "trendsWithAOMNIndicators", "trendsWithAOMNIndicators", "trendsWithAOMNIndicators"))
UnRateMachinetradeModel(Symbol = "^GSPC"      , src = "yahoo", Predictors = c("UNRATE","UMCSENT","Earnings", "SP500PriceEarningsRat"), IndicatorGeneratorFUN = c("trendsWithAOMXIndicators", "trendsWithAOMNIndicators", "trendsWithAOMNIndicators", "trendsWithAOMNIndicators"))

}
}
