% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{valueCurrentRetsClms}
\alias{valueCurrentRetsClms}
\title{get the currentrets columns}
\usage{
valueCurrentRetsClms(xTs = NULL)
}
\arguments{
\item{xTs}{xts object}
}
\value{
column names
}
\description{
\preformatted{

Value column names end in "currentrets" without "_wts".

}
}
\examples{
\dontrun{
# xTs <- xts(matrix(1:3, ncol = 3, dimnames = list(NULL,c("a","b","b_currentrets"))),zoo::as.Date(0))[0]
# valueCurrentRetsClmsClms(xTs)
# [1] "b_currentrets"
}
}
