% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{willShire5000MachineWts}
\alias{willShire5000MachineWts}
\title{add Willshire 5000 Index weights using Machine learning}
\usage{
willShire5000MachineWts(xTs = NULL, NumbReplicaCopiesMultiple = NULL)
}
\arguments{
\item{xTs}{xts object}

\item{NumbReplicaCopiesMultiple}{passed to NumbReplicaCopies
This means how much more ( e.g. "focused" data ) is replicated compared to "all" data}
}
\value{
xts object with merged data into xTs
}
\description{
\preformatted{

This is the workhorse function. This is where the magic/logic happens.
Use any other columns (called indicators) that do not have the weights (_wts)
suffix and do not have the same root name compared to each and every
other *_wts column.

}
}
