% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{wilshire5000indEomData}
\alias{wilshire5000indEomData}
\title{Wilshire 5000 Index price}
\usage{
wilshire5000indEomData()
}
\value{
xts object of end of month returns
}
\description{
Wilshire 5000 Index price
}
